### 2025-06-24 #8468 @editorialbot
[PRE REVIEW]: jetfuelburn: A Python package for calculating fuel burn of commercial aircraft
**Submitting author:** @michaelweinold (<a href="http://orcid.org/0000-0003-4859-2650">Michael Philipp Weinold</a>)
**Repository:** https://github.com/sustainableaviation/jetfuelburn
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
