### 2025-08-08 27f867c
moved to src/jetfuelburn file structure
