+++ src/jetfuelburn/__init__.py
@@ -2,4 +2,4 @@
 from pint import UnitRegistry
 ureg = UnitRegistry()
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
