+++ CHANGES.md
@@ -3,6 +3,12 @@
 The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `1.0.2` (08. August 2025)
+
+### Improvements
+
+- Moved to `src/jetfuelburn` directory structure.
+
 ## `1.0.1` (08. August 2025)
 
 ### Fixes
