### 2025-08-15 sustainableaviation/jetfuelburn#22 @michaelweinold
JOSS Publication
- https://github.com/openjournals/joss-reviews/issues/8468
- https://github.com/openjournals/joss-reviews/issues/9280
