### 2025-09-29 c389533
added notice to eea model
