### 2025-09-30 #8468 @kyleniemeyer
@michaelweinold a few comments:
- I can see a `tests` directory, but no information about how to run the tests or if you have continuous integration set up
- The documentation website looks great, but the README should still have some basic information about the package, how to install, and possibly how to run the test suite. (Badges at the top to indicate CI status, etc., are also often very helpful indicators)
