### 2025-09-30 #8468 @michaelweinold
Updated the readme file in https://github.com/sustainableaviation/jetfuelburn/commit/35f72ccbdf81c4b5bb512010bcd1e5b973fc5f33.
