### 2025-10-07 1abfcbc
renamed aux to utility
