+++ .github/workflows/test_package.yml
@@ -32,9 +32,9 @@ jobs:
                 python-version: ["3.10", "3.11", "3.12", "3.13"]
 
         steps:
-            - uses: actions/checkout@v4
+            - uses: actions/checkout@v5
             - name: Set up Python ${{ matrix.python-version }}
-              uses: actions/setup-python@v5
+              uses: actions/setup-python@v6
               with:
                 python-version: ${{ matrix.python-version }}
             - name: Display Python version
