+++ docs/api/physics.md
@@ -1,1 +1,1 @@
-::: jetfuelburn.aux.physics
+::: jetfuelburn.utility.physics
