+++ docs/api/allocation.md
@@ -1,1 +1,1 @@
-::: jetfuelburn.aux.allocation
+::: jetfuelburn.utility.allocation
