### 2025-10-16 c4f87c8
#30 added warning to aim model docstrings
