### 2025-10-21 #8468 @aliaksei135
I'd be happy to review
