### 2025-10-23 05e0980
added model comparison
