+++ mkdocs.yaml
@@ -5,7 +5,7 @@ copyright: © Paul Scherrer Institute (PSI) and Contributors | Code under MIT Li
 nav: 
   - Home: 'index.md'
   - About: 'about.md'
-  #- Comparing Models: 'comparison.md'
+  - Comparing Models: 'comparison.md'
   - Getting Started:
     - Installation: 'installation.md'
     - Working with Physical Units: 'interactive/units.md'
