### 2025-10-24 4af829e
added notebook download button
