### 2025-10-24 a381125
added documentation for non-jetfuelburn models
