### 2025-10-24 2805bfe
added documentation for non-jetfuelburn models
