### 2025-10-28 a55ef0f
updated documentation of external models
