+++ docs/api/reducedorder_external.md
@@ -4,16 +4,78 @@
 
 OpenAP (Open Aircraft Performance Model and Toolkit) is an open-source Python library developed by researchers at TU Delft for aircraft performance and emission calculations. The library also includes modules for drag, thrust, kinematic modeling, and trajectory generation.
 
-**Fuel and Emission Module:**
+### Fuel and Emission Module
 
-The OpenAP fuel module uses polynomial models derived from the ICAO Engine Emissions Databank and the Acropole model. The `FuelFlow()` class computes fuel consumption for different flight phases (taxi, takeoff, climb, cruise, descent) based on aircraft mass, true airspeed, and altitude. The `Emission()` class calculates pollutant emissions (NOx, CO, HC) using ICAO engine data.
+The OpenAP fuel module uses polynomial models derived from the ICAO Engine Emissions Databank and the Acropole model.
 
-!!! note
-    The OpenAP Model requires a full flight trajectory as input for fuel calculations.
+- The `FuelFlow()` class computes instantaneous fuel consumption for different flight phases (taxi, takeoff, climb, cruise, descent) based on aircraft mass, true airspeed (TAS), and altitude.  
+- The `Emission()` class calculates pollutant emissions (NOₓ, CO, HC) using corresponding ICAO engine emission data.
+
+### Full Flight Trajectory Input Requirement
+The OpenAP model requires a full flight trajectory as a time-series input to compute fuel consumption and emissions across all flight phases. This trajectory is typically provided as a tabular structure (e.g., a Pandas DataFrame) describing the aircraft’s state at regular time intervals.
+
+| Required Kinematic Input Columns | Description                                                     |
+|----------------------------------|-----------------------------------------------------------------|
+| mass or m0                       | Aircraft mass (Take-off mass is needed for initialization)      |
+| tas                              | True Airspeed throughout the flight                             |
+| h or altitude                    | Altitude throughout the flight                                  |
+| vs or vertical_rate              | Vertical speed throughout the flight                            |
+
+### Trajectory Acquisition
+Flight trajectories can be obtained in two ways:
+
+1. Generation: Using the `FlightGenerator` class in the `openap.gen` module to synthesize representative trajectories.
+2. Actual Flight Data: Importing real flight trajectories from ADS-B data providers such as OpenSky Network, Flightradar24, or airplanes.live.
+
+!!! example "Usage Example"
+    ```python
+    from openap import FuelFlow
+    import pandas as pd
+    
+    # Initialize FuelFlow model for A319
+    fuelflow = FuelFlow('A319')
+    mass_takeoff_assumed = 66300  # kg
+
+    # Load flight trajectory data
+    df = pd.read_csv(
+        'data/flight_a319_opensky.csv',
+        parse_dates=['timestamp'],
+        dtype={'icao24': str},
+    )
+
+    # Calculate time deltas between consecutive points
+    df = df.assign(d_ts=lambda d: d.timestamp.diff().dt.total_seconds().bfill())
+    
+    # Calculate fuel burn at each trajectory point
+    mass_current = mass_takeoff_assumed
+    fuelflow_every_step = []
+    fuel_every_step = []
+
+    for i, row in df.iterrows():
+        # Calculate instantaneous fuel flow
+        ff = fuelflow.enroute(
+            mass=mass_current,
+            tas=row.groundspeed,
+            alt=row.altitude,
+            vs=row.vertical_rate,
+        )
+        fuel = ff * row.d_ts
+        fuelflow_every_step.append(ff)
+        fuel_every_step.append(fuel)
+        mass_current -= fuel
[…]
