### 2025-10-28 b079c92
updated documentation of external models
