+++ docs/api/statistics_external.md
@@ -2,12 +2,12 @@
 
 ## Travel Impact Model (Google)
 
-The Google Travel Impact Model (TIM) estimates flight fuel consumption and associated emissions using the Tier 3A methodology (flight-specific, Origin-Destination approach) defined in the EMEP/EEA Annex 1.A.3.a Aviation 2023. For each scheduled flight, fuel consumption is calculated by combining data from standard databases for two phases:
+The Google [Travel Impact Model (TIM)](https://travelimpactmodel.org) estimates flight fuel consumption and associated emissions using the Tier 3A methodology (flight-specific, Origin-Destination approach) defined in the EMEP/EEA Annex 1.A.3.a Aviation 2023. For each scheduled flight, fuel consumption is calculated by combining data from standard databases for two phases:
 
-- **LTO phase (Taxi, Takeoff, Landing):** Fixed fuel burn calculated using ICAO Aircraft Engine Emissions Databank (AEED) standards (LTO cycle fuel flow/emission indices).
+- **LTO phase (Taxi, Takeoff, Landing):** Fixed fuel burn calculated using [ICAO Aircraft Engine Emissions Databank (AEED)](https://www.easa.europa.eu/en/domains/environment/icao-aircraft-engine-emissions-databank) standards (LTO cycle fuel flow/emission indices).
 - **CCD phase (Climb, Cruise, Descent):** Variable fuel burn derived from EUROCONTROL Advanced Emission Model (AEM)/Base of Aircraft Data (BADA) performance modeling, adjusted for real-world routing and calculated via linear interpolation/extrapolation for stage length.
 
-Total fuel is converted to CO₂-equivalent using ISO 14083-compliant Well-to-Wake (WTW) life-cycle factors (3.8359 kg CO₂e/kg fuel). The final passenger estimate is derived via a three-step apportionment process: allocating total emissions between cargo and passenger payloads using the mass-based approach, normalizing passenger count via statistical load factors, and applying cabin class weightings based on actual or median seat configuration data.
+Total fuel is converted to CO₂-equivalent using [ISO 14083](https://www.iso.org/standard/78864.html)-compliant Well-to-Wake (WTW) life-cycle factors (3.8359 kg CO₂e/kg fuel). The final passenger estimate is derived via a three-step apportionment process: allocating total emissions between cargo and passenger payloads using the mass-based approach, normalizing passenger count via statistical load factors, and applying cabin class weightings based on actual or median seat configuration data.
 
 This model is integrated into the Google Flights interface and displays the CO₂-equivalent values to users searching for flight options:
 
@@ -20,7 +20,9 @@ This model is integrated into the Google Flights interface and displays the CO�
 !!! note
     The Travel Impact Model in Google Flights only shows emissions for _future_ flights (which can still be booked). It does not provide emissions data for past flights.
 
+!!! reference "See Also"
+    - [Google Travel Impact Model Website](https://travelimpactmodel.org)
+    - [Google Travel Impact Model API Documentation](https://developers.google.com/travel/impact-model)
+
 !!! reference "References"
-    - [Official Website](http://www.travelimpactmodel.org/)
-    - [GitHub Repository](https://github.com/google/travel-impact-model)
-    - [EMEP/EEA Air Pollutant Emission Inventory Guidebook 2023 (Annex 1.A.3.a Aviation)](https://www.eea.europa.eu/en/analysis/publications/emep-eea-guidebook-2023)
+    [EMEP/EEA Air Pollutant Emission Inventory Guidebook 2023 (Annex 1.A.3.a Aviation)](https://www.eea.europa.eu/en/analysis/publications/emep-eea-guidebook-2023)

[…]
