### 2025-11-01 #9280 @xoolive
Hello,

Nice package! I would just regret that the paper is a bit light into showcasing the possibilities of the library. Do you think you could include some results with nice plots about the library possibilities?

Also don't forget the community guidelines, they are in the check list
