+++ CONTRIBUTING.MD
@@ -1,0 +1,26 @@
+# Contributing
+
+We welcome contributions!
+
+If you have any questions, [start a discussion](https://github.com/sustainableaviation/jetfuelburn/discussions) in this repository.
+
+## Contributing to the Code
+
+If you want to contribute to the package, please follow the [GitHub contribution workflow (fork, branch, PR)](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests).
+
+### Code Style
+
+Contributors must follow the [NumPy docstring style guide for documentation](https://numpydoc.readthedocs.io/en/latest/format.html) and [PEP 8](https://www.python.org/dev/peps/pep-0008/) for code style.
+
+## Reporting Bugs or Errors
+
+Something is not working as expected? You have two options:
+
+### 🥈 Report an Error
+
+Please open a new issue in this [repository](https://github.com/sustainableaviation/jetfuelburn/issues), describing the error and where you found it.
+A member of the developer community will then take care of the issue, but it may take some time for your issue to be resolved.
+
+### 🥇  Fix an Error yourself
+
+If you have a solution to the error, please follow the [GitHub contribution workflow (fork, branch, PR)](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests).
