### 2025-11-02 98019f7
updated paper
