### 2025-11-02 0482b4f
updated citations
