+++ joss/paper.bib
@@ -27,6 +27,13 @@
   year={2010},
   doi={10.2514/6.2010-9156}
 }
+@inproceedings{weinold2025csfm,
+  title={Estimating Fuel Consumption of Commercial Aircraft},
+  author={Weinold, Michael P and Marty, Noe and Wild, Peter and McKenna, Russell},
+  booktitle={CSFM Symposium 2025},
+  year={2025},
+  doi={10.3929/ethz-b-000749349}
+}
 @inproceedings{yanto2017efficient,
   title={Efficient fast approximation for aircraft fuel consumption for decision-making and policy analysis},
   author={Yanto, Jefry and Liem, Rhea P},
