### 2025-11-02 adf150e
updated cross-references
