### 2025-11-02 #9280 @michaelweinold
Hello @xoolive! Thank you for your efforts in reviewing our submission. [A contribution guide has now been added.](https://github.com/sustainableaviation/jetfuelburn/blob/main/CONTRIBUTING.MD) The paper has been updated with two figures showcasing possible applications of the model.
