+++ docs/about.md
@@ -2,7 +2,21 @@
 
 ## Authors
 
-The `jetfuelburn` package was written by [Michael Weinold](https://www.linkedin.com/in/michaelweinold/) under the supervision of [Prof. Russell McKenna](https://esc.ethz.ch/people/person-detail.mckenna.html) as part of his doctoral research at ETH Zurich and Paul Scherrer Institute (PSI).
+The `jetfuelburn` package was developed by [Michael Weinold](https://www.linkedin.com/in/michaelweinold/) under the supervision of [Prof. Russell McKenna](https://esc.ethz.ch/people/person-detail.mckenna.html) as part of his doctoral research at ETH Zurich and Paul Scherrer Institute (PSI).
+
+### Other Contributors
+
+| Contributor | Role | Contributions | Date |
+|-------------|------|---------------|------|
+| [Noe Marty](https://www.linkedin.com/in/noemarty/) | Semester Project Student at ETH Zurich | [Comparison of Methods](https://doi.org/10.3929/ethz-b-000749349) | 2025 |
+
+## Publications
+
+To cite the use of `jetfuelburn` in academic work, please use the main publication in the [Journal of Open Source Software (JOSS)](http://joss.theoj.org):
+
+### Publications using `jetfuelburn`
+
+Weinold, M.P., Marty, N., Wild, P., McKenna, R. (2025). Estimating Fuel Consumption of Commercial Aircraft. _Poster presented at the ETH Zurich CSFM Symposium_, Zurich, Switzerland, October 02, 2025. [doi:10.3929/ethz-b-000749349](https://doi.org/10.3929/ethz-b-000749349)  
 
 ## Funding
 
