+++ docs/about.md
@@ -14,10 +14,6 @@ The `jetfuelburn` package was developed by [Michael Weinold](https://www.linkedi
 
 To cite the use of `jetfuelburn` in academic work, please use the main publication in the [Journal of Open Source Software (JOSS)](http://joss.theoj.org):
 
-### Publications using `jetfuelburn`
-
-Weinold, M.P., Marty, N., Wild, P., McKenna, R. (2025). Estimating Fuel Consumption of Commercial Aircraft. _Poster presented at the ETH Zurich CSFM Symposium_, Zurich, Switzerland, October 02, 2025. [doi:10.3929/ethz-b-000749349](https://doi.org/10.3929/ethz-b-000749349)  
-
 ## Funding
 
 Funding was provided by the [Swiss Innovation Agency Innosuisse](https://innosuisse.admin.ch) through the ["WISER" Flagship Project](https://wiser-climate.com) (PFFS-21-72). Michael Weinold gratefully acknowledges the support of the [Swiss Study Foundation](https://www.studyfoundation.ch).
