### 2025-11-03 78302e2
updated homepage figure
