### 2025-11-03 279f886
updated homepage figure
