+++ docs/index.md
@@ -1,6 +1,6 @@
 # JetFuelBurn
 
-`jetfuelburn` is a Python package for calculating fuel burn in aircraft. It is designed to be used in the context of environmental impact assessment of air travel, aircraft performance analysis and optimisation. It includes helper function for basic atmospheric physics and the allocation of fuel burn to different cabin classes.
+`jetfuelburn` is a Python package for calculating fuel burn in aircraft. It is designed to be used in the context of environmental impact assessment of air travel. It includes helper function for basic atmospheric physics and the allocation of fuel burn to different cabin classes.
 
 The package implements models from textbooks and scientific publications. Some are based on basic aerodynamic equations like the [Breguet range equation](https://en.wikipedia.org/wiki/Range_(aeronautics)), while others are more complex and ultimately based on [EUROCONTROL BADA](https://www.eurocontrol.int/model/bada) or [Piano X](https://www.lissys.uk/index2.html) simulation data.
 
