### 2025-11-03 8f768db
removed comparison figures from paper
