### 2025-11-05 f4dac57
udpated homepage wording
