### 2025-11-07 e57098b
updated EU FEL documentation
