+++ docs/api/eu_fel.md
@@ -2,18 +2,11 @@
 
 ## Overview
 
-The EU Flight Emissions Label (FEL) is a voluntary environmental labelling scheme established under Article 14 of Regulation (EU) 2023/2405 (ReFuelEU Aviation). It enables transparent, standardized disclosure of verified CO₂-equivalent emissions for commercial flights departing from or arriving at EU airports, covering all flights operated by participating airlines.
+The [EU Flight Emissions Label (FEL)](https://www.flightemissions.eu/en) is a voluntary environmental labelling scheme established under [Article 14 of Regulation (EU) 2023/2405 (ReFuelEU Aviation)](https://eur-lex.europa.eu/eli/reg/2023/2405/oj/eng). The label discloses CO₂-equivalent emissions for commercial flights departing from or arriving at EU airports, covering all flights operated by participating airlines.
 
-Once an operator becomes a label-holder, the FEL must be displayed for all flights in scope, ensuring consistent and non-selective reporting across its network.
+As of now, participation is voluntary. Once an operator chooses to take part, the FEL must be displayed for all flights in scope, ensuring consistent and non-selective reporting across its network.
 
-The FEL is implemented by EASA through its Sustainability Portal and integrates verified operational data such as fuel burn, passenger load, and SAF share to generate per-flight emissions metrics. Participation is voluntary, but once an operator opts in, labels must cover every applicable flight included within the scheme's scope.
-
-## Purpose
-
-- Provide trusted, comparable flight-level CO₂ data to passengers  
-- Support consumer transparency and counter greenwashing  
-- Highlight verified operational performance (fuel burn + SAF usage)  
-- Complement ReFuelEU Aviation sustainable-fuel obligations
+The FEL is implemented by EASA through its [Sustainability Portal](https://www.easa.europa.eu/en/domains/environment/sustainability-portal) and uses verified operational data such as fuel burn, passenger load, and SAF share to generate per-flight emissions metrics.
 
 ## Label Structure
 
@@ -31,55 +24,55 @@ Each FEL record represents a unique combination:
 ### Regulatory Overview
 | Reference | Description |
 |:--|:--|
-| **Article 14**, Reg. (EU) 2023/2405 | Establishes voluntary Flight Emissions Label scheme |
-| **Annex II**, Reg. (EU) 2024/3170 | Defines official methodology for calculating flight emissions |
-| **Article 9 (1)** | Prohibits duplicate SAF claims across EU ETS, CORSIA and FEL |
-| **Article 14 (9)** | Requires operators to pay a fee to EASA for label issuance |
-| **Article 17 (7)** | Commission review by July 2027 to assess making FEL mandatory |
+| [**§ 14**, Reg. (EU) 2023/2405 (ReFuelEU Aviation)](https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX:32023R2405#art_14) | Establishes the voluntary Flight Emissions Label scheme |
+| [**Annex II**, Implementing Reg. (EU) 2024/3170](https://eur-lex.europa.eu/eli/reg_impl/2024/3170/oj/eng#anx_II) | Defines the official methodology for calculating flight emissions |
+| [**§ 9 (1)**, Reg. (EU) 2023/2405](https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX:32023R2405#art_9) | Prohibits duplicate SAF claims across EU ETS, CORSIA and FEL |
[…]
