+++ docs/api/eu_fel.md
@@ -35,44 +35,31 @@ Each FEL record represents a unique combination:
 The following extracts provide the legal foundation for the Flight Emissions Label (FEL). They are quoted directly from EU law with context for application within JetFuelBurn.
 
 ### § 14 — Establishment of the Flight Emissions Label
-> **Regulation (EU) 2023/2405 — ReFuelEU Aviation**  
-> [📘 Full text on EUR-Lex](https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX:32023R2405#art_14)
 
-!!! reference "§ 14 – Voluntary Environmental Labelling Scheme"
-    > (1) “A voluntary environmental labelling scheme enabling the environmental performance of flights to be measured is hereby established.”  
-    >  
-    > (2) “Labels issued pursuant to this Article shall apply to aircraft operators falling within the scope of this Regulation for flights covered by this Regulation departing from Union airports. Where an aircraft operator requests the issuance of a label under this Article, it shall request such a label for all its flights covered by this Regulation departing from Union airports.”  
-    >  
-    > “Aircraft operators may request the issuance of labels under this Article also for their flights covered by this Regulation arriving at Union airports. Where an aircraft operator requests the issuance of a label under this subparagraph, it shall request such a label for all its flights arriving at Union airports.”  
-    >  
-    > …  
-    >  
-    > (9) “In order to finance the costs of the service provided by the Agency, the issuing of a label at the request of an aircraft operator shall be subject to the payment of a charge … The amount of the charge shall be defined pursuant to Article 126(4) of Regulation (EU) 2018/1139.”  
-    >  
-    > …  
-    >  
-    > (12) “By 1 July 2027, the Commission shall identify and assess … with a view in particular to establish a compulsory environmental labelling scheme … The Commission shall present a report … to the European Parliament and to the Council.”  
-    — *§ 14 (1)–(2), (9), (12) Reg. (EU) 2023/2405 (ReFuelEU Aviation), OJ L 2023/2405, 31 Oct 2023.*
+> (1) “A voluntary environmental labelling scheme enabling the environmental performance of flights to be measured is hereby established.”  
+>  
+> (2) “Labels issued pursuant to this Article shall apply to aircraft operators falling within the scope of this Regulation for flights covered by this Regulation departing from Union airports. Where an aircraft operator requests the issuance of a label under this Article, it shall request such a label for all its flights covered by this Regulation departing from Union airports.”  
+>  
+> “Aircraft operators may request the issuance of labels under this Article also for their flights covered by this Regulation arriving at Union airports. Where an aircraft operator requests the issuance of a label under this subparagraph, it shall request such a label for all its flights arriving at Union airports.” (...)
+>  
+> (9) “In order to finance the costs of the service provided by the Agency, the issuing of a label at the request of an aircraft operator shall be subject to the payment of a charge … The amount of the charge shall be defined pursuant to Article 126(4) of Regulation (EU) 2018/1139.” (...) 
+>  
+> (12) “By 1 July 2027, the Commission shall identify and assess … with a view in particular to establish a compulsory environmental labelling scheme … The Commission shall present a report … to the European Parliament and to the Council.”  
+>
+> — [*§ 14 (1)–(2), (9), (12) Reg. (EU) 2023/2405 (ReFuelEU Aviation), OJ L 2023/2405, 31 Oct 2023.*](https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX:32023R2405#art_14)
 
[…]
