### 2025-11-14 aab6a8f
updated faq and api reference
