### 2025-11-14 999c2d8
updated faq and api reference
