+++ mkdocs.yaml
@@ -16,7 +16,7 @@ nav:
     - PL/R-Diagrams: 'api/diagrams.md'
     - Range Equation: 'api/breguet.md'
     - Statistics Models: 'api/statistics.md'
-    - Reduced Order Models: 'api/reducedorder.md'
+    #- Reduced Order Models: 'api/reducedorder.md'
     - Combined Models: 'api/combined.md'
     - Closed Form Models: 'api/closedform.md' 
     - Cabin Class Fuel Allocation: 'api/allocation.md'
