+++ joss/paper.md
@@ -40,6 +40,16 @@ Aerospace engineering researchers have proposed numerous methods for estimating 
 
 \clearpage
 
+# Example Use Case
+
+The `jetfuelburn` package can be used to compare different fuel burn models for specific flight routes and aircraft types. \autoref{fig:comparison} shows a comparison of fuel burn estimates for five different route/aircraft-combinations, using several models implemented in the `jetfuelburn` package. The results show that while all models generally agree on the trend of increasing fuel burn with distance, there are significant differences in the absolute values predicted by each model. These differences become larger as the flight distance increases. This highlights the importance of selecting an appropriate model for specific applications, as the choice of model can have a significant impact on the results of environmental assessments.
+
+![Comparison of per-passenger (economy-class) carbon emissions derived from fuel burn estimates for different routes/aircraft-combinations. A conversion factor of 2.63 kg CO$_2$ per liter of Jet A/A-1 fuel was used. Colors indicate model type: orange - statistical, red: machine-learning, blue: reduced-order, green: closed-form. For a geographical representation of the routes, compare \autoref{fig:map}. Adapted from [@weinold2025csfm, Figure 4] \label{fig:comparison}](_media/fuel_calc_comparison.svg){height="5.5cm"}
+
+![Overview of the routes shown in \autoref{fig:comparison}. Note that in cases where a flight along the great-circle between origin and destination is not possible due to eg. geopolitical reasons, a multiplier can be applied. \label{fig:map}](_media/map_flights_emissions_flight_routes.png){height="5cm"}
+
+\clearpage
+
 # Fuel Calculation Model Categories
 
 The `jetfuelburn` package includes different methods for calculating fuel burn of commercial aircraft. These methods can be broadly categorized into four groups:
