### 2025-11-15 #9280 @michaelweinold
I have now updated the comparison figure. Happy to go ahead with the publication process, @kyleniemeyer!
