+++ joss/paper.bib
@@ -69,12 +69,13 @@
   publisher={John Wiley \& Sons},
   doi={10.1002/9781118534786}
 }
-@article{burzlaff2017aircraft,
+@misc{burzlaff2017aircraft,
   title={Aircraft Fuel Consumption - Estimation and Visualization},
+  type={University Student Project},
   author={Burzlaff, Marcus},
   year={2017},
-  publisher={HAW Hamburg},
-  url={https://www.fzt.haw-hamburg.de/pers/Scholz/arbeiten/TextBurzlaff.pdf}
+  institution={HAW Hamburg},
+  doi={10.7910/DVN/2HMEHB}
 }
 @article{keiser2023life,
   title={Life cycle assessment in aviation: A systematic literature review of applications, methodological approaches and challenges},
