### 2025-12-27 #9280 @michaelweinold
- [x] can you edit the title on the Zenodo archive to match the paper?
- [x] throughout the paper, make sure "e.g.," is formatted correctly (with periods after each letter, and a comma)
- [x] In the "Reduced Order Models" section, the last sentence needs some editing:
Publications implemented in the initial version of jetfuelburn include ...
- [x] The Burzlaff citation appears to be missing some information. Is this a report, or something else?
    _This is the final report of a university project conducted by Burzlaff under the supervision of Scholz. The bibtex entry type has been changed to `misc`._
- [x] The Yanto and Liem citation is formatted as a journal article, but that is a conference paper.
    _This entry was already listed as `inproceedings`. I don't think there is much else I can do to influence the bibliography formatting?_
