### 2026-01-10 f7fc69d
updated citations
