### 2026-01-10 0b4c82c
added JOSS publication
