+++ mkdocs.yaml
@@ -56,6 +56,7 @@ plugins:
 markdown_extensions:
   - admonition
   - pymdownx.superfences
+  - pymdownx.details
   - toc:
       permalink: true
   - pymdownx.arithmatex:
