### 2026-01-10 4b6270e
added JOSS publication
