+++ docs/about.md
@@ -4,15 +4,39 @@
 
 The `jetfuelburn` package was developed by [Michael Weinold](https://www.linkedin.com/in/michaelweinold/) under the supervision of [Prof. Russell McKenna](https://esc.ethz.ch/people/person-detail.mckenna.html) as part of his doctoral research at ETH Zurich and Paul Scherrer Institute (PSI).
 
-### Other Contributors
+## Contributors
 
 | Contributor | Role | Contributions | Date |
 |-------------|------|---------------|------|
 | [Noe Marty](https://www.linkedin.com/in/noemarty/) | Semester Project Student at ETH Zurich | [Comparison of Methods](https://doi.org/10.3929/ethz-b-000749349) | 2025 |
 
-## Publications
+## Citing the Tool
 
-To cite the use of `jetfuelburn` in academic work, please use the main publication in the [Journal of Open Source Software (JOSS)](http://joss.theoj.org):
+To cite the use of `jetfuelburn` in academic work, please the main publication:
+
+> Weinold and McKenna, (2026). JetFuelBurn: A Python package for calculating fuel burn of commercial aircraft. _Journal of Open Source Software_, 11(117), 9280, [doi:10.21105/joss.09280](https://doi.org/10.21105/joss.09280).
+
+??? quote "BibTeX Entry"
+    ```bibtex
+    @article{weinold2026jetfuelburn,
+    title={JetFuelBurn: A Python package for calculating fuel burn of commercial aircraft},
+    author={Weinold, Michael P and McKenna, Russell},
+    journal={Journal of Open Source Software},
+    volume={11},
+    number={117},
+    pages={9280},
+    year={2026},
+    doi={10.21105/joss.09280}
+    }
+    ```
+
+## Publications using the Tool
+
+| Year | Authors | Title |
+|------|---------|-------|
+| 2025 | Weinold et al. | [Estimating Fuel Consumption of Commercial Aircraft](https://doi.org/10.3929/ethz-b-000749349) |
+| 2026 | Weinold and McKenna | [JetFuelBurn: A Python package for calculating fuel burn of commercial aircraft](https://doi.org/10.21105/joss.09280) |
+
 
 ## Funding
 
