+++ LICENSE.txt
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright © 2025 Paul Scherrer Institute (PSI) and Contributors
+Copyright © 2026 Paul Scherrer Institute (PSI) and Contributors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
