+++ src/jetfuelburn/statistics.py
@@ -4,6 +4,91 @@ import json
 from importlib import resources
 
 
+class aeromaps:
+    r"""
+    This class implements the the reduced-order fuel burn model of the AeroMaps software tool by Planès et al. (2023).
+
+    In this model, fuel burn calculations are based on statistical average values in MJ/ASK (megajoule per available seat-kilometer) for three representative aircraft types.
+    Data is stored in an Excel file included in the AeroMAPS repository: [`aeromaps/resources/data/data.xlsx`](https://github.com/AeroMAPS/AeroMAPS/blob/5febefe03cd69815a6949b67c43906c209ac14b1/aeromaps/resources/data/data.xlsx)
+
+    A conversion factor of 
+    
+    References
+    ----------
+    Planès, T., Delbecq, S., & Salgas, A. (2023).
+    AeroMAPS: a framework for performing multidisciplinary assessment of prospective scenarios for air transport.
+    In _Journal of Open Aviation Science_.
+    doi:[10.59490/joas.2023.7147](https://doi.org/10.59490/joas.2023.7147)
+
+    Planès, T., Salgas, A., Pollet, F., & Delbecq, S. (2025).
+    Simulation and analyses of ICAO transition scenarios using the AeroMAPS open-source framework.
+    In _Towards Sustainable Aviation Summit (TSAS 2025)_.
+    doi:[10.60711/TSAS25.20250216.7026241092200926](https://doi.org/10.60711/TSAS25.20250216.7026241092200926)
+
+    Pollet, F., Planès, T., & Delbecq, S. (2024).
+    A Comprehensive Methodology for Performing Prospective Life Cycle Assessments of Future Air Transport Scenarios.
+    In _34th Congress of the International Council of the Aeronautical Sciences_.
+    hal:[hal-04703961](https://hal.science/hal-04703961v1)
+    """
+
+    _statistical_data = {}
+    with resources.open_text("jetfuelburn.data.AeroMaps", "aeromaps_data.csv") as file:
+        csv_reader = csv.DictReader(file)
+        for row in csv_reader:
+            year = int(row.pop('year'))
+            _statistical_data[year] = {k: float(v) for k, v in row.items()}
+
+
+    @staticmethod
+    def available_years() -> list[int]:
+        """
+        Returns a sorted list of available years included in the model.
+        """
+        return sorted(aeromaps._statistical_data.keys())
+
+
+    @staticmethod
+    def available_aircraft(year: int) -> list[str]:
+        """
+        Given a year, returns a sorted list of available aircraft types 
+        (e.g., ['long_range', 'medium_range', 'short_range']).
+        """
+        if year not in aeromaps._statistical_data:
+            return []
+        return sorted(aeromaps._statistical_data[year].keys())
+
+
+    @staticmethod
+    @ureg.check(
+        None, # acft_type
+        None, # year
+        '[length]',
+    )
+    def calculate_fuel_consumption(
+        acft_type: str,
+        year: int,
+        R: float,
+    ) -> float:
+        r"""
+        $$
+        F = R / E_{spec}
+        $$
+        """
+        if R.magnitude < 0:
+            raise ValueError(f"Range must not be negative.")
+        else:
+            R = R.to('km')
+        if year not in aeromaps._statistical_data:
+            raise ValueError(f"Year '{year}' not found in model data. Please select one of the following: {aeromaps.available_years()}")
+        if acft_type not in aeromaps._statistical_data[year]:
+            raise ValueError(f"Aircraft type '{acft_type}' not found in model data for year '{year}'. Please select one of the following: {aeromaps.available_aircraft(year)}")
+        
[…]
