### 2026-01-14 4146dfa
added aeromaps class
