+++ src/jetfuelburn/reducedorder.py
@@ -1,9 +1,241 @@
+# %%
 import csv
 import json
 import math
 from importlib import resources
 from jetfuelburn import ureg
 from jetfuelburn.utility.physics import _calculate_dynamic_pressure
+
+
+class sacchi_etal:
+    r"""
+    This class implements the reduced-order fuel burn model of Sacchi et al. (2023):
+
+    ![Sacchi et al. Visual Abstract](../_static/reduced_order_sacchi.svg)
+
+    In this model, fuel burn calculations are based on a weight-based power law.
+    The regression coefficients were obtained by fitting mission parameters to fuel burn data obtained
+    from the European Environmental Agency [EMEP/EEA air pollutant emission inventory guidebook - 2013](https://www.eea.europa.eu/en/analysis/publications/emep-eea-guidebook-2013) 
+    which in turn is based on a combination of Eurocontrol BADA flight trajectory simulation model data and the ICAO Engine Emissions Databank (EDB).
+    Cruise segment fuel burn is calculated using the [Eurocontrol BADA](https://www.eurocontrol.int/model/bada) model. 
+    Climb and descent segment fuel burn is calculated using the values provided in the ICAO Engine Emissions Databank.
+
+    Fuel burn is calculated according to Supplement 1 of Sacchi et al. (2023):
+    
+    $$
+        F_{total} = \eta(t) \cdot \left( 
+        \underbrace{a_{LTO} \cdot W^{b_{LTO}}}_{\text{LTO}} + 
+        \underbrace{a_{CD} \cdot W^{b_{CD}}}_{\text{Climb/Descent}} + 
+        \underbrace{(a_{cruise} \cdot W^{b_{cruise}} + c_{cruise}) \cdot R}_{\text{cruise}} 
+        \right)
+    $$
+
+    where:
+
+    | Symbol         | Dimension         | Description                                                            |
+    |----------------|-------------------|------------------------------------------------------------------------|
+    | $F_{total}$    | [mass]            | Total fuel consumption of the flight                                   |
+    | $W$            | [mass]            | Operating empty weight of the aircraft                                 |
+    | $R$            | [length]          | Range of the aircraft (=mission distance)                              |
+    | $t$            | [time]            | Scenario year                                                          |
+    | $a_{LTO}$      |                   | Coefficient 'a' for LTO phase                                          |
+    | $b_{LTO}$      |                   | Exponent 'b' for LTO phase                                             |
+    | $a_{CD}$       |                   | Coefficient 'a' for Climb/Descent phase                                |
+    | $b_{CD}$       |                   | Exponent 'b' for Climb/Descent phase                                   |
+    | $a_{cruise}$   |                   | Coefficient 'a' for Cruise phase                                     |
+    | $b_{cruise}$   |                   | Exponent 'b' for Cruise phase                                          |
+    | $c_{cruise}$   |                   | Constant 'c' for Cruise phase (fuel per km)                            |
+    | $r_{hist}$     | [dimensionless]   | Historical annual improvement rate (2004-2018)                         |
+    | $r_{fut}$      | [dimensionless]   | Future annual improvement rate (post-2018)                             |
+
+    Efficiency $\eta(t)$ for years after 2018 is defined through two constant annual improvement factors for the periods 2004-2018 and post-2018:
+
+    $$
+        \eta(t) = (1 + r_{hist})^{14} \cdot (1 + r_{fut})^{(t - 2018)}
+    $$
+
+    where:
+
[…]
