+++ CHANGES.md
@@ -3,6 +3,13 @@
 The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `1.2.1` (19. January 2026)
+
+### Improvements
+
+- Extend official support to [Python 3.10-3.14](https://devguide.python.org/versions/).
+- Updated GitHub Actions workflows to use latest action versions.
+
 ## `1.2.0` (14. January 2026)
 
 ### Improvements
