+++ .github/workflows/test_package.yml
@@ -29,10 +29,10 @@ jobs:
         runs-on: ubuntu-latest
         strategy:
             matrix:
-                python-version: ["3.10", "3.11", "3.12", "3.13"]
+                python-version: ["3.10", "3.11", "3.12", "3.13", "3.14"]
 
         steps:
-            - uses: actions/checkout@v5
+            - uses: actions/checkout@v6
             - name: Set up Python ${{ matrix.python-version }}
               uses: actions/setup-python@v6
               with:
