+++ src/jetfuelburn/reducedorder.py
@@ -7,6 +7,156 @@ from jetfuelburn import ureg
 from jetfuelburn.utility.physics import _calculate_dynamic_pressure
 
 
+
+class montlaur_etal:
+    r"""
+    This class implements the reduced-order fuel burn model of Montlaur et al. (2025):
+
+    $$
+        \text{Fuel}_{ASK} = \beta_0 + \frac{\beta_1}{D} + \beta_2 D + \beta_3 S + \beta_4 (D \cdot S)
+    $$
+
+    where:
+
+    | Symbol         | Unit                                | Description                                                    |
+    |----------------|-------------------------------------|----------------------------------------------------------------|
+    | $F/ASK$        | [mass] / ([length] $\cdot$ [seats]) | Fuel per Available Seat Kilometer                              |
+    | $D$            | [length]                            | Flight distance                                                |
+    | $S$            | [dimensionless]                     | Number of available seats                                      |
+    | $\beta_{0..4}$ | varies                              | Regression coefficients                                        |
+
+    The model uses specific coefficients depending on the aircraft classification:
+
+    | Coefficient            | Small Aircraft (Eq 2)                 | Large Aircraft (Eq 3)                 |
+    |------------------------|---------------------------------------|---------------------------------------|
+    | $\beta_0$ (Intercept)  | $34.67$                               | $0.7361$                              |
+    | $\beta_1$ ($1/D$)      | $6608$                                | $6651$                                |
+    | $\beta_2$ ($D$)        | $-1.196 \times 10^{-3}$               | $5.989 \times 10^{-4}$                |
+    | $\beta_3$ ($S$)        | $-0.1354$                             | $6.152 \times 10^{-2}$                |
+    | $\beta_4$ ($D \cdot S$)| $1.338 \times 10^{-5}$                | $-1.014 \times 10^{-6}$               |
+
+    References
+    ----------
+    Montlaur, A., Trapote-Barreira, C., & Delgado, L. (2025).
+    Analytical models of flight fuel consumption and Non-CO2 emissions as a function of aircraft capacity.
+    _Applied Sciences_, 15(17), 9688. 
+    doi:[10.3390/app15179688](https://doi.org/10.3390/app15179688)
+
+    See Also
+    --------
+    [`emissions_fuel_estimation.emissions_fuel_model.compute_fuel_ask()`](https://github.com/luis-uow/emissions_fuel_estimation/blob/6d415b38e6089f29ff46a86646db0d5cc04744ba/emissions_fuel_model.py#L9)
+
+    Notes
+    -----
+    The model differentiates between "Small Aircraft" and "Large Aircraft" based on seat capacity and flight distance, 
+    as defined in Table 1 of Montlaur et al. (2020):
+
+    | Model Designation | Seat Capacity ($S$)         | Flight Distance ($D$)         | Representative Aircraft                                              |
+    |-------------------|-----------------------------|-------------------------------|------------------------------------------------------------------------|
+    | "Model D, E"      | $50 \le S \le 172$          | $100 \le D \le 5000$ km       | **CAT D:** B38M, A319<br>**CAT E:** CRJ2, CRJ9, E190                  |
+    | "Model B, D"      | $172 \le S \le 350$         | $200 \le D \le 12000$ km      | **CAT B:** A359, A339, B77W, B788, B789<br>**CAT D:** A21N, A20N, B38M |
+
+    Example
+    -------
+    ```pyodide install='jetfuelburn'
+    import jetfuelburn
+    from jetfuelburn import ureg
+    from jetfuelburn.reducedorder import montlaur_etal
[…]
