### 2026-01-19 90ea9d8
updated montlaur_etal class
