+++ src/jetfuelburn/reducedorder.py
@@ -7,13 +7,21 @@ from jetfuelburn import ureg
 from jetfuelburn.utility.physics import _calculate_dynamic_pressure
 
 
-
 class montlaur_etal:
     r"""
     This class implements the reduced-order fuel burn model of Montlaur et al. (2025):
 
+    ![Montlaur et al. Visual Abstract](../_static/reduced_order_montlaur_etal.svg)
+
+    In this model, fuel burn calculations are based on a regression model.
+    The regression coefficients were obtained by fitting mission parameters to fuel burn data obtained
+    from the Eurocontrol IMPACT model.
+
+    Fuel burn is calculated according to the function `compute_fuel_ask()` in [the code supplement](https://github.com/luis-uow/emissions_fuel_estimation) 
+    to Montlaur et al. (2025):
+    
     $$
-        \text{Fuel}_{ASK} = \beta_0 + \frac{\beta_1}{D} + \beta_2 D + \beta_3 S + \beta_4 (D \cdot S)
+        F/ASK = \beta_0 + \frac{\beta_1}{D} + \beta_2 D + \beta_3 S + \beta_4 (D \cdot S)
     $$
 
     where:
@@ -23,24 +31,19 @@ class montlaur_etal:
     | $F/ASK$        | [mass] / ([length] $\cdot$ [seats]) | Fuel per Available Seat Kilometer                              |
     | $D$            | [length]                            | Flight distance                                                |
     | $S$            | [dimensionless]                     | Number of available seats                                      |
-    | $\beta_{0..4}$ | varies                              | Regression coefficients                                        |
-
-    The model uses specific coefficients depending on the aircraft classification:
-
-    | Coefficient            | Small Aircraft (Eq 2)                 | Large Aircraft (Eq 3)                 |
-    |------------------------|---------------------------------------|---------------------------------------|
-    | $\beta_0$ (Intercept)  | $34.67$                               | $0.7361$                              |
-    | $\beta_1$ ($1/D$)      | $6608$                                | $6651$                                |
-    | $\beta_2$ ($D$)        | $-1.196 \times 10^{-3}$               | $5.989 \times 10^{-4}$                |
-    | $\beta_3$ ($S$)        | $-0.1354$                             | $6.152 \times 10^{-2}$                |
-    | $\beta_4$ ($D \cdot S$)| $1.338 \times 10^{-5}$                | $-1.014 \times 10^{-6}$               |
+    | $\beta_{0-4}$  | varies                              | Regression coefficients                                        |
 
     References
     ----------
     Montlaur, A., Trapote-Barreira, C., & Delgado, L. (2025).
     Analytical models of flight fuel consumption and Non-CO2 emissions as a function of aircraft capacity.
     _Applied Sciences_, 15(17), 9688. 
-    doi:[10.3390/app15179688](https://doi.org/10.3390/app15179688)
+    doi:[10.3390/app15179688](https://doi.org/10.3390/app15179688)  
+
+    Montlaur, A., Delgado, L., & Trapote-Barreira, C. (2021). 
+    Analytical models for CO2 emissions and travel time for short-to-medium-haul flights considering available seats. 
+    _Sustainability_, 13(18), 10401. 
+    doi:[10.3390/su131810401](https://doi.org/10.3390/su131810401)
 
     See Also
     --------
@@ -70,19 +73,19 @@ class montlaur_etal:
     """
 
     _REGRESSION_COEFFICIENTS_MODEL_B_D = {
-        "intercept": 34.67,
-        "inv_dist": 6608.0,
-        "dist": -1.196e-3,
-        "seats": -0.1354,
-        "interaction": 1.338e-5
[…]
