### 2026-01-22 5859d50
added outline of functions
