+++ src/jetfuelburn/data/Airports/__init__.py
@@ -1,0 +1,1 @@
+# https://github.com/ip2location/ip2location-iata-icao
