### 2026-01-22 7cfb375
added outline of functions
