### 2026-01-22 5e0b5f5
updated functions
