### 2026-01-22 01286ac
updated functions
