### 2026-01-22 42d7f03
added ipyleaflet to docs requirements
