### 2026-01-22 8781020
added ipyleaflet to docs requirements
