### 2026-01-23 fad2503
added lazy loading to airport distance calculator
