### 2026-01-23 5912d8d
updated distance calculation functions
