+++ CHANGES.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## `1.5.2` (23. January 2026)
 
 ### Fixes
+
 - Updated the file I/O handling in the `jetfuelburn.utility.geography` module to avoid repeated file openings when calculating Great Circle distances.
 
 ## `1.5.1` (22. January 2026)
