### 2026-01-29 36985a0
updated classical breguet function docstrings
