+++ mkdocs.yaml
@@ -84,6 +84,7 @@ markdown_extensions:
   - pymdownx.snippets
   - pymdownx.highlight:
       use_pygments: false
+  - attr_list
 
 extra_javascript:
   - _javascripts/katex.js
