### 2026-01-29 80b3d6a
updated classical breguet function docstrings
