### 2026-01-29 d7cd7a6
updated classical breguet function docstrings
