### 2026-01-29 d292dfb
updated functions
