### 2026-01-29 50823fd
updated functions
