+++ src/jetfuelburn/utility/physics.py
@@ -45,8 +45,8 @@ def _calculate_atmospheric_temperature(altitude):
 
     Returns
     -------
-    ureg.Quantity (float) [temperature]
-        Air temperature (°C).
+    ureg.Quantity [temperature]
+        Air temperature in degrees Celsius.
 
     Raises
     ------
@@ -206,24 +206,13 @@ def _calculate_atmospheric_density(altitude):
 
     Parameters
     ----------
-    altitude : float [distance]
-        Altitude above sea level
-
-    Notes
-    -----
-    Compare also the function
-    [`atmos()` in `openap/extra/aero.py`](https://github.com/TUDelft-CNS-ATM/openap/blob/39619977962fe6b4a86ab7efbefa70890eecfe36/openap/extra/aero.py#L48C5-L48C10)
-    by [Junzi Sun](https://github.com/junzis). Note that `jetfuelburn` function has been re-written completely and does not build on the `openap` code.
-
-    References
-    --------
-    - Temperature: [Eqn.(1.6) in Sadraey (2nd Edition, 2024)](https://doi.org/10.1201/9781003279068)
-    - [Density Equations entry on Wikipedia](https://en.wikipedia.org/wiki/Barometric_formula#Density_equations)
+    altitude : Quantity [length]
+        Altitude above sea level (0 to 20,000 meters).
 
     Returns
     -------
-    ureg.Quantity (float) [mass/volume]
-        Air density (kg/m³).
+    ureg.Quantity [mass/volume]
+        Air density in kg/m³.
 
     Raises
     ------
@@ -274,10 +263,9 @@ def _calculate_atmospheric_density(altitude):
 
     if altitude <= 11000 * ureg.m:
         current_temp = _calculate_atmospheric_temperature(altitude).to(ureg.K)
-        exponent = (g * M / (R * lapse_rate)) - 1
+        exponent = (-g / (R * lapse_rate)) - 1
         rho = rho_0 * (current_temp / temperature_0) ** exponent
     else:
-        # This part was already correct
         rho = rho_1 * math.exp(
             -g * M * (altitude - 11000 * ureg.m) / (R * temperature_stratosphere)
         )
@@ -285,6 +273,7 @@ def _calculate_atmospheric_density(altitude):
     return rho.to(ureg.kg / ureg.m**3)
 
     return rho.to(ureg.kg / ureg.m**3)
+
 
 
 @ureg.check(
@@ -322,7 +311,7 @@ def _calculate_dynamic_pressure(
     References
     --------
     - [Dynamic Pressure entry on Wikipedia](https://en.wikipedia.org/wiki/Dynamic_pressure)
-    - [Eqn.(2.63) in Young (2018)](https://doi.org/10.1002/9781118534786)
+    - Eqn. (2.63) in in Young, T. M. (2018). Performance of the Jet Transport Airplane. _John Wiley & Sons_. doi:[10.1002/9781118534786](https://doi.org/10.1002/9781118534786)
 
     Parameters
     ----------
