### 2026-01-29 4db1600
updated functions
