### 2026-01-29 9008607
added speed of sound function draft
