+++ src/jetfuelburn/utility/physics.py
@@ -45,63 +45,6 @@ def _calculate_atmospheric_temperature(altitude):
 
     Returns
     -------
-    ureg.Quantity [temperature]
-        Air temperature in degrees Celsius.
-
-    Raises
-    ------
-    ValueError
-        If altitude is less than 0 m or greater than 20,000 m.
-
-    Example
-    -------
-    ```pyodide install='jetfuelburn'
-    import jetfuelburn
-    from jetfuelburn import ureg
-    from jetfuelburn.utility.physics import _calculate_atmospheric_temperature
-    _calculate_atmospheric_temperature(altitude=10000*ureg.m)
-    ```
-    """
-    if altitude < 0 * ureg.m:
-        raise ValueError("Altitude must not be < 0.")
-    elif altitude > 20000 * ureg.m:
-        raise ValueError("Altitude must not be > 20000 m.")
-
-    temperature_0 = 288.15 * ureg.K            # Sea-level standard temperature
-    lapse_rate = 0.0065 * (ureg.K / ureg.m)    # Temperature lapse rate
-    temperature_stratosphere = 216.65 * ureg.K # Constant temp in lower stratosphere
-
-    if altitude <= 11000 * ureg.m:
-        temperature = temperature_0 - lapse_rate * altitude
-    else:
-        temperature = temperature_stratosphere
-
-    return temperature.to(ureg.celsius)
-
-
-@ureg.check('[length]')
-def _calculate_atmospheric_density(altitude):
-    r"""
-    Computes the air density as a function of altitude up to 20,000 meters 
-    based on the International Standard Atmosphere (ISA):
-
-    <img src="https://upload.wikimedia.org/wikipedia/commons/a/a8/International_Standard_Atmosphere.svg" width="250">
-
-    Density in the Troposphere (<= 11,000 m) is calculated according to
-
-    Temperature in the lower Stratosphere (> 11,000 m) is assumed constant at -56.5°C (216.65 K).
-
-    References
-    ----------
-    Eqn.(1.6) in Sadraey, M. H. (2017). Aircraft Performance: An Engineering Approach . _CRC Press_. doi:[10.1201/9781003279068](https://doi.org/10.1201/9781003279068)
-
-    Parameters
-    ----------
-    altitude : Quantity [length]
-        Altitude above sea level (0 to 20,000 meters).
-
-    Returns
-    -------
     ureg.Quantity (float) [temperature]
         Air temperature (°C).
 
@@ -211,13 +154,6 @@ def _calculate_atmospheric_density(altitude):
 
     Returns
     -------
-    ureg.Quantity [mass/volume]
-        Air density in kg/m³.
-
-    Raises
-    ------
-    ValueError
-        If altitude is less than 0 m or greater than 20,000 m.
     ureg.Quantity (float) [mass/volume]
         Air density (kg/m³).
 
@@ -272,9 +208,6 @@ def _calculate_atmospheric_density(altitude):
 
     return rho.to(ureg.kg / ureg.m**3)
 
-    return rho.to(ureg.kg / ureg.m**3)
-
-
 
 @ureg.check(
     '[speed]',
@@ -308,10 +241,13 @@ def _calculate_dynamic_pressure(
     --------
     [Dynamic Pressure entry on Wikipedia](https://en.wikipedia.org/wiki/Dynamic_pressure)
 
+    See Also
+    --------
+    [Dynamic Pressure entry on Wikipedia](https://en.wikipedia.org/wiki/Dynamic_pressure)
+
     References
     --------
-    - [Dynamic Pressure entry on Wikipedia](https://en.wikipedia.org/wiki/Dynamic_pressure)
[…]
