### 2026-01-29 a9095c2
added speed of sound function draft
