+++ src/jetfuelburn/utility/physics.py
@@ -37,8 +37,8 @@ def _calculate_atmospheric_temperature(altitude):
 
     Returns
     -------
-    ureg.Quantity [temperature]
-        Air temperature in degrees Celsius.
+    ureg.Quantity (float) [temperature]
+        Air temperature (°C).
 
     Raises
     ------
@@ -120,8 +120,8 @@ def _calculate_atmospheric_density(altitude):
 
     Returns
     -------
-    ureg.Quantity [mass/volume]
-        Air density in kg/m³.
+    ureg.Quantity (float) [mass/volume]
+        Air density (kg/m³).
 
     Raises
     ------
@@ -163,7 +163,6 @@ def _calculate_atmospheric_density(altitude):
     return rho.to(ureg.kg / ureg.m**3)
 
 
-
 @ureg.check(
     '[speed]',
     '[length]'
@@ -188,10 +187,13 @@ def _calculate_dynamic_pressure(
     | $V$    | [speed]         | aircraft speed      |
 
 
+    See Also
+    --------
+    [Dynamic Pressure entry on Wikipedia](https://en.wikipedia.org/wiki/Dynamic_pressure)
+
     References
     --------
-    - [Dynamic Pressure entry on Wikipedia](https://en.wikipedia.org/wiki/Dynamic_pressure)
-    - Eqn. (2.63) in in Young, T. M. (2018). Performance of the Jet Transport Airplane. _John Wiley & Sons_. doi:[10.1002/9781118534786](https://doi.org/10.1002/9781118534786)
+    Eqn. (2.63) in in Young, T. M. (2018). Performance of the Jet Transport Airplane. _John Wiley & Sons_. doi:[10.1002/9781118534786](https://doi.org/10.1002/9781118534786)
 
     Parameters
     ----------
@@ -202,8 +204,8 @@ def _calculate_dynamic_pressure(
 
     Returns
     -------
-    float
-        Dynamic pressure [Pa]
+    ureg.Quantity (float) [pressure]
+        Dynamic pressure (Pascals).
 
     Example
     -------
@@ -217,7 +219,7 @@ def _calculate_dynamic_pressure(
     )
     ```
     """
-    air_density = _calculate_atmospheric_conditions(altitude)['density']
+    air_density = _calculate_atmospheric_density(altitude)
     dynamic_pressure = 0.5 * air_density * speed ** 2
     return dynamic_pressure.to(ureg.Pa)
 
@@ -254,15 +256,23 @@ def _calculate_aircraft_velocity(
     altitude : float [length]
         Flight altitude above sea level
 
-    References
+    Returns
+    -------
+    ureg.Quantity (float) [speed]
+        Aircraft velocity (km/h).
+
+    See Also
     --------
-    - [Mach Number entry on Wikipedia](https://en.wikipedia.org/wiki/Mach_number#Calculation)
-    - [Eqn.(1.33)-(1.34) in Sadraey (2nd Edition, 2024)](https://doi.org/10.1201/9781003279068)
+    [Mach Number entry on Wikipedia](https://en.wikipedia.org/wiki/Mach_number#Calculation)
+
+    References
+    ----------
+    Eqn.(1.33)-(1.34) Sadraey, M. H. (2017). Aircraft Performance: An Engineering Approach . _CRC Press_. doi:[10.1201/9781003279068](https://doi.org/10.1201/9781003279068)
 
     Returns
     -------
-    float
-        Aircraft velocity [km/h]
+    ureg.Quantity (float) [speed]
+        Aircraft velocity in kilometers per hour.
 
     Example
     -------
@@ -276,11 +286,70 @@ def _calculate_aircraft_velocity(
     )
     ```
     """
-
[…]
