### 2026-01-29 b78782b
added speed of sound function draft
