+++ CHANGES.md
@@ -1,6 +1,6 @@
 # `jetfuelburn` Changelog
 
-The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## `2.0.0` (29. January 2026)
