+++ src/jetfuelburn/data/OpenAP/__init__.py
@@ -1,0 +1,14 @@
+# https://github.com/TUDelft-CNS-ATM/openap/tree/master/openap/data/aircraft
+# https://github.com/TUDelft-CNS-ATM/openap/tree/master/openap/data/dragpolar
+
+# Licensing Note:
+# The LGPL3 license of these repositories does not apply to the re-use of aerodynamic data, since in Switzerland 
+# only "intellectual creations with individual character" are protected by copyright law.
+# Only Swiss copyright law applies, since the Netherlands and Switzerland are parties to the Berne Convention
+# which states that the extent of copyright protection is governed exclusively by the laws of the country where protection is claimed.
+
+# References:
+# Federal Act on Copyright and Related Rights (CopA), Art. 2, Par. 1
+# https://www.fedlex.admin.ch/eli/cc/1993/1798_1798_1798/en#tit_2
+# Berne Convention
+# https://www.fedlex.admin.ch/eli/cc/1955/1092_1114_1122/de
