+++ src/jetfuelburn/utility/aerodynamics.py
@@ -1,0 +1,4 @@
+import math
+from jetfuelburn import ureg
+
+
