+++ src/jetfuelburn/data/USDOT/__init__.py
@@ -1,10 +1,17 @@
-"""
-# Files
+# https://www.transtats.bts.gov/Fields.asp?gnoyr_VQ=FIH
 
-## US Department of Transportation (DOT)
+# Licensing Note:
+# This data is a work prepared by an officer or employee of the United States Government and 
+# is therefore not subject to copyright protection (Public Domain) under 17 U.S.C. § 105.
+# Furthermore, in Switzerland, only "intellectual creations with individual character" 
+# are protected by copyright law, which excludes factual statistical data.
+# Only Swiss copyright law applies, since the USA and Switzerland are parties to the Berne Convention
+# which states that the extent of copyright protection is governed exclusively by the laws of the country where protection is claimed.
 
-### `L_AIRCRAFT_TYPE.csv`
-
-- Source: https://www.transtats.bts.gov/Fields.asp?gnoyr_VQ=FIH
-- Description: Air Carrier Summary: T2: U.S. Air Carrier TRAFFIC And Capacity Statistics by Aircraft Type, Field: Aircraft Type
-"""
+# References:
+# 17 U.S. Code § 105 - United States Government works
+# https://www.law.cornell.edu/uscode/text/17/105
+# Federal Act on Copyright and Related Rights (CopA), Art. 2, Par. 1
+# https://www.fedlex.admin.ch/eli/cc/1993/1798_1798_1798/en#tit_2
+# Berne Convention
+# https://www.fedlex.admin.ch/eli/cc/1955/1092_1114_1122/de
