### 2026-01-31 a240159
added function and test draft
