+++ src/jetfuelburn/utility/physics.py
@@ -2,24 +2,19 @@ import math
 from jetfuelburn import ureg
 
 
-@ureg.check(
-    '[length]' # altitude
-)
-def _calculate_atmospheric_conditions(altitude: float) -> dict[float, float]:
+@ureg.check('[length]')
+def _calculate_atmospheric_temperature(altitude):
     r"""
-    Computes the air density and temperature as a function of altitute, for altitudes up to 20,000 meters.
-
-    All calculations are based on the ISA (International Standard Atmosphere):
+    Computes the air temperature as a function of altitude up to 20,000 meters 
+    based on the International Standard Atmosphere (ISA):
 
     <img src="https://upload.wikimedia.org/wikipedia/commons/a/a8/International_Standard_Atmosphere.svg" width="250">
-
-    Temperature in the Troposphere is calculated according to:
-
+    
+    Temperature in the Troposphere (<= 11,000 m) is calculated according to
+    
     $$
-        T=T_0 - L * h
+    T = T_0 - L \cdot h
     $$
-
-    in the lower Stratosphere, it is simply -56.5°C.
 
     where:
 
@@ -28,82 +23,146 @@ def _calculate_atmospheric_conditions(altitude: float) -> dict[float, float]:
     | $T$    | [temperature]         | temperature at altitude $h$         | N/A          |
     | $T_0$  | [temperature]         | temperature at sea level            | 288.15 K     |
     | $L$    | [temperature/length]  | temperature lapse rate (Troposphere)| 0.0065 K/m   |
-    | $h$    | [length]              | altitude                            | N/A          |
 
-    Density is calculated using the _exponential approximation_:
+    Temperature in the lower Stratosphere (> 11,000 m) is assumed constant at -56.5°C (216.65 K).
 
-    $$
-    \rho = \rho_0 * e^{-\frac{g*h}{R*T}}
-    $$
-
-    where:
-
-    | Symbol  | Dimension             | Description                         | Value        |
-    |---------|-----------------------|-------------------------------------|--------------|
-    | $\rho$  | [mass/volume]         | air density at altitude $h$         | N/A          |
-    | $\rho_0$| [mass/volume]         | air density at sea level            | 1.225 kg/m³  |
-    | $g$     | [acceleration]        | acceleration due to gravity         | 9.80665 m/s² |
-    | $R$     | [specific gas constant| specific gas constant for air       | 287 J/(kg*K) |
-    | $T$     | [temperature]         | temperature at altitude $h$         | N/A          |
+    References
+    ----------
+    Eqn.(1.6) in Sadraey, M. H. (2017). Aircraft Performance: An Engineering Approach . _CRC Press_. doi:[10.1201/9781003279068](https://doi.org/10.1201/9781003279068)
 
     Parameters
     ----------
-    altitude : float [distance]
-        Altitude above sea level
-
-    Notes
-    -----
-    Compare also the function
-    [`atmos()` in `openap/extra/aero.py`](https://github.com/TUDelft-CNS-ATM/openap/blob/39619977962fe6b4a86ab7efbefa70890eecfe36/openap/extra/aero.py#L48C5-L48C10)
-    by [Junzi Sun](https://github.com/junzis). Note that `jetfuelburn` function has been re-written completely and does not build on the `openap` code.
-
-    References
-    --------
-    - Temperature: [Eqn.(1.6) in Sadraey (2nd Edition, 2024)](https://doi.org/10.1201/9781003279068)
-    - [Density Equations on Wikipedia](https://en.wikipedia.org/wiki/Barometric_formula#Density_equations)
[…]
