+++ CHANGES.md
@@ -3,6 +3,12 @@
 The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `2.0.0` (02. February 2026)
+
+### Breaking Changes
+
+- Updated the `utility.physics` module XXXX
+
 ## `1.5.2` (23. January 2026)
 
 ### Fixes
