### 2026-01-31 2846c31
added function and test draft
