+++ tests/test_physics.py
@@ -1,36 +1,132 @@
 import pytest
+from jetfuelburn import ureg
+from jetfuelburn.utility.physics import (
+    _calculate_atmospheric_temperature,
+    _calculate_atmospheric_density,
+    _calculate_dynamic_pressure,
+    _calculate_aircraft_velocity,
+    _calculate_speed_of_sound,
+)
 
 from jetfuelburn.utility.tests import approx_with_units
-from jetfuelburn.utility.physics import _calculate_atmospheric_conditions, _calculate_aircraft_velocity
-from .fixtures.physics import atmospheric_case_fixture, mach_speed_case_fixture
+
+from .fixtures.physics import (
+    atmospheric_case_fixture,
+    mach_speed_case_fixture,
+)
 
 
-def test_calculate_atmospheric_conditions(atmospheric_case_fixture):
-    make_atmospheric_case, altitudes = atmospheric_case_fixture
+def test_calculate_atmospheric_temperature_valid(atmospheric_case_fixture):
+    """
+    [cite_start]Validates temperature calculation against Table A.1 data points[cite: 4, 36, 54].
+    """
+    make_case, altitudes = atmospheric_case_fixture
     
     for altitude in altitudes:
-        input_data, expected_output = make_atmospheric_case(altitude)
-        calculated_output = _calculate_atmospheric_conditions(altitude=input_data)
-
+        _, expected_data = make_case(altitude)
+        
+        # We need to test the temperature function specifically
+        # Note: The function expects altitude in meters, but handles Pint quantities correctly.
+        # We use a relaxed absolute tolerance because the PDF table rounds to 2 decimal places.
+        result = _calculate_atmospheric_temperature(altitude)
+        
         assert approx_with_units(
-            value_check=calculated_output['density'],
-            value_expected=expected_output['density'],
-            rel=1e-2
+            result, 
+            expected_data['temperature'], 
+            abs=1e-1 # 0.1 degree tolerance due to table rounding
         )
 
-
-def test_calculate_aircraft_velocity(mach_speed_case_fixture):
-    make_mach_speed_case, altitudes = mach_speed_case_fixture
+def test_calculate_atmospheric_density_valid(atmospheric_case_fixture):
+    """
+    [cite_start]Validates density calculation against Table A.1 data points[cite: 4, 36, 54].
+    """
+    make_case, altitudes = atmospheric_case_fixture
     
     for altitude in altitudes:
-        input_data, expected_output = make_mach_speed_case(altitude)
-        output = _calculate_aircraft_velocity(
-            mach_number=input_data['mach_number'],
+        _, expected_data = make_case(altitude)
+        
+        result = _calculate_atmospheric_density(altitude)
+        
+        # Density table values have 4 decimal places.
+        assert approx_with_units(
+            result, 
+            expected_data['density'], 
+            rel=1e-2 # 1% relative tolerance to account for constant variations
+        )
+
+def test_calculate_aircraft_velocity(mach_speed_case_fixture):
+    """
+    [cite_start]Validates Mach to TAS conversion using speed of sound data derived from Table A.1[cite: 4, 36, 54].
+    """
+    make_case, altitudes = mach_speed_case_fixture
+    
+    for altitude in altitudes:
+        input_data, expected_speed = make_case(altitude)
+        
+        result = _calculate_aircraft_velocity(
+            mach_number=input_data['mach_number'], 
             altitude=input_data['altitude']
         )
         
-        assert approx_with_units(
-            value_check=output,
-            value_expected=expected_output,
-            rel=1e-2
-        )
[…]
