### 2026-01-31 a4b9f4e
added function and test draft
