+++ tests/data/__init__.py
@@ -1,0 +1,2 @@
+# https://onlinelibrary.wiley.com/doi/pdf/10.1002/9781118534786.app1
+
