### 2026-02-01 139658a
added isa table and removed python 3.10
