+++ tests/data/__init__.py
@@ -1,2 +1,12 @@
-# https://onlinelibrary.wiley.com/doi/pdf/10.1002/9781118534786.app1
+# https://doi.org/10.1002/9781118534786.app1
 
+# Licensing Note:
+# In Switzerland, only "intellectual creations with individual character" are protected by copyright law.
+# Only Swiss copyright law applies, since the United States and Switzerland are parties to the Berne Convention
+# which states that the extent of copyright protection is governed exclusively by the laws of the country where protection is claimed.
+
+# References:
+# Federal Act on Copyright and Related Rights (CopA), Art. 2, Par. 1
+# https://www.fedlex.admin.ch/eli/cc/1993/1798_1798_1798/en#tit_2
+# Berne Convention
+# https://www.fedlex.admin.ch/eli/cc/1955/1092_1114_1122/de
