### 2026-02-01 9210225
added isa table and removed python 3.10
