+++ tests/test_physics.py
@@ -1,4 +1,7 @@
+# %%
+from pathlib import Path
 import pytest
+import csv
 from jetfuelburn import ureg
 from jetfuelburn.utility.physics import (
     _calculate_atmospheric_temperature,
@@ -10,123 +13,195 @@ from jetfuelburn.utility.physics import (
 
 from jetfuelburn.utility.tests import approx_with_units
 
-from .fixtures.physics import (
-    atmospheric_case_fixture,
-    mach_speed_case_fixture,
-)
+def load_isa_csv():
+    r"""
+    Helper function to load a CSV file containing 
+    parameters of the International Standard Atmosphere (ISA).
+    Returns a list of dictionaries where values are float-converted.
+
+    References
+    ----------
+    "International Standard Atmosphere (ISA) Table" in Young, T. M. (2017).
+    _Performance of the Jet Transport Airplane: Analysis Methods, Flight Operations, and Regulations._
+    John Wiley & Sons.
+    doi:[10.1002/9781118534786.app1](https://doi.org/10.1002/9781118534786.app1)
+
+    See Also
+    --------
+    [International Standard Atmosphere (ISA) entry on Wikipedia](https://en.wikipedia.org/wiki/International_Standard_Atmosphere)
+
+    Returns
+    -------
+    List[Dict[str, float]]
+        A list of dictionaries with ISA parameters, where each dictionary corresponds to a row in the CSV file.  
+        Of the form:
+        ```
+        [
+            {
+                'H_ft': 0.0,
+                'H_m': 0.0,
+                'theta': 1.0,
+                'T_K': 288.15,
+                'T_C': 15.0,
+                'delta': 1.0,
+                'P_N/m^2': 101325.0,
+                'P_lb/ft^2': 2116.21,
+                'sigma': 1.0,
+                'rho_kg/m^3': 1.225,
+                'rho_slug/ft^3': 0.002377,
+                'a_m/s': 340.3,
+                'a_ft/s': 1116.0,
+                'a_kt': 661.5
+            },
+            ...
+        ]
+        ```
+    """
+    csv_path = Path(__file__).parent / "data" / "isa.csv"
+    data = []
+    with open(csv_path, mode='r') as f:
+        reader = csv.DictReader(f)
+        for row in reader:
+            clean_row = {k: float(v) for k, v in row.items()}
+            data.append(clean_row)
+    return data
 
 
-def test_calculate_atmospheric_temperature_valid(atmospheric_case_fixture):
-    """
-    [cite_start]Validates temperature calculation against Table A.1 data points[cite: 4, 36, 54].
-    """
-    make_case, altitudes = atmospheric_case_fixture
-    
-    for altitude in altitudes:
-        _, expected_data = make_case(altitude)
-        
-        # We need to test the temperature function specifically
-        # Note: The function expects altitude in meters, but handles Pint quantities correctly.
-        # We use a relaxed absolute tolerance because the PDF table rounds to 2 decimal places.
+ISA_DATA = load_isa_csv()
+
+
+class TestCalculateAtmosphericTemperature:
+    """Test suite for _calculate_atmospheric_temperature."""
+
+    @pytest.mark.parametrize("row", ISA_DATA)
+    def test_values_against_isa_table(self, row):
+        """Checks if calculated temperature matches the ISA table values."""
+        altitude = row['H_m'] * ureg.meter
+        expected = ureg.Quantity(row['T_C'], ureg.degC)
         result = _calculate_atmospheric_temperature(altitude)
         
-        assert approx_with_units(
-            result, 
-            expected_data['temperature'], 
-            abs=1e-1 # 0.1 degree tolerance due to table rounding
-        )
[…]
