+++ .github/workflows/test_package.yml
@@ -29,7 +29,7 @@ jobs:
         runs-on: ubuntu-latest
         strategy:
             matrix:
-                python-version: ["3.10", "3.11", "3.12", "3.13", "3.14"]
+                python-version: ["3.11", "3.12", "3.13", "3.14"]
 
         steps:
             - uses: actions/checkout@v6
