### 2026-02-02 411106a
added interpolation function
