+++ src/jetfuelburn/utility/aerodynamics.py
@@ -1,4 +1,2 @@
 import math
 from jetfuelburn import ureg
-
-
