+++ figures/interpolation/interpolation.py
@@ -1,0 +1,208 @@
+#%%
+# runs code as interactive cell 
+# https://code.visualstudio.com/docs/python/jupyter-support-py
+
+# IMPORTS #######################################
+
+# sys
+import os
+import bisect
+# plotting
+import matplotlib.pyplot as plt
+import matplotlib.patches as patches
+# unit conversion
+cm = 1/2.54 # for inches-cm conversion
+# data science
+import numpy as np
+
+# SETUP #########################################
+
+plt.rcParams.update({
+    "font.family": "Arial",
+    'font.size': 12
+})
+
+# DATA IMPORT ###################################
+
+def _interpolate(
+    x_val: float | int,
+    x_list: list[float | int],
+    y_list: list[float | int]
+):
+    """
+    Finds y_val for a given x_val using linear interpolation.
+    """
+    # Boundary checks removed for visualization purposes
+    
+    i = bisect.bisect_right(x_list, x_val)
+    x0, x1 = x_list[i-1], x_list[i]
+    y0, y1 = y_list[i-1], y_list[i]
+
+    k = (y1 - y0) / (x1 - x0)
+    y_val = y0 + k * (x_val - x0)
+    
+    return y_val, x0, x1, y0, y1, k
+
+# DATA MANIPULATION #############################
+
+# 1. Define known data points (The "List")
+x_known = [0, 10, 20, 30]
+y_known = [10, 50, 90, 60]
+
+# 2. Define target value to interpolate
+x_target = 16
+
+# 3. Calculate interpolation
+y_target, x0, x1, y0, y1, k = _interpolate(x_target, x_known, y_known)
+
+# FIGURE ########################################
+
+# SETUP ######################
+
+fig, ax = plt.subplots(
+    num = 'main',
+    nrows = 1,
+    ncols = 1,
+    dpi = 300,
+    figsize=(20*cm, 10*cm),
+)
+
+# AXIS LIMITS ################
+
+ax.set_xlim(0, 32)
+ax.set_ylim(0, 110)
+
+# TICKS AND LABELS ###########
+
+ax.minorticks_on()
+ax.tick_params(axis='x', which='minor', bottom=True)
+
+# Manually set ticks to match our data points
+x_ticks = x_known
+ax.set_xticks(x_ticks + [x_target])
+ax.set_xticklabels([str(x) for x in x_ticks] + [r"$x_{val}$"])
+
+y_ticks = [0, 20, 40, 60, 80, 100]
+ax.set_yticks(y_ticks)
+
+# GRIDS ######################
+
+ax.grid(which='major', axis='y', linestyle='-', linewidth = 0.5, alpha=0.5)
+ax.grid(which='minor', axis='y', linestyle=':', linewidth = 0.5, alpha=0.5)
+ax.grid(which='major', axis='x', linestyle='-', linewidth = 0.5, alpha=0.5)
+
+# AXIS LABELS ################
+
+ax.set_xlabel('Input (x)')
+ax.set_ylabel("Output (y)")
+
+# PLOTTING ###################
+
+# 1. Plot the "Known" function (Piecewise Linear)
+plt.plot(
+    x_known,
+    y_known,
+    color = 'black',
+    linestyle = '-',
+    linewidth = 1.5,
+    marker = 'o',
+    markersize = 6,
+    label = 'Known Data',
+    zorder = 2
+)
+
+# 2. Plot the interpolated point
+plt.scatter(
+    [x_target],
+    [y_target],
+    color = 'tab:red',
+    marker = '*',
+    s = 200,
+    label = 'Interpolated Point',
+    zorder = 10
+)
+
[…]
