+++ mkdocs.yaml
@@ -22,6 +22,7 @@ nav:
     - Cabin Class Fuel Allocation: 'api/allocation.md'
     - Physics Utility Functions: 'api/physics.md'
     - Geography Utility Functions: 'api/geography.md'
+    - Math Utility Functions: 'api/math.md'
   - API Reference (External):
     - EU Flight Emissions Label: 'api/eu_fel.md'
     - Reduced Order Models: 'api/reducedorder_external.md'
