+++ src/jetfuelburn/data/JSBSim/__init__.py
@@ -1,0 +1,14 @@
+# https://sourceforge.net/p/flightgear/fgaddon/HEAD/tree/trunk/Aircraft/
+# https://github.com/JSBSim-Team/jsbsim
+
+# Licensing Note:
+# The licenses of these repositories do not apply to the re-use of basic aerodynamics data, since in Switzerland 
+# only "intellectual creations with individual character" are protected by copyright law.
+# Only Swiss copyright law applies, since the United States and Switzerland are parties to the Berne Convention
+# which states that the extent of copyright protection is governed exclusively by the laws of the country where protection is claimed.
+
+# References:
+# Federal Act on Copyright and Related Rights (CopA), Art. 2, Par. 1
+# https://www.fedlex.admin.ch/eli/cc/1993/1798_1798_1798/en#tit_2
+# Berne Convention
+# https://www.fedlex.admin.ch/eli/cc/1955/1092_1114_1122/de
