+++ src/jetfuelburn/utility/aerodynamics.py
@@ -1,2 +1,3 @@
 import math
 from jetfuelburn import ureg
+
