### 2026-02-02 715ecc8
added draft of drag polar class
