+++ docs/api/aerodynamics.md
@@ -1,0 +1,1 @@
+::: jetfuelburn.utility.aerodynamics
