### 2026-02-02 f125b14
added draft of drag polar class
