+++ src/jetfuelburn/utility/aerodynamics.py
@@ -3,6 +3,10 @@ import math
 import json
 from importlib import resources
 from jetfuelburn.utility.math import _interpolate
+from jetfuelburn.utility.physics import (
+    _calculate_dynamic_pressure,
+    _calculate_aircraft_velocity
+)
 from jetfuelburn import ureg
 
 
@@ -20,94 +24,63 @@ class jsbsim_drag_polars:
         """
         Returns a sorted list of available ICAO aircraft designators included in the model.
         """
-        return sorted(jsbsim_drag_polar._aircraft_data.keys())
+        return sorted(jsbsim_drag_polars._aircraft_data.keys())
 
     @staticmethod
     @ureg.check(
         None,
-        None,
-        None,
+        '[force]', # lift
+        '[]', # Mach number
+        '[length]', # altitude
     )
     def calculate_drag(
         acft: str,
-        CL: float,
-        Mach: float,
+        L: float,
+        M: float,
+        H: float,
     ) -> dict:
         r"""
-        Given an aircraft name, Lift Coefficient (CL), and Mach number, calculates the total Drag Coefficient (CD).
+        asdf
+        """        
+        if hasattr(L, "magnitude"): L = L.magnitude
+        if hasattr(M, "magnitude"): M = M.magnitude
 
-        The calculation follows a three-step process:
-        1. **Inverse Lift Lookup:** Finds the Angle of Attack ($\alpha$) required for the requested $C_L$.
-        2. **Parasite Drag Lookup:** Finds $C_{D0}$ based on that $\alpha$.
-        3. **Wave Drag Lookup:** Finds compressibility penalties based on Mach.
-
-        | Symbol     | Dimension         | Description                                                            |
-        |------------|-------------------|------------------------------------------------------------------------|
-        | $C_D$      | [dimensionless]   | Total Drag Coefficient                                                 |
-        | $C_L$      | [dimensionless]   | Lift Coefficient                                                       |
-        | $M$        | [dimensionless]   | Mach Number                                                            |
-
-        Parameters
-        ----------
-        acft : str
-            ICAO Aircraft Designator (e.g., 'B737', 'B788').
-        CL : float or ureg.Quantity
-            Lift Coefficient [dimensionless].
-        Mach : float or ureg.Quantity
-            Mach Number [dimensionless].
-
-        Returns
-        -------
-        dict
-            'CD_total' : ureg.Quantity
-                Total Drag Coefficient [dimensionless]
-            'CD_parasite' : ureg.Quantity
-                Base parasite drag component [dimensionless]
-            'CD_induced' : ureg.Quantity
-                Induced drag component [dimensionless]
-            'CD_wave' : ureg.Quantity
-                Compressibility/Mach drag component [dimensionless]
-            'alpha_rad' : ureg.Quantity
-                Angle of Attack [radians]
-
-        Raises
-        ------
-        ValueError
-            If the ICAO Aircraft Designator is not found in the model data.
-        """
-        
-        # Ensure inputs are magnitude only if they were passed as Quantities
-        if hasattr(CL, "magnitude"): CL = CL.magnitude
-        if hasattr(Mach, "magnitude"): Mach = Mach.magnitude
-
-        if acft not in jsbsim_drag_polar._aircraft_data:
+        if acft not in jsbsim_drag_polars._aircraft_data:
[…]
