+++ src/jetfuelburn/utility/physics.py
@@ -4,11 +4,7 @@ from jetfuelburn import ureg
 
 @ureg.check('[length]')
 def _calculate_atmospheric_temperature(altitude):
-@ureg.check('[length]')
-def _calculate_atmospheric_temperature(altitude):
     r"""
-    Computes the air temperature as a function of altitude up to 20,000 meters 
-    based on the International Standard Atmosphere (ISA):
     Computes the air temperature as a function of altitude up to 20,000 meters 
     based on the International Standard Atmosphere (ISA):
 
@@ -16,11 +12,7 @@ def _calculate_atmospheric_temperature(altitude):
     
     Temperature in the Troposphere (<= 11,000 m) is calculated according to
     
-    
-    Temperature in the Troposphere (<= 11,000 m) is calculated according to
-    
     $$
-    T = T_0 - L \cdot h
     T = T_0 - L \cdot h
     $$
 
@@ -97,36 +89,10 @@ def _calculate_atmospheric_density(altitude):
 
     $$
     \rho = \rho_1 e^{\frac{-g M (h - h_1)}{R T_s}}
-    \rho = \rho_0 \left( \frac{T}{T_0} \right)^{\frac{-g}{L R} - 1}
-    $$
-
-    Temperature in the lower Stratosphere (> 11,000 m) is calculated according to
-
-    $$
-    \rho = \rho_1 e^{\frac{-g M (h - h_1)}{R T_s}}
     $$
 
     where:
 
-    | Symbol  | Dimension                   | Description                                | Value                |
-    |---------|-----------------------------|--------------------------------------------|----------------------|
-    | $\rho$  | [mass/volume]               | air density at altitude $h$                | N/A                  |
-    | $\rho_0$| [mass/volume]               | air density at sea level                   | 1.225 kg/m³          |
-    | $\rho_1$| [mass/volume]               | air density at tropopause (11,000 m)       | 0.36391 kg/m³        |
-    | $T$     | [temperature]               | temperature at altitude $h$                | N/A                  |
-    | $T_0$   | [temperature]               | temperature at sea level                   | 288.15 K             |
-    | $T_s$   | [temperature]               | constant temperature in lower Stratosphere | 216.65 K             |
-    | $h$     | [length]                    | altitude above sea level                   | variable             |
-    | $h_1$   | [length]                    | altitude at tropopause                     | 11,000 m             |
-    | $g$     | [length/time²]              | acceleration due to gravity                | 9.80665 m/s²         |
-    | $L$     | [temperature/length]        | temperature lapse rate (Troposphere)       | 0.0065 K/m           |
-    | $R$     | [energy/(mass*temperature)] | specific gas constant for air              | 287.052874 J/(kg·K)  |
-    | $M$     | [mass/mole]                 | molar mass of dry air                      | 0.0289644 kg/mol     |
-
-    References
-    ----------
-    - Troposphere: Eqn. (4.14) in Young, T. M. (2018). Performance of the Jet Transport Airplane. _John Wiley & Sons_. doi:[10.1002/9781118534786](https://doi.org/10.1002/9781118534786)
-    - Stratosphere: Eqn. (4.16) in Young, T. M. (2018). Performance of the Jet Transport Airplane. _John Wiley & Sons_. doi:[10.1002/9781118534786](https://doi.org/10.1002/9781118534786)
     | Symbol  | Dimension                   | Description                                | Value                |
[…]
