### 2026-02-02 a053f4f
updated lift classes
