### 2026-02-02 ad1196b
updated lift classes
