+++ src/jetfuelburn/rangeequation.py
@@ -11,10 +11,6 @@ import math
     '[speed]',
     '[time]/[length]' # [mg/Ns] = s/m
 )
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
->>>>>>> 36384dd (updated functions)
 def calculate_fuel_consumption_breguet_improved(
     R: float,
     LD: float,
@@ -43,11 +39,6 @@ def calculate_fuel_consumption_breguet_improved(
     '[speed]',
     '[time]/[length]' # [mg/Ns] = s/m
 )
-<<<<<<< HEAD
-=======
->>>>>>> ccabb7f (updated module and function names)
-=======
->>>>>>> 36384dd (updated functions)
 def calculate_fuel_consumption_breguet(
     R: float,
     LD: float,
