+++ src/jetfuelburn/utility/aerodynamics.py
@@ -1,6 +1,7 @@
 # %%
 import math
 import json
+import csv
 from importlib import resources
 from jetfuelburn.utility.math import _interpolate
 from jetfuelburn.utility.physics import (
@@ -42,9 +43,6 @@ class jsbsim_drag_polars:
         r"""
         asdf
         """        
-        if hasattr(L, "magnitude"): L = L.magnitude
-        if hasattr(M, "magnitude"): M = M.magnitude
-
         if acft not in jsbsim_drag_polars._aircraft_data:
             raise ValueError(f"ICAO Aircraft Designator '{acft}' not found in model data.")
         
@@ -55,27 +53,27 @@ class jsbsim_drag_polars:
             speed=_calculate_aircraft_velocity(M, H),
             altitude=H
         )
-        C_L = L * q * S
+        C_L = L / (q * S)
         
         lift_by_alpha: dict = data["lift_table"]
         alpha_rad = _interpolate(
-            x_val=C_L,
+            x_val=C_L.magnitude,
             x_list=lift_by_alpha["cl"],
             y_list=lift_by_alpha["alpha"]
         )
 
-        c_D0_by_alpha: dict = data["c_D0_table"]
+        c_D0_by_alpha: dict = data["cd0_table"]
         c_D_parasitic = _interpolate(
             x_val=alpha_rad,
             x_list=c_D0_by_alpha["alpha"],
-            y_list=c_D0_by_alpha["c_D0"]
+            y_list=c_D0_by_alpha["cd0"]
         )
 
         wave_drag_by_mach: dict = data["wave_drag_table"]
         c_D_wave = _interpolate(
             x_val=M,
             x_list=wave_drag_by_mach["mach"],
-            y_list=wave_drag_by_mach["c_D_wave"]
+            y_list=wave_drag_by_mach["cd_wave"]
         )
 
         c_D_induced = data["k_factor"] * (C_L ** 2)
@@ -83,4 +81,169 @@ class jsbsim_drag_polars:
         
         D = q * S * c_D_total
         
+        return D
+    
+
+
+class openap_drag_polars:
+    r"""
+    The class implements a low-speed drag polar model based on data published with the OpenAP model.
+
+    Warning
+    -------
+    The data used in this model is based on a XXX stoachastic model based on ADS-B data.....
+
+    See Also
+    --------
+    [`jetfuelburn.utility.aerodynamics.jsbsim_drag_polars`][]
+
+    References
+    ----------
+    - Sun, J., Hoekstra, J. M., & Ellerbroek, J. (2020). 
+      OpenAP: An open-source aircraft performance model for air transportation studies and simulations. 
+      _Aerospace_, 7(8), 104.
+      doi:[10.3390/aerospace7080104](https://doi.org/10.3390/aerospace7080104)
+    - Sun, J., Hoekstra, J. M., & Ellerbroek, J. (2020).
+      Estimating aircraft drag polar using open flight surveillance data and a stochastic total energy model.
+      _Transportation Research Part C: Emerging Technologies_, 114, 391-404. 
+      doi:[10.1016/j.trc.2020.01.026](https://doi.org/10.1016/j.trc.2020.01.026)
+    """
+
+    # Parsed from your CSV
+    _aircraft_data = {}
+    with resources.open_text("jetfuelburn.data.OpenAP", "data.csv") as file:
+        reader = csv.DictReader(file)
+        for row in reader:
+            icao = row["icao_code"]
+            _aircraft_data[icao] = {
+                "wing_area_m2": float(row["wing_area_m2"]),
+                "CD0": float(row["CD0"]),
+                "K": float(row["K"]),
+            }
+    
+
+    @staticmethod
+    def available_aircraft() -> list[str]:
[…]
