+++ src/jetfuelburn/utility/physics.py
@@ -1,3 +1,4 @@
+# %%
 import math
 from jetfuelburn import ureg
 
