### 2026-02-03 25644c9
udpated range equation functions
