+++ src/jetfuelburn/rangeequation.py
@@ -1,7 +1,6 @@
 from jetfuelburn import ureg
+from jetfuelburn import g
 import math
-
-
 
 
 @ureg.check(
@@ -9,27 +8,104 @@ import math
     '[]',
     '[mass]',
     '[speed]',
-    '[time]/[length]' # [mg/Ns] = s/m
+    '[speed]',
+    '[time]/[length]' # [mg/Ns] = s/m,
+    '[]',
+    '[]',
 )
 def calculate_fuel_consumption_breguet_improved(
     R: float,
     LD: float,
     m_after_cruise: float,
-    v_cruise: float,
-    TSFC_cruise: float,
+    V: float,
+    V_headwind: float,
+    TSFC: float,
+    lost_fuel_fraction: float = 0.0152,
+    recovered_fuel_fraction: float = 0.001,
 ) -> float:
     r"""
 
+    $$
+    W_{\text{fuel}} = W_{\text{LDG}} \left( \frac{1}{\exp\left\{ \frac{-R}{H}\left(1 - \frac{V_{\text{HW}}}{V}\right)} \right\} - \frac{\Delta W_{\text{lost}}}{W_{\text{TO}}} + \frac{\Delta W_{\text{rec}}}{W_{\text{TO}}}} - 1 \right)
+    $$
+
+    with
+
+    $$
+    H = \frac{L/D \cdot V}{TSFC \cdot g}
+    $$
+
+    where:
+
+    | Symbol                      | Dimension         | Description                                                    |
+    |-----------------------------|-------------------|----------------------------------------------------------------|
+    | $m_{f}$                     | [mass]            | Fuel required for the mission of the aircraft                  |
+    | $m_{LDG}$                   | [mass]            | Mass of the aircraft after cruise segment                      |
+    | $R$                         | [distance]        | Range of the aircraft (=mission distance)                      |
+    | $TSFC$                      | [time/distance]   | Thrust Specific Fuel Consumption of the aircraft during cruise |
+    | $g$                         | [acceleration]    | Acceleration due to gravity                                    |
+    | $L/D$                       | [dimensionless]   | Lift-to-Drag ratio of the aircraft                             |
+    | $V$                         | [speed]           | Average speed of the aircraft (TAS)                            |
+    | $V_{HW}$                    | [speed]           | Average speed of headwind component                            |
+    | $\frac{\Delta m_L}{m_{TO}}$ | [dimensionless]   | Lost fuel fraction                                             |
+    | $\frac{\Delta m_R}{m_{TO}}$ | [dimensionless]   | Recovered fuel fraction                                        | 
+
+    Warnings
+    --------
+    Unlike the equation in Randle et al. (2011), this function uses mass [kg] 
+    instead of weight [N] for the fuel mass.
+
+    Notes
+    -----
+    Assumes a lost fuel fraction $\Delta W_{lost}=0.0152$ 
+    and a recovered fuel fraction $\Delta W_{rec}=0.001$ as per 
+    section D of Randle et al. (2011). 
+    Values can be adjusted through function parameters.
+
     References
     --------
-    Randle, W. E., Hall, C. A., & Vera-Morales, M. (2011). Improved range equation based on aircraft flight data. _Journal of Aircraft_. doi:[10.2514/1.C031262](https://doi.org/10.2514/1.C031262)
+    Eqn. 19 in Randle, W. E., Hall, C. A., & Vera-Morales, M. (2011). 
[…]
