### 2026-02-03 d17920e
udpated range equation functions
