### 2026-02-03 6ceb499
updated range equation docs
