### 2026-02-03 7546b7c
updated range equation docs
