+++ tests/test_rangeequation.py
@@ -1,0 +1,183 @@
+import pytest
+from jetfuelburn import ureg, g
+import pint
+from jetfuelburn.rangeequation import (
+    calculate_fuel_consumption_breguet,
+    calculate_fuel_consumption_breguet_improved
+)
+from jetfuelburn.utility.tests import approx_with_units
+
+from .fixtures.rangeequation import breguet_range_fuel_calculation_data_1
+
+class TestCalculateFuelConsumptionBreguet:
+    
+    def test_valid_input_units(self):
+        """
+        Test that the function accepts valid units and returns a mass in kg.
+        """
+        R = 3000 * ureg.kilometer
+        LD = 18
+        m_after = 60 * ureg.metric_ton
+        V = 850 * ureg.kph
+        TSFC = 0.6 * (ureg.lb / ureg.lbf / ureg.hour)
+
+        result = calculate_fuel_consumption_breguet(R, LD, m_after, V, TSFC)
+        
+        assert result.check('[mass]')
+        assert result.units == ureg.kg
+        assert result.magnitude > 0
+
+    def test_zero_range(self):
+        """
+        Test that zero range results in zero fuel consumption.
+        """
+        R = 0 * ureg.meter
+        LD = 18
+        m_after = 60000 * ureg.kg
+        V = 200 * ureg.meter / ureg.second
+        TSFC = 15 * (ureg.mg / ureg.N / ureg.second)
+
+        result = calculate_fuel_consumption_breguet(R, LD, m_after, V, TSFC)
+        
+        assert result.magnitude == 0
+        assert result.units == ureg.kg
+
+    @pytest.mark.parametrize("invalid_input", [
+        {"R": -100 * ureg.km, "msg": "Range must be greater than zero"},
+        {"LD": 0.5, "msg": "Lift-to-Drag ratio must be greater than 1"},
+        {"m_after_cruise": -50 * ureg.kg, "msg": "Mass after cruise must be greater than zero"},
+        {"V": 0 * ureg.kph, "msg": "Cruise speed must be greater than zero"},
+        {"TSFC": -5 * (ureg.mg / ureg.N / ureg.s), "msg": "Thrust Specific Fuel Consumption must be greater than zero"},
+    ])
+    def test_raises_value_error_on_invalid_magnitudes(self, invalid_input):
+        """
+        Test that specific value errors are raised for physically impossible inputs.
+        """
+        # Default valid values
+        params = {
+            "R": 2000 * ureg.nmi,
+            "LD": 18,
+            "m_after_cruise": 100 * ureg.metric_ton,
+            "V": 800 * ureg.kph,
+            "TSFC": 17 * (ureg.mg / ureg.N / ureg.s)
+        }
+        
+        expected_msg = invalid_input.pop("msg")
+        params.update(invalid_input)
+
+        with pytest.raises(ValueError, match=expected_msg):
+            calculate_fuel_consumption_breguet(**params)
+
+    def test_raises_dimensionality_error(self):
+        """
+        Test that passing incorrect units (e.g., time instead of length) raises an error
+        via the @ureg.check decorator.
+        """
+        with pytest.raises((ValueError, pint.errors.DimensionalityError)):
+            calculate_fuel_consumption_breguet(
+                R=500 * ureg.second, # Invalid: Time instead of Length
+                LD=18,
+                m_after_cruise=50000 * ureg.kg,
+                V=200 * ureg.mps,
+                TSFC=0.0001 * (ureg.sec / ureg.meter)
+            )
+
+    def test_calculate_fuel_consumption_breguet_from_fixture(self, breguet_range_fuel_calculation_data_1):
+        """
+        Test using a data fixture.
+        """
+        input_data, expected_data = breguet_range_fuel_calculation_data_1
[…]
