### 2026-02-03 1d1ba00
updated range equation function
