### 2026-02-03 b47fe46
updated range equation function
