+++ src/jetfuelburn/rangeequation.py
@@ -2,6 +2,105 @@
 from jetfuelburn import ureg
 from jetfuelburn import g
 import math
+from jetfuelburn.utility.physics import _calculate_atmospheric_density
+
+
+@ureg.check(
+    '[length]',
+    '[length]',
+    '[]',
+    '[]',
+    '[mass]',
+    '[area]',
+    '[speed]',
+    '[time]/[length]', # [mg/Ns] = s/m
+)
+def calculate_fuel_consumption_arctan(
+    R: float | int,
+    h: float | int,
+    K: float | int,
+    C_D0: float | int,
+    m_after_cruise: float | int,
+    S: float | int,
+    V: float | int,
+    TSFC: float | int,
+) -> float:
+    r"""
+
+    $$
+        m_f = \frac{(B + m_{LDG}^2) \tan(\theta)}{\sqrt{B} - m_{LDG} \tan(\theta)}
+    $$
+
+    with
+
+    \begin{align*}
+    \theta &= \frac{R g TSFC}{2 E_{max} V} \\
+    B &= \left( \frac{C_{D_0}}{K} \right) \left( \frac{\rho V^2 S}{2g} \right)^2
+    \end{align*}
+
+    where:
+
+    | Symbol     | Dimension         | Description                                                            |
+    |------------|-------------------|------------------------------------------------------------------------|
+    | $m_{fuel}$ | [mass]            | Fuel required for the mission of the aircraft                          |
+    | $R$        | [distance]        | Range of the aircraft (=mission distance)                              |
+    | $TSFC$     | [time/distance]   | Thrust Specific Fuel Consumption of the aircraft during cruise         |
+    | $g$        | [acceleration]    | Acceleration due to gravity                                            |
+    | $V$        | [speed]           | Average cruise speed of the aircraft (TAS)                             |
+    | $E_{max}$  | [energy/mass]     | Maximum lift-to-drag ratio according to drag parabola                  |
+    | $C_{D_0}$  | [dimensionless]   | Zero-lift drag coefficient of the aircraft                             |
+    | $K$        | [dimensionless]   | Lift-dependent drag factor of the aircraft                             |
+    | $\rho$     | [mass/volume]     | Air density at cruise altitude                                         |
+    | $S$        | [area]            | Wing reference area of the aircraft                                    |
+
+    References
+    ----------
+    Eqn. 13.25a ff. in 
+    Young, T. M. (2018). 
+    Performance of the Jet Transport Airplane. 
+    _John Wiley & Sons_. 
+    doi:[10.1002/9781118534786](https://doi.org/10.1002/9781118534786)
+
+    Example
+    -------
+    ```pyodide install='jetfuelburn'
+    import jetfuelburn
+    from jetfuelburn import ureg
+    from jetfuelburn.rangeequation import calculate_fuel_consumption_arctan
+    calculate_fuel_consumption_arctan(
+        R=2000*ureg.nmi,
+        h=35000*ureg.feet,
+        K=0.045,
+        C_D0=0.02,
+        m_after_cruise=100*ureg.metric_ton,
+        S=122.6*ureg.meter**2,
+        V=800*ureg.kph,
+        TSFC=17*(ureg.mg/ureg.N/ureg.s),
+    )
+    ```
+
+    """
+    if R==0 * ureg.meter:
+        return 0 * ureg.kg
+    if R < 0 * ureg.meter:
+        raise ValueError("Range must be greater than zero.")
+    if h < 0 * ureg.meter:
+        raise ValueError("Altitude must be greater than zero.")
+    if m_after_cruise < 0 * ureg.kg:
[…]
