### 2026-02-03 1e32e8e
added helper function
