+++ src/jetfuelburn/utility/engines.py
@@ -1,0 +1,127 @@
+import pint
+from jetfuelburn import ureg
+from jetfuelburn.utility.physics import _calculate_atmospheric_temperature
+
+@ureg.check(
+    '[time]/[length]', # [mg/Ns] = s/m
+    '[dimensionless]',
+    '[dimensionless]',
+    '[length]',
+    '[length]',
+    None
+)
+def calculate_corrected_tsfc(
+    tsfc_reported: pint.Quantity[float | int],
+    M_reported: pint.Quantity[float | int],
+    M_actual: pint.Quantity[float | int],
+    h_reported: pint.Quantity[float | int],
+    h_actual: pint.Quantity[float | int],
+    beta: float | int | pint.Quantity[float | int],
+) -> pint.Quantity[float | int]:
+    r"""
+    Given a reported thrust-specific fuel consumption (TSFC) at reference conditions,
+    calculates the actual TSFC under different flight conditions according to an empirical correction formula 
+    proposed by Martinez-Val & Perez (1992) and used by Cavcar (2006).
+
+    TSFC is calculated as:
+
+    $$
+    \text{TSFC}_{\text{actual}} = \text{TSFC}_{\text{reported}} \cdot \left( \frac{M_{\text{actual}}}{M_{\text{reported}}} \right)^{\beta} \cdot \sqrt{\frac{\Theta_{\text{actual}}}{\Theta_{\text{reported}}}}
+    $$
+
+    where:
+
+    | Symbol                         | Dimension        | Description                                                                                                               |
+    |--------------------------------|------------------|---------------------------------------------------------------------------------------------------------------------------|
+    | $\text{TSFC}_{\text{actual}}$  | [time]/[length]  | The actual thrust-specific fuel consumption under the given conditions ($M_{\text{actual}}$, $h_{\text{actual}}$).        |
+    | $\text{TSFC}_{\text{reported}}$| [time]/[length]  | The reported thrust-specific fuel consumption at the reference conditions ($M_{\text{reported}}$, $h_{\text{reported}}$). |
+    | $M_{\text{actual}}$            | [dimensionless]  | The actual Mach number during the flight.                                                                                 |
+    | $M_{\text{reported}}$          | [dimensionless]  | The Mach number at which the TSFC was reported.                                                                           |
+    | $h_{\text{actual}}$            | [length]         | The actual altitude during the flight.                                                                                    |
+    | $h_{\text{reported}}$          | [length]         | The altitude at which the TSFC was reported.                                                                              |
+    | $\Theta_{\text{actual}}$       | [dimensionless]  | The actual atmospheric temperature ratio, calculated as $T_{\text{actual}} / T(h=0).                                      |
+    | $\Theta_{\text{reported}}$     | [dimensionless]  | The reported atmospheric temperature ratio, calculated as $T_{\text{reported}} / T(h=0)$.                                 |
+    | $\beta$                        | [dimensionless]  | An empirically derived exponent that captures how TSFC changes with Mach number.                                          |
+
+    Notes
+    -----
+    Cavcar (2006) suggests for $\beta$:
+    
+    > 0.2–0.4 for low bypass turbofan engines and 0.4–0.7 for high bypass turbofan engines.  
+    - P.126 in Martinez-Val & Perez (1992)
+
+    See Also
+    --------
+    [Thrust-specific fuel consumption entry on Wikipedia](https://en.wikipedia.org/wiki/Thrust-specific_fuel_consumption)
+
[…]
