### 2026-02-05 1b831e0
added draft of numeric range calculation class
