### 2026-02-05 fabb609
added draft of numeric range calculation class
