+++ src/jetfuelburn/rangeequation.py
@@ -1,8 +1,130 @@
 # %%
+import pint
 from jetfuelburn import ureg
-from jetfuelburn import g
 import math
+from typing import Callable
+
 from jetfuelburn.utility.physics import _calculate_atmospheric_density
+from jetfuelburn.utility.code import _validate_function
+
+
+
+class IntegratedRangeCalculator:
+    """
+    Performs numerical integration of the Specific Air Range (SAR) to estimate 
+    cruise performance. Supports both static and dynamic TSFC models.
+    """
+    def __init__(
+        self, 
+        func_or_scalar_tsfc: pint.Quantity[int | float] | Callable,
+        func_or_scalar_lift_to_drag: int | float | pint.Quantity[int | float] | Callable, 
+    ):
+        r"""
+        """
+        if isinstance(func_or_scalar_tsfc, pint.Quantity):
+            self.func_tsfc = lambda M, H, F: func_or_scalar_tsfc
+        elif callable(func_or_scalar_tsfc):
+            # If dynamic, validate it and use it directly
+            self._validate_function(
+                func=func_or_scalar_tsfc,
+                required_params={'M', 'H', 'F'},
+                expected_return_type=pint.Quantity
+            )
+            self.func_tsfc = func_or_scalar_tsfc
+        else:
+            raise TypeError("tsfc must be a pint.Quantity or a callable.")
+
+        if isinstance(func_or_scalar_lift_to_drag, (int, float, pint.Quantity)):
+            self.func_drag = lambda L, M, h: L / func_or_scalar_lift_to_drag
+        elif callable(func_or_scalar_lift_to_drag):
+            self._validate_function(
+                func=func_or_scalar_lift_to_drag,
+                required_params={'L', 'M', 'h'},
+                expected_return_type=pint.Quantity
+            )
+            self.func_drag = func_or_scalar_lift_to_drag
+        else:
+            raise TypeError("lift_to_drag must be a scalar or a callable.")
+        
+    
+    @ureg.check(
+        None,
+        '[mass]',
+        '[mass]',
+        '[length]',
+        '[speed]',
+        None
+    )
+    def compute_integrated_range(
+        self,
+        m_start: pint.Quantity[int | float],
+        m_end: pint.Quantity[int | float],
+        h: pint.Quantity[int | float],
+        V: pint.Quantity[int | float],
+        num_intervals: int = 20,
+    ):
+        r"""
+        asdf
+
+        """
+        m_start = m_start.to("kg")
+        m_end = m_end.to("kg")
+        m_increment = (m_end - m_start) / (num_intervals - 1)
+        m_stations = [(m_start + i * m_increment) for i in range(num_intervals)]
+
+        # 2. Performance Parameters
+        ranges = [Q_(0, "nautical_mile")]
+        times = [Q_(0, "hour")]
+        list_SAR_at_station = []
+        list_fuel_burn_per_distance = [] # New list for Fuel Burn per Distance
+        #TODO: convert speed to mach number based on altitude
+
+        # 3. Loop
+        for m_station in m_stations:
+            L = m_station * ureg.gravity
+            D = self.func_drag(
+                L=L,
+                M=,
+                h=h
+            )
+            TSFC = self.func_tsfc(
+                M=M,
+                h=h,
+                F=D
+            )
+            fuel_flow = D * TSFC
+            
+            
[…]
