+++ src/jetfuelburn/utility/code.py
@@ -1,49 +1,28 @@
 # %%
-from typing import Callable
+from jetfuelburn import ureg
 import inspect
+from typing import Callable
+import pint
 
-def _validate_function(
+
+def _validate_physics_function_parameters(
     func: Callable,
-    required_params: set,
-    expected_return_type: type | None = None
+    required_params: set[str],
 ) -> None:
     r"""
-    Given a callable (= a function), a set of required parameter names
-    and an optional expected return type,
-    checks if the function accepts all required parameters
-    and (if provided) has the correct return type hint.
-
-    For instance, for a function:
-
-    ```python
-    def my_func(a, b, c) -> int:
-        return a + b + c
-    ```
-
-    the function can be validated as:
-
-    ```python
-    _validate_function(
-    my_func, {'a', 'b', 'c'}, expected_return_type=int)
-    ```
+    Validates that a physics function has exactly the expected parameters.
 
     Parameters
     ----------
-    func : Callable
-        The function to validate.
-    required_params : set
-        A set of parameter names that the function must accept.
-    expected_return_type : type | None, optional
-        If provided, the function return type hint must match this type.  Defaults to None.
+    func: Callable
+        The physics function to validate.
+    required_params: set[str]
+        A set of required parameter names.
 
     Raises
     ------
     ValueError
-        If the function is missing required parameters or has an invalid return type.
-
-    Returns
-    -------
-    None
+        If the function's parameters do not match the expected names.
     """
     sig = inspect.signature(func)
     func_params = sig.parameters.keys()
@@ -53,16 +32,48 @@ def _validate_function(
             f"Function {func.__name__} missing parameters: {missing}. "
             f"Must accept: {required_params}"
         )
-    
-    if expected_return_type is not None:
-        return_annotation = sig.return_annotation
-        
-        if return_annotation is inspect.Signature.empty:
-            raise ValueError(f"Function '{func.__name__}' is missing a return type hint.")
 
-        if return_annotation != expected_return_type:
-            raise ValueError(
-                f"Function '{func.__name__}' has invalid return type. "
-                f"Expected: {expected_return_type.__name__}, "
-                f"Got: {getattr(return_annotation, '__name__', return_annotation)}"
-            )
+
+def _normalize_physics_function_or_scalar(function_or_scalar: int | float | pint.Quantity | Callable) -> Callable:
+        r"""
+        Given a scalar or Callable, wraps scalars in a lambda function. 
+        Callables are returned as-is. 
+
+        See Also
+        --------
+        [`jetfuelburn.rangeequation.IntegratedRangeCalculator`][]
+
+        Notes
+        -----
+        For scalars, the resulting lambda function will accept any arguments and ignore them, returning the constant value.  
+        For instance, for:
+
+        ```python
+        const_func = _normalize_physics_function(42)
+        ```
+
+        Calling `const_func(x, y, z)` for any `x`, `y`, `z` will return `42`.
+
+        Parameters
+        ----------
+        function_or_scalar: int | float | pint.Quantity | Callable
+            A scalar value or a function that computes a physics parameter.
+
+        Returns
+        -------
+        Callable
+            A function that returns the scalar value or the original Callable.
+
+        Raises
+        ------
+        TypeError
+            If the input is neither a scalar nor a Callable.
+        """
+        if isinstance(function_or_scalar, (pint.Quantity, float, int)):
+            return lambda *args, **kwargs: function_or_scalar
+        
+        elif callable(function_or_scalar):
+            return function_or_scalar
+            
+        else:
+            raise TypeError(f"Input must be Quantity, scalar, or Callable. Got {type(function_or_scalar)}")
