+++ src/jetfuelburn/rangeequation.py
@@ -4,127 +4,91 @@ from jetfuelburn import ureg
 import math
 from typing import Callable
 
-from jetfuelburn.utility.physics import _calculate_atmospheric_density
-from jetfuelburn.utility.code import _validate_function
+from jetfuelburn.utility.physics import (
+    _calculate_atmospheric_density,
+    _calculate_aircraft_velocity,
+)
+from jetfuelburn.utility.code import (
+    _validate_physics_function_parameters,
+    _normalize_physics_function_or_scalar,
+)
 
 
-
-class IntegratedRangeCalculator:
-    """
-    Performs numerical integration of the Specific Air Range (SAR) to estimate 
-    cruise performance. Supports both static and dynamic TSFC models.
-    """
-    def __init__(
-        self, 
-        func_or_scalar_tsfc: pint.Quantity[int | float] | Callable,
-        func_or_scalar_lift_to_drag: int | float | pint.Quantity[int | float] | Callable, 
-    ):
-        r"""
-        """
-        if isinstance(func_or_scalar_tsfc, pint.Quantity):
-            self.func_tsfc = lambda M, H, F: func_or_scalar_tsfc
-        elif callable(func_or_scalar_tsfc):
-            # If dynamic, validate it and use it directly
-            self._validate_function(
-                func=func_or_scalar_tsfc,
-                required_params={'M', 'H', 'F'},
-                expected_return_type=pint.Quantity
-            )
-            self.func_tsfc = func_or_scalar_tsfc
-        else:
-            raise TypeError("tsfc must be a pint.Quantity or a callable.")
-
-        if isinstance(func_or_scalar_lift_to_drag, (int, float, pint.Quantity)):
-            self.func_drag = lambda L, M, h: L / func_or_scalar_lift_to_drag
-        elif callable(func_or_scalar_lift_to_drag):
-            self._validate_function(
-                func=func_or_scalar_lift_to_drag,
-                required_params={'L', 'M', 'h'},
-                expected_return_type=pint.Quantity
-            )
-            self.func_drag = func_or_scalar_lift_to_drag
-        else:
-            raise TypeError("lift_to_drag must be a scalar or a callable.")
-        
-    
-    @ureg.check(
-        None,
-        '[mass]',
-        '[mass]',
-        '[length]',
-        '[speed]',
-        None
+@ureg.check(
+    '[mass]',
+    '[length]',
+    '[length]',
+    '[]',
+    None,
+    None,
+    '[mass]',
+)
+def calculate_integrated_range(
+    m_end: pint.Quantity,
+    R: pint.Quantity,
+    h: pint.Quantity,
+    M: float,
+    TSFC: pint.Quantity | Callable,
+    LD: float | pint.Quantity | Callable,
+    integration_mass_step: pint.Quantity = 100 * ureg.kg
+) -> pint.Quantity:
+    r"""
+    Example
+    -------
+    ```pyodide install='jetfuelburn'
+    import jetfuelburn
+    from jetfuelburn import ureg
+    from jetfuelburn.rangeequation import calculate_integrated_range
+    calculate_integrated_range(
+        m_end=100*ureg.metric_ton,
+        R=2000*ureg.nmi,
+        h=35000*ureg.feet,
+        M=0.78,
+        TSFC=17*(ureg.mg/ureg.N/ureg.s),
+        LD=18,
     )
-    def compute_integrated_range(
-        self,
-        m_start: pint.Quantity[int | float],
-        m_end: pint.Quantity[int | float],
-        h: pint.Quantity[int | float],
-        V: pint.Quantity[int | float],
-        num_intervals: int = 20,
-    ):
-        r"""
-        asdf
+    """
[…]
