### 2026-02-06 b1ccfe2
renamed physics function
