### 2026-02-07 d3598cd
updated tsfc function and tests
