+++ src/jetfuelburn/utility/physics.py
@@ -1,4 +1,3 @@
-# %%
 import math
 from jetfuelburn import ureg
 
