+++ README.md
@@ -1,5 +1,9 @@
 # JetFuelBurn
 
+[![PyPI Downloads](https://img.shields.io/pypi/dm/jetfuelburn?label=PyPI%20Downloads&logo=pypi&logoColor=white)](https://pypi.org/project/jetfuelburn/)
+[![License: MIT](https://img.shields.io/pypi/l/jetfuelburn?label=License&logo=open-source-initiative&logoColor=white)](https://pypi.org/project/jetfuelburn/)
+[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/jetfuelburn?logo=python&logoColor=white)](https://pypi.org/project/jetfuelburn/)
+
 A Python package for calculating fuel burn of commercial aircraft.  
 Maintainance Team: [@michaelweinold](https://github.com/michaelweinold)
 
