### 2026-02-07 292f2e3
updated tsfc function and tests
