### 2026-02-08 56ec965
updated aerodyamics functions
