+++ src/jetfuelburn/utility/aerodynamics.py
@@ -2,6 +2,7 @@
 import json
 import csv
 from importlib import resources
+from typing import Callable
 import functools
 from jetfuelburn.utility.math import _interpolate
 from jetfuelburn.utility.physics import (
@@ -14,7 +15,8 @@ from jetfuelburn import data, ureg
 
 class jsbsim_drag_polars:
     r"""
-    The class implements a component-based drag polar model derived from open-source JSBSim flight dynamics data.
+    The class implements a component-based drag polar model derived from open-source 
+    [JSBSim](https://github.com/JSBSim-Team/jsbsim) flight dynamics data.
     """
 
     _aircraft_data = {}
@@ -37,12 +39,43 @@ class jsbsim_drag_polars:
     )
     def calculate_drag(
         acft: str,
-        L: pint.Quantity[float | int],
-        M: float | int | pint.Quantity[float | int],
-        h: pint.Quantity[float | int],
-    ) -> pint.Quantity[float | int]:
+        L: pint.Quantity,
+        M: float | int | pint.Quantity,
+        h: pint.Quantity,
+    ) -> pint.Quantity:
         r"""
-        asdf
+        Given lift, Mach number, and altitude, calculates the drag force for a specified aircraft
+        using a component-based drag polar model derived from JSBSim flight dynamics data.
+
+        Drag is calculated as:
+
+        \begin{align*}
+        D &= q \cdot S \cdot C_D \\
+        C_D &= C_{D0} + C_{D_{induced}} + C_{D_{wave}} \\
+        C_{D_{induced}} &= k \cdot C_L^2
+        \end{align*}
+
+        where:
+
+        | Symbol           | Dimension             | Description                        |
+        |------------------|-----------------------|------------------------------------|
+        | $D$              | [force]               | Drag force                         |
+        | $q$              | [pressure]            | Dynamic pressure                   |
+        | $S$              | [area]                | Wing reference area                |
+        | $C_D$            | [dimensionless]       | Total drag coefficient             |
+        | $C_{D0}$         | [dimensionless]       | Parasitic drag coefficient at zero lift |
+        | $C_{D_{induced}}$| [dimensionless]       | Induced drag coefficient           |
+        | $C_{D_{wave}}$   | [dimensionless]       | Wave drag coefficient               |
+        | $k$              | [dimensionless]       | Induced drag factor                |
+        | $C_L$            | [dimensionless]       | Lift coefficient                   |
+        | $L$              | [force]               | Lift force                         |
+
+        See Also
+        --------
+        [JSBSim Flight Dynamics Data]
+
+        References
+        ----------
         """        
         if acft not in jsbsim_drag_polars._aircraft_data:
             raise ValueError(f"ICAO Aircraft Designator '{acft}' not found in model data.")
@@ -91,14 +124,56 @@ class jsbsim_drag_polars:
         '[]', # Mach number
         '[length]', # altitude
     )
-    def _calculate_lift_to_drag_physics(
+    def calculate_lift_to_drag(
         acft: str,
-        L: pint.Quantity[float | int],
-        M: float | int | pint.Quantity[float | int],
-        h: pint.Quantity[float | int],
-    ) -> pint.Quantity[float | int]:
+        L: pint.Quantity,
+        M: float | int | pint.Quantity,
+        h: pint.Quantity,
+    ) -> pint.Quantity:
         r"""
-        asdf
+        Given lift, Mach number, and altitude, 
+        calculates the lift-to-drag ratio for a specified aircraft 
+        using the JSBSim-based drag polar model.
+
+        See Also
+        --------
[…]
