+++ src/jetfuelburn/utility/aerodynamics.py
@@ -10,13 +10,33 @@ from jetfuelburn.utility.physics import (
     _calculate_airspeed_from_mach
 )
 import pint
-from jetfuelburn import data, ureg
+from jetfuelburn import ureg
 
 
 class jsbsim_drag_polars:
     r"""
     The class implements a component-based drag polar model derived from open-source 
     [JSBSim](https://github.com/JSBSim-Team/jsbsim) flight dynamics data.
+
+    ```python exec="true" html="true"
+    from jetfuelburn.figures.aerodynamics import figure_jsbsim_dragpolar_mach_effects
+    fig = figure_jsbsim_dragpolar_mach_effects()
+    print(fig.to_html(full_html=False, include_plotlyjs="cdn"))
+    # https://pawamoy.github.io/markdown-exec/gallery/#with-plotly
+    ```
+
+    Warning
+    -------
+    The data used in this model is based on data collected by 
+
+    See Also
+    --------
+    [`jetfuelburn.utility.aerodynamics.openap_drag_polars`][]
+
+    References
+    ----------
+    Berndt, J. S., & JSBSim Development Team. (2011). 
+    [JSBSim: An open source, platform-independent, flight dynamics model in C++]((https://jsbsim.sourceforge.net/JSBSimReferenceManual.pdf)). 
     """
 
     _aircraft_data = {}
@@ -33,9 +53,9 @@ class jsbsim_drag_polars:
     @staticmethod
     @ureg.check(
         None,
-        '[force]', # lift
-        '[]', # Mach number
-        '[length]', # altitude
+        '[force]',
+        '[]',
+        '[length]',
     )
     def calculate_drag(
         acft: str,
@@ -51,32 +71,76 @@ class jsbsim_drag_polars:
 
         \begin{align*}
         D &= q \cdot S \cdot C_D \\
-        C_D &= C_{D0} + C_{D_{induced}} + C_{D_{wave}} \\
-        C_{D_{induced}} &= k \cdot C_L^2
+        C_L &= \frac{L}{q \cdot S}
         \end{align*}
+
+        with the drag coefficient $C_D$ calculated as the sum of parasitic, induced, and wave drag components 
+        (drag build-up):
+
+        $$
+        C_D = C_{D0}(\alpha) + C_{D_{i}} + C_{D_{wave}} = C_{D0}(\alpha) + k \cdot C_L^2 + C_{D_{wave}}
+        $$
 
         where:
 
-        | Symbol           | Dimension             | Description                        |
-        |------------------|-----------------------|------------------------------------|
-        | $D$              | [force]               | Drag force                         |
-        | $q$              | [pressure]            | Dynamic pressure                   |
-        | $S$              | [area]                | Wing reference area                |
-        | $C_D$            | [dimensionless]       | Total drag coefficient             |
+        | Symbol           | Dimension             | Description                             |
+        |------------------|-----------------------|-----------------------------------------|
+        | $D$              | [force]               | Drag force                              |
+        | $L$              | [force]               | Lift force                              |
+        | $q$              | [pressure]            | Dynamic pressure                        |
+        | $S$              | [area]                | Wing reference area                     |
+        | $C_D$            | [dimensionless]       | Total drag coefficient                  |
         | $C_{D0}$         | [dimensionless]       | Parasitic drag coefficient at zero lift |
-        | $C_{D_{induced}}$| [dimensionless]       | Induced drag coefficient           |
-        | $C_{D_{wave}}$   | [dimensionless]       | Wave drag coefficient               |
-        | $k$              | [dimensionless]       | Induced drag factor                |
-        | $C_L$            | [dimensionless]       | Lift coefficient                   |
-        | $L$              | [force]               | Lift force                         |
[…]
