+++ src/jetfuelburn/figures/aerodynamics.py
@@ -1,7 +1,11 @@
 # %%
 import plotly.graph_objects as go
+from jetfuelburn import ureg
 
-from jetfuelburn.utility.aerodynamics import openap_drag_polars
+from jetfuelburn.utility.aerodynamics import (
+    openap_drag_polars,
+    jsbsim_drag_polars
+)
 
 def figure_openap_dragpolar():
     """
@@ -48,7 +52,6 @@ def figure_openap_dragpolar():
         ))
 
     fig.update_layout(
-        title='Drag Polar Comparison (OpenAP Model)',
         xaxis_title='Drag Coefficient (C<sub>D</sub>)',
         yaxis_title='Lift Coefficient (CL)',
         font=dict(family='Arial', size=11),
@@ -67,6 +70,146 @@ def figure_openap_dragpolar():
             gridcolor='lightgrey',
             zeroline=True,
             zerolinecolor='grey'
+        ),
+        yaxis=dict(
+            showgrid=True, 
+            gridcolor='lightgrey',
+            zeroline=True,
+            zerolinecolor='grey'
+        ),
+        plot_bgcolor='white',
+        width=600,
+        height=400
+    )
+
+    return fig
+
+
+def figure_jsbsim_dragpolar_mach_effects():
+    """
+    Creates a Plotly figure showing the effect of Mach number on the drag polar 
+    using the JSBSim component-based model.
+
+    The plot demonstrates how Drag Coefficient (CD) increases for the same 
+    Lift Coefficient (CL) as Mach number rises (wave drag).
+
+    Returns:
+        go.Figure: Plotly figure object with CL vs CD curves for various Mach numbers
+    """
+    acft = "B788"
+    
+    # Define conditions
+    # We use a fixed altitude to calculate a reference dynamic pressure (q)
+    altitude = 30000 * ureg.foot 
+    mach_values = [0.4, 0.6, 0.7, 0.76, 0.78, 0.80, 0.82]
+    
+    # Helper to estimate air density at altitude (Standard Atmosphere) 
+    # to calculate valid Lift forces for the model inputs.
+    # rho_0 = 1.225 kg/m^3
+    # at 30k ft (~9144m), sigma ~ 0.37
+    rho = 0.458 * ureg('kg/m^3') # Approx density at 30k ft
+    
+    # Retrieve Wing Area (S) to calculate Lift Force (L) inputs
+    # We access the private dictionary because we need S to sweep CL correctly
+    if acft not in jsbsim_drag_polars._aircraft_data:
+         raise ValueError(f"Aircraft {acft} not found")
+         
+    acft_data = jsbsim_drag_polars._aircraft_data[acft]
+    S = acft_data['wing_area_sqft'] * ureg.square_feet
+    
+    # Generate Target CL range
+    # 0.0 to 1.3 in steps of 0.05
+    target_cls = [x * 0.05 for x in range(27)]
+
+    fig = go.Figure()
+    
+    # Color scale for Mach numbers
+    colors = [
+        '#ffeebb', '#ffcc88', '#ffaa66', '#ff8844', 
+        '#ff6622', '#ff4400', '#cc0000'
+    ]
+
+    for i, mach in enumerate(mach_values):
+        
+        # 1. Calculate Dynamic Pressure (q) for this Mach
+        # Speed of Sound (a) at 30k ft is approx 303 m/s
+        a = 303 * ureg('m/s')
+        v = mach * a
+        q = 0.5 * rho * v**2
+        
+        cd_results = []
+        cl_results = []
+
+        for cl in target_cls:
+            if cl == 0:
+                # Avoid division by zero issues or zero lift logic if specific models dislike it,
+                # though typical polars start at CD0. 
[…]
