### 2026-02-09 82dfd55
updated range equation docstrings
