### 2026-02-09 c530a81
updated range equation function docstring
