### 2026-02-11 0f04b9b
updated range equation tests
