+++ tests/test_rangeequation.py
@@ -3,9 +3,13 @@ from jetfuelburn import ureg
 import pint
 from jetfuelburn.rangeequation import (
     calculate_fuel_consumption_breguet,
-    calculate_fuel_consumption_breguet_improved
+    calculate_fuel_consumption_breguet_improved,
+    calculate_fuel_consumption_stepclimb_arctan,
+    calculate_fuel_consumption_stepclimb_integration
 )
 from jetfuelburn.utility.tests import approx_with_units
+from jetfuelburn.utility.physics import _calculate_mach_from_airspeed
+from jetfuelburn.utility.aerodynamics import openap_drag_polars
 
 from .fixtures.rangeequation import breguet_range_fuel_calculation_data_1
 
@@ -181,3 +185,206 @@ class TestCalculateFuelConsumptionBreguetImproved:
                 V_headwind=50 * ureg.gram, # Invalid: Mass instead of Speed
                 TSFC=17 * (ureg.mg / ureg.N / ureg.s)
             )
+
+
+class TestCalculateFuelConsumptionStepclimbArctan:
+
+    def test_valid_input_units(self):
+        """Test valid calculation for the arctan method."""
+        result = calculate_fuel_consumption_stepclimb_arctan(
+            R=2000 * ureg.nmi,
+            h=35000 * ureg.feet,
+            K=0.045,
+            C_D0=0.02,
+            m_after_cruise=100 * ureg.metric_ton,
+            S=122.6 * ureg.meter**2,
+            V=800 * ureg.kph,
+            TSFC=17 * (ureg.mg / ureg.N / ureg.s),
+        )
+        assert result.check('[mass]')
+        assert result.units == ureg.kg
+        assert result.magnitude > 0
+
+    def test_zero_range(self):
+        """Test zero range returns zero mass."""
+        result = calculate_fuel_consumption_stepclimb_arctan(
+            R=0 * ureg.nmi,
+            h=35000 * ureg.feet,
+            K=0.045,
+            C_D0=0.02,
+            m_after_cruise=100 * ureg.metric_ton,
+            S=122.6 * ureg.meter**2,
+            V=800 * ureg.kph,
+            TSFC=17 * (ureg.mg / ureg.N / ureg.s),
+        )
+        assert result.magnitude == 0
+
+    @pytest.mark.parametrize("invalid_input", [
+        {"R": -100 * ureg.km, "msg": "Range must be greater than zero"},
+        {"h": -100 * ureg.ft, "msg": "Altitude must be greater than zero"},
+        {"m_after_cruise": -50 * ureg.kg, "msg": "Mass after cruise must be greater than zero"},
+        {"S": 1.0 * ureg.m**2, "msg": "Lift-to-Drag ratio must be greater than 1"}, # Error msg corresponds to check S<=1 in source
+        {"V": 0 * ureg.kph, "msg": "Cruise speed must be greater than zero"},
+        {"TSFC": -5 * (ureg.mg / ureg.N / ureg.s), "msg": "Thrust Specific Fuel Consumption must be greater than zero"},
+    ])
+    def test_raises_value_error_on_invalid_magnitudes(self, invalid_input):
+        params = {
+            "R": 2000 * ureg.nmi,
+            "h": 35000 * ureg.feet,
+            "K": 0.045,
+            "C_D0": 0.02,
+            "m_after_cruise": 100 * ureg.metric_ton,
+            "S": 122.6 * ureg.meter**2,
+            "V": 800 * ureg.kph,
+            "TSFC": 17 * (ureg.mg / ureg.N / ureg.s),
+        }
+        
+        expected_msg = invalid_input.pop("msg")
+        params.update(invalid_input)
+
[…]
