### 2026-02-11 f576892
implemented black code style
