+++ tests/fixtures/reducedorder.py
@@ -1,6 +1,7 @@
 import pytest
 from jetfuelburn import ureg
 
+
 @pytest.fixture(scope="module")
 def fixture_aim2015_B787():
     """
@@ -13,8 +14,8 @@ def fixture_aim2015_B787():
     Notes
     -----
     - We assume: "(...) 95 kg for a passenger with luggage and an average of 4500 kg hold freight." - P.97, Dray et al. (2019)
-    - We further assume a 200km climb/descent distance. 
-    
+    - We further assume a 200km climb/descent distance.
+
     See Also
     --------
     [Pytest Documentation "Factories as Fixtures"](https://docs.pytest.org/en/stable/how-to/fixtures.html#factories-as-fixtures)
@@ -24,10 +25,10 @@ def fixture_aim2015_B787():
     [Panel 6 in Figure 4 in Dray et al. (2019)](https://doi.org/10.1016/j.tranpol.2019.04.013)
     """
     base_data = {
-        'acft_size_class': 6,
-        'D_climb': 200 * ureg.km,
-        'D_descent': 200 * ureg.km,
-        'PL': (359*95+4500) * ureg.kg,
+        "acft_size_class": 6,
+        "D_climb": 200 * ureg.km,
+        "D_descent": 200 * ureg.km,
+        "PL": (359 * 95 + 4500) * ureg.kg,
     }
 
     cases = {
@@ -39,11 +40,12 @@ def fixture_aim2015_B787():
     def _make_case(d):
         if d not in cases:
             raise ValueError(f"No expected output defined for d={d}")
-        input_data = base_data | {'D_cruise': d - (200+200) * ureg.km}
+        input_data = base_data | {"D_cruise": d - (200 + 200) * ureg.km}
         expected_output = cases[d]
         return input_data, expected_output
-    
+
     return _make_case, list(cases.keys())
+
 
 @pytest.fixture(scope="module")
 def fixture_seymour_B738():
@@ -53,7 +55,7 @@ def fixture_seymour_B738():
     The `cases` dictionary contains:
     - `key`: range [km]
     - `value`: expected value for calculated fuel burn [kg]
-    
+
     See Also
     --------
     [Pytest Documentation "Factories as Fixtures"](https://docs.pytest.org/en/stable/how-to/fixtures.html#factories-as-fixtures)
@@ -63,7 +65,7 @@ def fixture_seymour_B738():
     [Figure 1 in Jupyter Notebook `Reduced_Model_Shape_Exploration.ipynb`](https://github.com/kwdseymour/FEAT/blob/master/Reduced_Model_Shape_Exploration.ipynb)
     """
     base_data = {
-        'acft': 'B738',
+        "acft": "B738",
     }
 
     cases = {
@@ -74,22 +76,22 @@ def fixture_seymour_B738():
     def _make_case(r):
         if r not in cases:
             raise ValueError(f"No expected output defined for r={r}")
-        input_data = base_data | {'R': r}
+        input_data = base_data | {"R": r}
         expected_output = cases[r]
         return input_data, expected_output
-    
+
     return _make_case, list(cases.keys())
+
 
 @pytest.fixture(scope="module")
 def fixture_yanto_B739():
-
     """
     Fixture returning a factory function to generate test cases.
 
     The `cases` dictionary contains:
     - `key`: range [km]
     - `value`: expected value for calculated fuel burn [kg]
-    
+
     See Also
     --------
     [Pytest Documentation "Factories as Fixtures"](https://docs.pytest.org/en/stable/how-to/fixtures.html#factories-as-fixtures)
@@ -99,7 +101,7 @@ def fixture_yanto_B739():
[…]
