+++ src/jetfuelburn/__init__.py
@@ -1,5 +1,6 @@
 # https://pint.readthedocs.io/en/stable/getting/pint-in-your-projects.html#using-pint-in-your-projects
 from pint import UnitRegistry
+
 ureg = UnitRegistry()
 
 __version__ = "2.0.0"

[…]
