### 2026-02-11 3d6fd68
implemented black code style
