+++ tests/fixtures/statistics.py
@@ -1,3 +1,2 @@
 import pytest
 from jetfuelburn import ureg
-
