+++ src/jetfuelburn/data/AeroMaps/__init__.py
@@ -1,7 +1,7 @@
 # https://github.com/AeroMAPS/AeroMAPS/blob/main/aeromaps/resources/data/outputs.json
 
 # Licensing Note:
-# The GPL-3.0 license of this repository does not apply to the re-use of aerodynamic data, since in Switzerland 
+# The GPL-3.0 license of this repository does not apply to the re-use of aerodynamic data, since in Switzerland
 # only "intellectual creations with individual character" are protected by copyright law.
 # Only Swiss copyright law applies, since France and Switzerland are parties to the Berne Convention
 # which states that the extent of copyright protection is governed exclusively by the laws of the country where protection is claimed.
