+++ figures/fuel_calculation_segments/fuel_calculation_segments.py
@@ -1,5 +1,5 @@
-#%%
-# runs code as interactive cell 
+# %%
+# runs code as interactive cell
 # https://code.visualstudio.com/docs/python/jupyter-support-py
 
 # IMPORTS #######################################
@@ -7,8 +7,9 @@
 # plotting
 import matplotlib.pyplot as plt
 import matplotlib.font_manager as font_manager
+
 # unit conversion
-cm = 1/2.54 # for inches-cm conversion
+cm = 1 / 2.54  # for inches-cm conversion
 # time manipulation
 from datetime import datetime
 
@@ -21,10 +22,7 @@ from pathlib import PurePath, Path
 
 # SETUP #########################################
 
-plt.rcParams.update({
-    "font.family": "Arial",
-    'font.size': 12
-})
+plt.rcParams.update({"font.family": "Arial", "font.size": 12})
 
 # DATA IMPORT ###################################
 
@@ -43,38 +41,41 @@ alternate = (850, 0)
 # SETUP ######################
 
 fig, ax = plt.subplots(
-    num = 'main',
-    nrows = 1,
-    ncols = 1,
-    dpi = 300,
-    figsize=(20*cm, 10*cm), # A4=(210x297)mm,
-    sharex=True
+    num="main",
+    nrows=1,
+    ncols=1,
+    dpi=300,
+    figsize=(20 * cm, 10 * cm),  # A4=(210x297)mm,
+    sharex=True,
 )
 
 # AXIS LIMITS ################
 
 ax.set_xlim(0, 900)
-ax.set_ylim(0,32000)
+ax.set_ylim(0, 32000)
 
 # TICKS AND LABELS ###########
 
 ax.minorticks_on()
-ax.tick_params(axis='x', which='both', bottom=True)
-ax.tick_params(axis='y', which='both', bottom=True)
+ax.tick_params(axis="x", which="both", bottom=True)
+ax.tick_params(axis="y", which="both", bottom=True)
 
 import matplotlib.ticker as ticker
+
+
 def thousand_formatter(value, tick_number):
     """
     Formats the tick label with thousand separators: 1000 = 1'000.
     """
     return f"{int(value):,}".replace(",", "'")
 
+
 ax.yaxis.set_major_formatter(ticker.FuncFormatter(thousand_formatter))
 
 # GRIDS ######################
 
-ax.grid(which='both', axis='y', linestyle='-', linewidth=0.5)
-ax.grid(which='major', axis='x', linestyle='--', linewidth=0.5)
+ax.grid(which="both", axis="y", linestyle="-", linewidth=0.5)
+ax.grid(which="major", axis="x", linestyle="--", linewidth=0.5)
 
 # AXIS LABELS ################
 
@@ -86,75 +87,86 @@ ax.set_xlabel("Distance [nmi]")
 # Plot points
 points_main = [takeoff, cruise_1, cruise_2, destination]
 x_climb, y_climb = zip(*points_main)
-ax.plot(x_climb, y_climb, marker='o', linestyle='-', color='k', label='Main Segment (Fuel burned)')
+ax.plot(
+    x_climb,
+    y_climb,
+    marker="o",
+    linestyle="-",
+    color="k",
+    label="Main Segment (Fuel burned)",
+)
 
 points_alternate = [destination, cruise_3, cruise_4, alternate]
 x_alternate, y_alternate = zip(*points_alternate)
-ax.plot(x_alternate, y_alternate, marker='o', linestyle='--', color='k', label='Diversion Segment (Fuel uplifted)')
+ax.plot(
+    x_alternate,
+    y_alternate,
+    marker="o",
+    linestyle="--",
[…]
