+++ src/jetfuelburn/data/Airports/__init__.py
@@ -1,7 +1,7 @@
 # https://github.com/ip2location/ip2location-iata-icao
 
 # Licensing Note:
-# The CC-BY-SA-4.0 license of this repository does not apply to the re-use of basic geo-spatial data, since in Switzerland 
+# The CC-BY-SA-4.0 license of this repository does not apply to the re-use of basic geo-spatial data, since in Switzerland
 # only "intellectual creations with individual character" are protected by copyright law.
 # Only Swiss copyright law applies, since Malaysia and Switzerland are parties to the Berne Convention
 # which states that the extent of copyright protection is governed exclusively by the laws of the country where protection is claimed.

[…]
