### 2026-02-11 6bacbf6
added black code style linting workflow
