+++ .github/workflows/black.yml
@@ -1,0 +1,11 @@
+# https://black.readthedocs.io/en/stable/integrations/github_actions.html
+name: Lint
+
+on: [push, pull_request]
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v5
+      - uses: psf/black@stable
