+++ src/jetfuelburn/combined.py
@@ -1,327 +1,0 @@
-from jetfuelburn import ureg
-from jetfuelburn.rangeequation import calculate_fuel_consumption_range_equation
-
-
-@ureg.check(
-    "[mass]",  # payload
-    "[mass]",  # oew
-    "[]",  # number_of_engines
-    "[mass]/[time]",  # fuel_flow_per_engine_idle
-    "[mass]/[time]",  # fuel_flow_per_engine_takeoff
-    "[speed]",  # speed_cruise
-    "[time]/[length]",  # tsfc_cruise, [mg/Ns] = s/m
-    "[]",  # lift_to_drag
-    None,  # dict_climb_segments_origin_to_destination
-    None,  # dict_descent_segments_origin_to_destination
-    "[length]",  # R_cruise_origin_to_destination
-    "[time]",  # time_taxi
-    None,  # dict_climb_segments_destination_to_alternate
-    None,  # dict_descent_segments_destination_to_alternate
-    "[length]",  # R_cruise_destination_to_alternate
-)
-def calculate_fuel_consumption_combined_model(
-    payload: float,
-    oew: float,
-    number_of_engines: int,
-    fuel_flow_per_engine_idle: float,
-    fuel_flow_per_engine_takeoff: float,
-    speed_cruise: float,
-    tsfc_cruise: float,
-    lift_to_drag: float,
-    dict_climb_segments_origin_to_destination: dict,
-    dict_descent_segments_origin_to_destination: dict,
-    R_cruise_origin_to_destination: float,
-    time_taxi: float = 26 * ureg.min,
-    dict_climb_segments_destination_to_alternate: dict = None,
-    dict_descent_segments_destination_to_alternate: dict = None,
-    R_cruise_destination_to_alternate: float = 0 * ureg.km,
-) -> dict:
-    r"""
-    Given aircraft performance parameters, payload, climb/descent segment information and a flight distances,
-    calculates the required fuel mass $m_F$ for a given mission.
-
-    ![Diagram](../_static/combined.svg)
-
-    For `dict_climb_segments`, only 'takeoff' is required; other segments are optional.
-    For `dict_descent_segments`, only 'approach' is required; other segments are optional.
-
-    ```python
-    dict_climb_segments = {
-        'takeoff': <flight_segment_dict>,
-        'climb_to_10000ft': <flight_segment_dict>,
-        'climb_10000ft_to_20000ft': <flight_segment_dict>,
-        (...)
-    }
-    dict_descent_segments = {
-        'descent_cruise_to_10000ft': <flight_segment_dict>,
-        'descent_10000ft_to_5000ft': <flight_segment_dict>,
-        'approach': <flight_segment_dict>,
-        (...)
-    }
-    ```
-
-    Each `flight_segment_dict` must contain a time key-value pair and either an absolute
-    `fuel_flow_per_engine` value-pair or a relative `fuel_flow_per_engine_relative_to_takeoff` value-pair.
-
-    ```python
-    flight_segment_dict_absolute = {
-        'time': 5*ureg.min,
-        'fuel_flow_per_engine': 0.205*(ureg.kg/ureg.sec),
-    }
-    flight_segment_dict_relative = {
-        'time': 180*ureg.s,
-        'fuel_flow_per_engine_relative_to_takeoff': 0.82
-    }
-    ```
-
-    Notes
-    -----
-
-    - [ICAO Engine Emissions Databank LTO segments](https://web.archive.org/web/20220725161602/https://www.icao.int/environmental-protection/Documents/EnvironmentalReports/2022/ENVReport2022_Art17.pdf):
-
-    | Segment Name | Duration (official LTO cycle) | Thrust  |
-    |--------------|-------------------------------|---------|
[…]
