+++ .github/workflows/coverage.yml
@@ -1,0 +1,38 @@
+name: Coverage Badge
+on:
+  push:
+    branches:
+      - main
+
+permissions:
+  contents: write # Required to push the badge
+
+jobs:
+  update-badge:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+
+      - name: Setup Python
+        uses: actions/setup-python@v6
+        with:
+          python-version: '3.10'
+
+      - name: Install & Test
+        run: |
+          pip install pytest pytest-cov
+          pytest --cov=./ 
+
+      # 1. Generate the badge
+      - name: Generate Badge
+        uses: tj-actions/coverage-badge-py@v2
+        with:
+          output: coverage.svg
+
+      # 2. Push to a separate 'badges' branch (Clean & Org-owned)
+      - name: Deploy Badge
+        uses: JamesIves/github-pages-deploy-action@v4
+        with:
+          branch: badges # The branch the badge will live on
+          folder: .      # The folder to upload ('.' captures the SVG we just made)
+          clean: false   # Don't delete other files in that branch if they exist
