+++ .github/workflows/coverage.yml
@@ -3,9 +3,10 @@ on:
   push:
     branches:
       - main
+  workflow_dispatch:
 
 permissions:
-  contents: write # Required to push the badge
+  contents: write
 
 jobs:
   update-badge:
@@ -16,7 +17,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v6
         with:
-          python-version: '3.10'
+          python-version: '3.14'
 
       - name: Install & Test
         run: |
