+++ docs/api/closedform.md
@@ -1,1 +1,0 @@
-::: jetfuelburn.closedform
