### 2026-02-11 e49450e
improved test coverage
