+++ tests/test_allocation.py
@@ -1,44 +1,138 @@
 import pytest
 
-from .fixtures.allocation import allocation_two_classes, allocation_all_classes
+from .fixtures.allocation import (
+    allocation_one_class,
+    allocation_two_classes,
+    allocation_all_classes,
+)
 
 from jetfuelburn.utility.allocation import footprint_allocation_by_area
 
 
-@pytest.mark.parametrize(
-    "fixture_name", ["allocation_two_classes", "allocation_all_classes"]
-)
-def test_footprint_allocation_by_area(request, fixture_name):
-    fixture = request.getfixturevalue(fixture_name)
-    input_data = fixture
+class TestFootprintAllocationByArea:
 
-    dict_calculated_fuel = footprint_allocation_by_area(
-        fuel_per_flight=input_data["fuel_per_flight"],
-        size_factor_eco=input_data["size_factor_eco"],
-        size_factor_premiumeco=input_data["size_factor_premiumeco"],
-        size_factor_business=input_data["size_factor_business"],
-        size_factor_first=input_data["size_factor_first"],
-        seats_eco=input_data["seats_eco"],
-        seats_premiumeco=input_data["seats_premiumeco"],
-        seats_business=input_data["seats_business"],
-        seats_first=input_data["seats_first"],
-        load_factor_eco=input_data["load_factor_eco"],
-        load_factor_premiumeco=input_data["load_factor_premiumeco"],
-        load_factor_business=input_data["load_factor_business"],
-        load_factor_first=input_data["load_factor_first"],
+    def test_footprint_allocation_by_area_input_validation(self):
+        """Tests that invalid inputs raise ValueError."""
+        with pytest.raises(ValueError, match="Fuel burn per flight must be >0."):
+            footprint_allocation_by_area(
+                fuel_per_flight=-1000,
+                size_factor_eco=1,
+                size_factor_premiumeco=0,
+                size_factor_business=1.5,
+                size_factor_first=0,
+                seats_eco=154,
+                seats_premiumeco=0,
+                seats_business=24,
+                seats_first=0,
+                load_factor_eco=0.9,
+                load_factor_premiumeco=0,
+                load_factor_business=0.5,
+                load_factor_first=0,
+            )
+
+    def test_invalid_load_factor(self):
+        """Tests that invalid load factors raise ValueError."""
+        with pytest.raises(
+            ValueError, match="Load factor \\(economy class\\) must be between 0 and 1."
+        ):
+            footprint_allocation_by_area(
+                fuel_per_flight=14000,
+                size_factor_eco=1,
+                size_factor_premiumeco=0,
+                size_factor_business=1.5,
+                size_factor_first=0,
+                seats_eco=154,
+                seats_premiumeco=0,
+                seats_business=24,
+                seats_first=0,
+                load_factor_eco=1.1,
+                load_factor_premiumeco=0,
+                load_factor_business=0.5,
+                load_factor_first=0,
+            )
+
+    def test_invalid_size_factor_eco(self):
+        """Tests that an invalid size_factor_eco raises a ValueError."""
+        with pytest.raises(
+            ValueError,
+            match="Economy class must have size factor 1 and at least one seat.",
+        ):
+            footprint_allocation_by_area(
+                fuel_per_flight=14000,
+                size_factor_eco=1.1,
+                size_factor_premiumeco=0,
+                size_factor_business=1.5,
+                size_factor_first=0,
+                seats_eco=154,
+                seats_premiumeco=0,
+                seats_business=24,
+                seats_first=0,
+                load_factor_eco=0.9,
+                load_factor_premiumeco=0,
+                load_factor_business=0.5,
+                load_factor_first=0,
[…]

[…]
