### 2026-02-11 d5ce6d0
moved shield around
