+++ tests/test_statistics.py
@@ -26,9 +26,9 @@ class TestUsdot:
         acft = "B787-800 Dreamliner"
         R = 1000 * ureg.nmi
         W = 1000 * ureg.kg
-        
+
         fuel = usdot.calculate_fuel_consumption_per_weight(year, acft, R, W)
-        
+
         assert fuel.magnitude > 0
         assert fuel.units == ureg.kg
 
@@ -37,37 +37,58 @@ class TestUsdot:
         year = 2024
         acft = "B787-800 Dreamliner"
         R = 1000 * ureg.nmi
-        
+
         fuel = usdot.calculate_fuel_consumption_per_seat(year, acft, R)
-        
+
         assert fuel.magnitude > 0
         assert fuel.units == ureg.kg
 
     def test_fuel_consumption_per_weight_error_handling(self):
         """Test error handling for fuel consumption per weight."""
-        with pytest.raises(ValueError, match="Range and/or weight must not be negative."):
-            usdot.calculate_fuel_consumption_per_weight(2024, "B787-800 Dreamliner", -100 * ureg.km, 1000 * ureg.kg)
-            
-        with pytest.raises(ValueError, match="Range and/or weight must not be negative."):
-            usdot.calculate_fuel_consumption_per_weight(2024, "B787-800 Dreamliner", 100 * ureg.km, -1000 * ureg.kg)
+        with pytest.raises(
+            ValueError, match="Range and/or weight must not be negative."
+        ):
+            usdot.calculate_fuel_consumption_per_weight(
+                2024, "B787-800 Dreamliner", -100 * ureg.km, 1000 * ureg.kg
+            )
+
+        with pytest.raises(
+            ValueError, match="Range and/or weight must not be negative."
+        ):
+            usdot.calculate_fuel_consumption_per_weight(
+                2024, "B787-800 Dreamliner", 100 * ureg.km, -1000 * ureg.kg
+            )
 
         with pytest.raises(ValueError, match="No data available for year '1800'."):
-            usdot.calculate_fuel_consumption_per_weight(1800, "B787-800 Dreamliner", 100 * ureg.km, 1000 * ureg.kg)
+            usdot.calculate_fuel_consumption_per_weight(
+                1800, "B787-800 Dreamliner", 100 * ureg.km, 1000 * ureg.kg
+            )
 
-        with pytest.raises(ValueError, match="US DOT Aircraft Designator 'ufo' not found in model data."):
-            usdot.calculate_fuel_consumption_per_weight(2024, "ufo", 100 * ureg.km, 1000 * ureg.kg)
+        with pytest.raises(
+            ValueError,
+            match="US DOT Aircraft Designator 'ufo' not found in model data.",
+        ):
+            usdot.calculate_fuel_consumption_per_weight(
+                2024, "ufo", 100 * ureg.km, 1000 * ureg.kg
+            )
 
     def test_fuel_consumption_per_seat_error_handling(self):
         """Test error handling for fuel consumption per seat."""
         with pytest.raises(ValueError, match="Range must not be negative."):
-            usdot.calculate_fuel_consumption_per_seat(2024, "B787-800 Dreamliner", -100 * ureg.km)
+            usdot.calculate_fuel_consumption_per_seat(
+                2024, "B787-800 Dreamliner", -100 * ureg.km
+            )
 
         with pytest.raises(ValueError, match="No data available for year '1800'."):
-            usdot.calculate_fuel_consumption_per_seat(1800, "B787-800 Dreamliner", 100 * ureg.km)
+            usdot.calculate_fuel_consumption_per_seat(
+                1800, "B787-800 Dreamliner", 100 * ureg.km
+            )
 
[…]
