### 2026-02-20 ce75542
fixed test failures and added pre-commit hook
