+++ src/jetfuelburn/rangeequation.py
@@ -1,4 +1,3 @@
-# %%
 import pint
 from jetfuelburn import ureg
 import math
