### 2026-02-20 6e3b833
fixed test failures and added pre-commit hook
