+++ src/jetfuelburn/__init__.py
@@ -3,4 +3,4 @@ from pint import UnitRegistry
 
 ureg = UnitRegistry()
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"
