+++ .github/workflows/update_license.yml
@@ -4,6 +4,7 @@ name: Update copyright year(s) in license file
 on:
     schedule:
         - cron: '0 3 1 1 *' # 03:00 AM on January 1
+    workflow_dispatch:
 
 jobs:
     update-license-year:
@@ -15,3 +16,4 @@ jobs:
             - uses: FantasticFiasco/action-update-license-year@v3
               with:
                   token: ${{ secrets.GITHUB_TOKEN }}
+                  path: LICENSE.txt
