+++ .github/workflows/black.yml
@@ -8,4 +8,4 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v5
-      - uses: psf/black@stable
+      - uses: psf/black@26.1.0
