### 2026-02-20 e95c419
gemini draft of function
