+++ src/jetfuelburn/rangeequation.py
@@ -5,6 +5,9 @@ from typing import Callable
 
 from jetfuelburn.utility.physics import (
     _calculate_atmospheric_density,
+    _calculate_atmospheric_temperature,
+    _calculate_speed_of_sound,
+    _calculate_dynamic_pressure,
     _calculate_airspeed_from_mach,
 )
 from jetfuelburn.utility.code import (
@@ -601,3 +604,183 @@ def calculate_fuel_consumption_breguet(
     else:
         m_fuel = m_after_cruise * (math.exp((R * TSFC * ureg.gravity) / (LD * V)) - 1)
         return m_fuel.to("kg")
+
+
+@ureg.check(
+    "[length]",
+    "[length]",
+    "[]",
+    "[mass]",
+    "[area]",
+    "[]",
+    "[]",
+    "[]",
+    "[]",
+    "[]",
+    "[time]/[length]",
+    "[]",
+    "[length]",
+    "[]",
+    "[]",
+    "[]",
+)
+def calculate_fuel_consumption_cavcar(
+    R: pint.Quantity,
+    h: pint.Quantity,
+    M: float,
+    m_after_cruise: pint.Quantity,
+    S: pint.Quantity,
+    AR: float,
+    e: float,
+    t_c: float,
+    lambda_deg: float,
+    kappa_a: float,
+    TSFC_ref: pint.Quantity,
+    M_ref: float,
+    h_ref: pint.Quantity,
+    beta: float = 0.5,
+    CDmin: float = 0.015,
+    CLmin: float = 0.1,
+) -> pint.Quantity:
+    r"""
+    Given a flight distance (=range) $R$ and aircraft performance parameters,
+    returns the fuel mass burned during the flight $m_f$ [kg] based on a
+    constant altitude-constant Mach number cruise flight schedule.
+
+    The calculation follows the approximate solution proposed by Cavcar (2006),
+    which accounts for compressibility effects on the drag polar and the variation
+    of thrust-specific fuel consumption (TSFC) with Mach number and altitude.
+
+    Fuel mass is calculated from the fuel weight fraction $\zeta = W_f / W_0$:
+
+    $$
+        W_f = \frac{R C_1}{\frac{a M^{1-\beta}}{c_{t_0} \sqrt{\theta} q S} - R C_2}
+    $$
+
+    where:
+
+    | Symbol      | Description                                                |
+    |-------------|------------------------------------------------------------|
+    | $R$         | Range of the aircraft (=mission distance)                  |
+    | $W_f$       | Fuel weight burned                                         |
+    | $a$         | Speed of sound at cruise altitude                          |
+    | $M$         | Mach number during cruise                                  |
+    | $\beta$     | TSFC Mach number exponent (typically 0.5 for turbofans)    |
+    | $c_{t_0}$   | TSFC constant at cruise altitude                           |
+    | $\theta$    | Relative temperature at cruise altitude                    |
+    | $q$         | Dynamic pressure at cruise altitude                        |
+    | $S$         | Wing reference area                                        |
+    | $C_1, C_2$  | Coefficients derived from the drag polar approximation     |
+
+    Notes
+    -----
+    This function implements the solution for $M < M_{crit}$. For $M > M_{crit}$,
+    the drag polar coefficients $C_{D_0}'$, $K_1'$, and $K_2'$ are modified to
+    account for wave drag.
+
+    References
+    ----------
+    Cavcar, A. (2006).
[…]
