### 2026-03-04 efc3b63
docs: switch from mkdocs to zensical
