+++ .readthedocs.yaml
@@ -1,15 +1,16 @@
 version: 2
 
 build:
-  os: "ubuntu-lts-latest"
+  os: "ubuntu-24.04"
   tools:
     python: "latest"
-
-mkdocs:
-  configuration: mkdocs.yaml
-
-python:
-  install:
-    - requirements: docs/_requirements.txt
-    - method: pip
-      path: .
+  jobs:
+    install:
+      - pip install -r docs/_requirements.txt
+      - pip install .
+    build:
+      html:
+        - zensical build
+    post_build:
+      - mkdir -p $READTHEDOCS_OUTPUT/html/
+      - cp -r site/* $READTHEDOCS_OUTPUT/html/
