+++ mkdocs.yaml
@@ -1,114 +1,0 @@
-# https://www.mkdocs.org/user-guide/configuration/
-site_name: JetFuelBurn Documentation
-repo_url: https://github.com/sustainableaviation/jetfuelburn
-copyright: © Paul Scherrer Institute (PSI) and Contributors | Code under MIT License
-nav:
-  - Home: 'index.md'
-  - About: 'about.md'
-  - Comparing Models:
-    - Aircraft Type Matrix: 'comparison/comparison.md'
-  - Getting Started:
-    - Installation: 'installation.md'
-    - Working with Physical Units: 'interactive/units.md'
-    - Fuel Burn Calculations: 'interactive/fuel.md'
-    - Cabin Class Fuel Allocation: 'interactive/allocation.md'
-  - API Reference:
-    - PL/R-Diagrams: 'api/diagrams.md'
-    - Range Equation: 'api/rangeequation.md'
-    - Statistics Models: 'api/statistics.md'
-    - Reduced Order Models: 'api/reducedorder.md'
-    #- Combined Models: 'api/combined.md'
-    #- Closed Form Models: 'api/closedform.md' 
-    - Cabin Class Fuel Allocation: 'api/allocation.md'
-  - API Reference (Utility):
-    - Aerodynamics: 'api/aerodynamics.md'
-    - Engines: 'api/engines.md'
-    - Physics: 'api/physics.md'
-    - Geography: 'api/geography.md'
-    - Math: 'api/math.md'
-  - API Reference (External):
-    - EU Flight Emissions Label: 'api/eu_fel.md'
-    - Reduced Order Models: 'api/reducedorder_external.md'
-    - Closed Form Models: 'api/closedform_external.md'
-  - FAQ: 'faq.md'
-  - Changelog: 'changelog.md'
-theme: # https://www.mkdocs.org/user-guide/choosing-your-theme/#readthedocs
-  name: material
-  custom_dir: docs/_overrides
-  logo: _static/logos/jetfuelburn_logo.svg
-  favicon: _static/logos/jetfuelburn_logo.svg
-  highlightjs: true
-  navigation_depth: 4
-  features:
-    - navigation.instant
-    - navigation.instant.progress
-    - navigation.expand
-    - navigation.path
-    - navigation.indexes
-    - navigation.top
-    - content.footnote.tooltips
-  palette:
-    - scheme: default
-      toggle:
-        icon: material/brightness-7 
-        name: Switch to dark mode
-    - scheme: slate
-      toggle:
-        icon: material/brightness-4
-        name: Switch to light mode
-plugins:
-  - search
-  - mkdocstrings: # https://mkdocstrings.github.io/python/usage/
-      handlers:
-        python:
-          options:
-            docstring_style: numpy
-            show_root_toc_entry: True
-            show_root_heading: true
-            heading_level: 2
-            show_source: true
-            filters: []
-            separate_signature: true
-  - markdown-exec
-  - autorefs
-  - include-markdown
-  - mkdocs-jupyter:
-      include: ["*.ipynb"]
-      execute: False
-      show_input: True
-      include_source: True
-  - mkdocs-pdf
-  
-markdown_extensions:
-  - admonition
-  - pymdownx.superfences
-  - pymdownx.details
-  - toc:
-      permalink: true
-  - pymdownx.arithmatex:
-      generic: true
-  - pymdownx.snippets
-  - pymdownx.highlight:
-      use_pygments: false
-  - attr_list
-
-extra_javascript:
-  - _javascripts/katex.js
[…]
