+++ docs/api/mapping.md
@@ -1,0 +1,1 @@
+::: jetfuelburn.utility.mapping
