+++ pyproject.toml
@@ -10,10 +10,12 @@ authors = [
 maintainers = [
   { name="Michael Weinold", email="michaelphilippweinold+jetfuelburn@gmail.com" }
 ]
+
 description = "A Python package for calculating fuel burn of commercial aircraft."
 readme = "README.md"
 license = {text = "MIT"}
 dynamic = ["version"]
+
 # https://pypi.org/classifiers/
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -27,21 +29,27 @@ classifiers = [
     "Operating System :: OS Independent",
     "Topic :: Scientific/Engineering"
 ]
+
 requires-python = ">=3.11"
-dependencies = [
-    "pint",
-]
 
 [project.urls]
 source = "https://github.com/sustainableaviation/jetfuelburn"
 homepage = "https://jetfuelburn.readthedocs.io"
 tracker = "https://github.com/sustainableaviation/jetfuelburn/issues"
 
+dependencies = [
+    "pint",
+]
+
 [project.optional-dependencies]
 testing = [
     "pytest",
     "pytest-cov",
     "python-coveralls",
+]
+plotting = [
+    "folium",
+    "pandas",
 ]
 
 [tool.setuptools]
