### 2026-03-15 b030926
added draft and figures
