+++ figures/ofp/ofp.py
@@ -1,0 +1,213 @@
+# %%
+# runs code as interactive cell
+# https://code.visualstudio.com/docs/python/jupyter-support-py
+
+# IMPORTS #######################################
+
+# sys
+import os
+from pathlib import Path
+
+# plotting
+import matplotlib.pyplot as plt
+import matplotlib.ticker as ticker
+
+# unit conversion
+cm = 1 / 2.54  # for inches-cm conversion
+
+# data science
+import numpy as np
+import pandas as pd
+
+# SETUP #########################################
+
+plt.rcParams.update({"font.family": "Arial", "font.size": 12})
+
+# DATA IMPORT ###################################
+
+# (Placeholder for data if needed in future)
+
+# DATA MANIPULATION #############################
+
+# FIGURE ########################################
+
+# SETUP ######################
+
+fig, ax = plt.subplots(
+    num="ofp",
+    nrows=1,
+    ncols=1,
+    dpi=300,
+    figsize=(20 * cm, 10 * cm),  # A4=(210x297)mm,
+)
+
+# AXIS LIMITS ################
+
+y_max = 35000
+x_max = 1000
+
+ax.set_xlim(-50, x_max)
+ax.set_ylim(0, y_max)
+
+# TICKS AND LABELS ###########
+
+ax.minorticks_on()
+ax.tick_params(axis="both", which="both", labelbottom=False, labelleft=False)
+
+# GRIDS ######################
+
+ax.grid(which="major", axis="both", linestyle="-", linewidth=0.5)
+ax.grid(which="minor", axis="both", linestyle=":", linewidth=0.5)
+
+# AXIS LABELS ################
+
+ax.set_xlabel("Distance or Time")
+ax.set_ylabel("Altitude")
+ax.yaxis.set_label_coords(-0.02, 0.5)
+
+# PLOTTING ###################
+
+# Lower third background: light orange
+ax.axhspan(0, y_max / 3, facecolor='moccasin', alpha=0.6, label="Climb Regime 1")
+
+# Upper two thirds background: light red
+ax.axhspan(y_max / 3, y_max, facecolor='mistyrose', alpha=0.6, label="Climb Regime 2")
+
+# TRAJECTORY ARROWS ##########
+
+border_y = y_max / 3
+mid_x = 205
+wp2_upper = (800, 30000)
+
+# Orange arrow: Up to the border
+ax.annotate(
+    "",
+    xy=(mid_x, border_y),
+    xytext=(0, 0),
+    arrowprops=dict(arrowstyle="->", color="orange", lw=2),
+)
+
+# Red arrow: From the border to the upper WP2
+ax.annotate(
+    "",
+    xy=wp2_upper,
+    xytext=(mid_x, border_y),
+    arrowprops=dict(arrowstyle="->", color="red", lw=2),
+)
+
+# STEEPER TRAJECTORY (BLACK) ####
+
+# First steeper segment (from 0,0)
+mid_x_steep = 150 # Reaches border earlier than mid_x (205)
+ax.annotate(
+    "",
+    xy=(mid_x_steep, border_y),
+    xytext=(0, 0),
+    arrowprops=dict(arrowstyle="->", color="black", lw=2, linestyle='--'),
+)
+
+# Second steeper segment (to 30000 alt)
+# Reaches 30000 alt at x=600 (before WP2 x=800)
+target_x_steep = 600
+ax.annotate(
+    "",
+    xy=(target_x_steep, 30000),
+    xytext=(mid_x_steep, border_y),
+    arrowprops=dict(arrowstyle="->", color="black", lw=2, linestyle='--'),
+)
[…]
