+++ src/jetfuelburn/utility/ofp.py
@@ -1,0 +1,214 @@
+import polars as pl
+import math
+
+def generate_4d_trajectory(df_ofp: pl.DataFrame, perf_data: dict, resolution_min: float = 1.0) -> pl.DataFrame:
+    """
+    Generate a four-dimensional (4D) trajectory from a flight plan using aircraft climb/descent performance data.
+    
+    The function interpolates positions (lat/lon) and altitudes (alt) mapping to a high-resolution timestamp
+    given the times to arrive at specific waypoints. Missing `timeto` values are inferred primarily by utilizing
+    the vertical rates defined in `perf_data` to calculate the required time to bridge altitude changes between
+    known points.
+    
+    If an aircraft climbs or descends and reaches the altitude of the next waypoint before arriving at the waypoint's
+    target time (inferred or given), it will level off at that target altitude until the waypoint is reached.
+
+    Args:
+        df_ofp (pl.DataFrame): A Polars DataFrame flight plan containing at least the columns: 
+            ['waypoint', 'lat', 'lon', 'timeto', 'altitude'] or ['waypoint', 'lat', 'lon', 'timeto', 'alt'].
+            'timeto' is the cumulative flight time in minutes from the origin. 
+        perf_data (dict): A dictionary describing the rate of climb (ROC) and rate of descent (ROD)
+            (in feet per minute) for different altitude brackets. Example:
+            {
+                "climb": [
+                    {"min_alt": 0, "max_alt": 10000, "rate": 800},
+                    {"min_alt": 10000, "max_alt": 20000, "rate": 500}
+                ],
+                "descent": [
+                    {"min_alt": 0, "max_alt": 10000, "rate": 1000},
+                    {"min_alt": 10000, "max_alt": 20000, "rate": 800}
+                ]
+            }
+        resolution_min (float, optional): The time resolution in minutes for the output trajectory. Defaults to 1.0.
+
+    Returns:
+        pl.DataFrame: A high-resolution trajectory DataFrame containing the columns:
+            ['timestamp', 'lat', 'lon', 'alt']
+    """
+    
+    # Standardize altitude column name
+    alt_col = 'alt' if 'alt' in df_ofp.columns else 'altitude'
+    if alt_col not in df_ofp.columns:
+        raise ValueError("DataFrame must contain either an 'alt' or 'altitude' column.")
+        
+    req_cols = ['lat', 'lon', 'timeto', alt_col]
+    for col in req_cols:
+        if col not in df_ofp.columns:
+            raise ValueError(f"DataFrame is missing required column: {col}")
+            
+    # Extract lists to iterate and mutate missing times
+    lats = df_ofp['lat'].to_list()
+    lons = df_ofp['lon'].to_list()
+    times = df_ofp['timeto'].to_list()
+    alts = df_ofp[alt_col].to_list()
+    
+    # Helper to calculate time required to traverse from start_alt to end_alt
+    def calculate_transit_time(start_alt, end_alt):
+        if start_alt == end_alt:
+            return 0.0
+            
+        is_climb = end_alt > start_alt
+        brackets = perf_data.get('climb', []) if is_climb else perf_data.get('descent', [])
+        
+        # Sort brackets by min_alt ascending for climb, descending max_alt for descent
+        if is_climb:
+            brackets = sorted(brackets, key=lambda x: x['min_alt'])
+        else:
+            brackets = sorted(brackets, key=lambda x: x['max_alt'], reverse=True)
+            
+        total_time = 0.0
+        current_alt = start_alt
+        
+        for bracket in brackets:
+            b_min = bracket['min_alt']
+            b_max = bracket['max_alt']
+            rate = bracket['rate']
+            
+            if is_climb:
+                if current_alt >= end_alt:
+                    break
[…]
