### 2026-03-15 90aca99
updated functions and figures
