+++ tests/test_mapping.py
@@ -2,89 +2,149 @@ import pytest
 import os
 import tempfile
 import csv
-from ipyleaflet import Map, Polyline, Marker, CircleMarker
+import ipyleaflet
 from jetfuelburn.utility.mapping import plot_ofp
+
 
 class TestMapping:
 
     @pytest.fixture
     def sample_csv(self):
-        """Create a temporary CSV file for testing."""
+        """Create a temporary CSV file with two valid rows and two invalid rows."""
         content = [
             ["waypoint", "lat", "lon", "custom_label"],
             ["ZRH", 47.4582, 8.5555, "Zurich"],
             ["SFO", 37.6188, -122.3751, "San Francisco"],
             ["Invalid", "not_a_float", 0.0, "Skip me"],
-            ["Missing", 0.0, "", "Skip me too"]
+            ["Missing", 0.0, "", "Skip me too"],
         ]
-        with tempfile.NamedTemporaryFile(mode='w', suffix='.csv', delete=False, newline='') as f:
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".csv", delete=False, newline="") as f:
             writer = csv.writer(f)
             writer.writerows(content)
             temp_path = f.name
-        
+
         yield temp_path
-        
+
+        if os.path.exists(temp_path):
+            os.remove(temp_path)
+
+    @pytest.fixture
+    def sample_csv_custom_cols(self):
+        """Create a temporary CSV file using non-default lat/lon column names."""
+        content = [
+            ["waypoint", "latitude", "longitude"],
+            ["ZRH", 47.4582, 8.5555],
+            ["SFO", 37.6188, -122.3751],
+        ]
+        with tempfile.NamedTemporaryFile(mode="w", suffix=".csv", delete=False, newline="") as f:
+            writer = csv.writer(f)
+            writer.writerows(content)
+            temp_path = f.name
+
+        yield temp_path
+
         if os.path.exists(temp_path):
             os.remove(temp_path)
 
     def test_plot_ofp_returns_map(self, sample_csv):
         """Test that plot_ofp returns an ipyleaflet Map."""
         m = plot_ofp(sample_csv)
-        assert isinstance(m, Map)
+        assert isinstance(m, ipyleaflet.Map)
 
     def test_plot_ofp_layer_counts(self, sample_csv):
-        """Test that the correct number of layers are added to the map."""
+        """Test that the correct number of layers are added to the map.
+
+        With 2 valid rows (ZRH, SFO) every row is either start or end, so:
+          1 TileLayer (base)
+        + 1 Polyline
+        + 1 Marker  (ZRH – start endpoint)
+        + 1 Marker  (ZRH – label DivIcon)
+        + 1 Marker  (SFO – end endpoint)
+        + 1 Marker  (SFO – label DivIcon)
+        = 6 layers
+        """
         m = plot_ofp(sample_csv)
-        
-        # layers in ipyleaflet Map (current implementation):
-        # 1. TileLayer (base)
-        # 2. Polyline
-        # 3. CircleMarker (ZRH)
-        # 4. CircleMarker (SFO)
-        
-        layer_types = [type(layer) for layer in m.layers]
+
+        layer_types = [type(layer).__name__ for layer in m.layers]
         print(f"Layers found: {layer_types}")
-        
-        assert any(isinstance(l, Polyline) for l in m.layers)
-        assert any(isinstance(l, CircleMarker) for l in m.layers)
-        # Markers were removed by user in Step 264
-        assert not any(isinstance(l, Marker) for l in m.layers)
-        
-        assert len(m.layers) == 4
+
+        assert any(isinstance(l, ipyleaflet.Polyline) for l in m.layers)
+        assert any(isinstance(l, ipyleaflet.Marker) for l in m.layers)
[…]
