+++ CHANGES.md
@@ -3,6 +3,14 @@
 The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `2.1.0` (15. March 2026)
+
+## Improvements
+
+- Added a new `mapping` utility module for mapping-related calculations.
+- Added a new `ofp` utility module for flight plan-related calculations.
+- Updated the installation instructions to include optional dependencies.
+
 ## `2.0.1` (20. February 2026)
 
 ### Fixes
