+++ src/jetfuelburn/utility/mapping.py
@@ -1,15 +1,21 @@
-import csv
-import io
-import urllib.request
-from ipyleaflet import Map, Polyline, CircleMarker
-from ipywidgets import HTML
+try:
+    import polars as pl
+    import ipyleaflet
+    import ipywidgets
+except ImportError as e:
+    raise ImportError(
+        f"Optional dependency missing: {e}. "
+        "Install all required packages with: pip install jetfuelburn[optionaldependencies]"
+    ) from e
 
 
 def plot_ofp(
     csv_path: str,
     label_column: str = "waypoint",
+    lat_col: str = "lat",
+    lon_col: str = "lon",
     output_html: str | None = None,
-) -> Map:
+) -> ipyleaflet.Map:
     r"""
     Reads a CSV file containing flight trajectory data from operational flight plans
     (OFP) and plots it on an interactive [ipyleaflet](https://ipyleaflet.readthedocs.io/) map.
@@ -20,14 +26,20 @@ def plot_ofp(
         Path to the CSV file or URL containing trajectory data.
         The CSV must contain at least the following columns:
 
-        | Column    | Type    | Description                                      |
-        |-----------|---------|--------------------------------------------------|
-        | `lat`     | `float` | Latitude of the waypoint (decimal degrees).      |
-        | `lon`     | `float` | Longitude of the waypoint (decimal degrees).     |
-        | `waypoint`| `str`   | Short name or designator for the waypoint.       |
+        | Column     | Type    | Description                                      |
+        |------------|---------|--------------------------------------------------|
+        | `lat`      | `float` | Latitude of the waypoint (decimal degrees).      |
+        | `lon`      | `float` | Longitude of the waypoint (decimal degrees).     |
+        | `waypoint` | `str`   | Short name or designator for the waypoint.       |
 
     label_column : str, optional
-        The name of the CSV column to use for labeling the waypoints (default: 'waypoint').
+        The name of the CSV column to use for labeling the waypoints on the map (default: `'waypoint'`).
+
+    lat_col : str, optional
+        The name of the CSV column containing latitude values (default: `'lat'`).
+
+    lon_col : str, optional
+        The name of the CSV column containing longitude values (default: `'lon'`).
 
     output_html : str, optional
         Path to save the generated map as an HTML file. If omitted, the map is not saved to disk.
@@ -40,7 +52,7 @@ def plot_ofp(
 
     Warnings
     -----
-    Rows with missing or invalid `lat` or `lon` values are automatically skipped.
+    Rows with missing or invalid lat/lon values are automatically skipped.
 
     Notes
     -----
@@ -54,6 +66,8 @@ def plot_ofp(
     plot_ofp(
         csv_path="https://raw.githubusercontent.com/sustainableaviation/jetfuelburn/refs/heads/main/tests/data/ofp.csv",
         label_column="waypoint",
+        lat_col="lat",
+        lon_col="lon",
     )
     ```
 
@@ -64,55 +78,75 @@ def plot_ofp(
     )
     ```
     """
-    if csv_path.startswith(("http://", "https://")):
-        with urllib.request.urlopen(csv_path) as response:
-            content = response.read().decode("utf-8")
-        f = io.StringIO(content)
+    df = pl.read_csv(csv_path)
+    if lat_col in df.columns and lon_col in df.columns:
+        df = df.with_columns(
+            pl.col(lat_col).cast(pl.Float64, strict=False),
+            pl.col(lon_col).cast(pl.Float64, strict=False),
+        ).drop_nulls(subset=[lat_col, lon_col])
     else:
-        f = open(csv_path, mode="r", encoding="utf-8")
[…]
