+++ src/jetfuelburn/utility/ofp.py
@@ -13,7 +13,12 @@ def generate_4d_trajectory(
     df_ofp: pl.DataFrame, 
     perf_data: dict, 
     resolution_min: float = 1.0, 
-    strategy: str = "leveloff" 
+    strategy: str = "leveloff", 
+    colname_wp: str = "waypoint",
+    colname_timeto: str = "timeto",
+    colname_alt: str = "alt",
+    colname_lat: str = "lat",
+    colname_lon: str = "lon",
 ) -> pl.DataFrame:
     r"""
     Generate a four-dimensional (4D) trajectory from a flight plan.
@@ -26,33 +31,82 @@ def generate_4d_trajectory(
     If an aircraft climbs or descends and reaches the altitude of the next waypoint before arriving at the waypoint's
     target time (inferred or given), it will level off at that target altitude until the waypoint is reached.
 
+    ??? info "Data Pipeline"
+
+        | waypoint | timeto | alt    | ...   |
+        |----------|--------|--------|-------|
+        | WP1      | 5      | 0      | ...   |
+        | WP2      | 5      | CLB    | ...   |
+        | WP3      | 3      | 10000  | ...   |
+        | WP4      | ...    | 10000  | ...   |
+
+        gives two separate flight segments, one with a climb:
+
+        | flight segment | waypoint | timeto | alt    | ... |
+        |----------------|----------|--------|--------|-----|
+        | segment 1      | WP1      | 5      | 0      | ... |
+        | segment 1      | WP2      | 5      | CLB    | ... |
+        | segment 1      | WP3      | ...    | 10000  | ... |
+
+        and one with constant altitude:
+
+        | flight segment | waypoint | timeto | alt    | ... |
+        |----------------|----------|--------|--------|-----|
+        | segment 2      | WP3      | 3      | 10000  | ... |
+        | segment 2      | WP4      | ...    | 10000  | ... |
+
+        With the aircraft performance data:
+
+        | climb regime      | min_alt | max_alt | rate |
+        |-------------------|---------|---------|------|
+        | initial climb     | 0       | 5000    | 2000 |
+        | final climb       | 5000    | 15000   | 1000 |
+
+        The function determines first performs a linear interpolation between the waypoint:
+
+        | timestamp           | alt    | flight segment | ... |
+        |---------------------|--------|----------------|-----|
+        | 2025-01-01 00:00:00 | 0      | segment 1      | ... |
+        | 2025-01-01 00:01:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:02:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:03:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:04:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:05:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:06:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:07:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:08:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:09:00 | CLB    | segment 1      | ... |
+        | 2025-01-01 00:10:00 | 10000  | segment 1      | ... |
+        | 2025-01-01 00:11:00 | 10000  | segment 2      | ... |
+        | 2025-01-01 00:12:00 | 10000  | segment 2      | ... |
+        | 2025-01-01 00:13:00 | 10000  | segment 2      | ... |
+        
+        Next, the function determines the rate of climb (ROC) or rate of descent (ROD) for each segment
+        based on the aircraft performance data and the altitude change between the waypoints.
+
[…]
