### 2026-03-15 6143bfa
started manual function re-write
