+++ src/jetfuelburn/utility/ofp.py
@@ -10,15 +10,16 @@ import math
 
 
 def generate_4d_trajectory(
-    df_ofp: pl.DataFrame, 
-    perf_data: dict, 
-    resolution_min: float = 1.0, 
-    strategy: str = "leveloff", 
+    df_ofp: pl.DataFrame,
+    perf_data: pl.DataFrame = None,
+    resolution_min: float = 1.0,
+    strategy: str = "leveloff",
     colname_wp: str = "waypoint",
     colname_timeto: str = "timeto",
     colname_alt: str = "alt",
     colname_lat: str = "lat",
     colname_lon: str = "lon",
+    timestamp_start: pd.Timestamp = pd.Timestamp('2025-01-01 00:00:00'),
 ) -> pl.DataFrame:
     r"""
     Generate a four-dimensional (4D) trajectory from a flight plan.
@@ -103,10 +104,27 @@ def generate_4d_trajectory(
 
     Parameters
     ----------
+    df_ofp : polars.DataFrame
+        Flight plan dataframe with columns corresponding to `colname_wp`, `colname_timeto`, corresponding
+        to time to navigate to the *next* waypoint, `colname_alt`, `colname_lat`, and `colname_lon`.
+    perf_data : polars.DataFrame
+        Aircraft performance dataframe with columns `state` (e.g., 'CLB', 'DES'), `min_alt`, `max_alt`, 
+        and `rate`.
+    resolution_min : float, default 1.0
+        The minimum time resolution in minutes for the generated trajectory.
+    strategy : str, default 'leveloff'
+        The altitude strategy to employ when arriving at altitude early.
+    colname_wp : str, default 'waypoint'
+    colname_timeto : str, default 'timeto'
+    colname_alt : str, default 'alt'
+    colname_lat : str, default 'lat'
+    colname_lon : str, default 'lon'
 
     Returns
     -------
     polars.DataFrame
+        A DataFrame with columns ``["timestamp", "lat", "lon", "alt"]`` at the requested
+        time resolution, with all null rows dropped.
 
     Warnings
     --------
@@ -129,176 +147,60 @@ def generate_4d_trajectory(
     If the rate of climb is such that the aircraft would reach the altitude defined at the next waypoint 
     before reaching the actual waypoint, the aircraft will level off at that altitude until the waypoint is reached.
     """
-    if strategy != "leveloff":
-        raise ValueError(f"Only strategy 'leveloff' implemented at the moment.")
+    if len(df_ofp) == 0:
+        raise ValueError("Empty flight plan provided.")
     for col in [colname_wp, colname_timeto, colname_alt, colname_lat, colname_lon]:
         if col not in df_ofp.columns:
-            raise ValueError(f"DataFrame is missing required column: {col}")
+            raise ValueError(f"Flight plan must contain column {col}.")
 
+    df = df_ofp.copy()
 
-    # Extract lists to iterate and mutate missing times
-    lats = df_ofp[colname_lat].to_list()
-    lons = df_ofp[colname_lon].to_list()
-    times = df_ofp[colname_timeto].to_list()
-    alts = df_ofp[colname_alt].to_list()
-    
-    # Helper to calculate time required to traverse from start_alt to end_alt
-    def calculate_transit_time(start_alt, end_alt):
-        if start_alt == end_alt:
-            return 0.0
-            
-        is_climb = end_alt > start_alt
-        brackets = perf_data.get('climb', []) if is_climb else perf_data.get('descent', [])
-        
-        # Sort brackets by min_alt ascending for climb, descending max_alt for descent
-        if is_climb:
-            brackets = sorted(brackets, key=lambda x: x['min_alt'])
-        else:
[…]
