### 2026-03-15 1e225e7
added data and draft of lookup function
