### 2026-03-15 a4e9129
added data and draft of lookup function
