+++ src/jetfuelburn/data/EurocontrolAPD/__init__.py
@@ -1,0 +1,6 @@
+# https://learningzone.eurocontrol.int/ilp/customs/ATCPFDB/default.aspx?
+
+# Note:
+# The data of this database is under copyright of EUROCONTROL
+# The example included in 'data.yaml' is only a fictious sample 
+# to test the functionality of the 'jetfuelburn.utility.ofp' module.
