### 2026-03-15 d2e2364
updated lookup function
