### 2026-03-15 d35c7f1
updated lookup function
