### 2026-03-15 b0a3e58
updated functions and requirements
