### 2026-03-15 401a166
updated functions and requirements
