### 2026-03-15 25ec504
added draft of ofp logic
