### 2026-03-15 1b6f899
completed initial draft
