+++ src/jetfuelburn/utility/ofp.py
@@ -47,14 +47,14 @@ def _get_aircraft_performance(
         max_alt = ureg(str(r["max_alt"])).to("ft")
         rate = ureg(str(r["rate"])).to("ft/min")
         
-        if min_alt ==  max_alt:
+        if min_alt == max_alt:
             raise ValueError(f"Degenerate altitude band (min == max) in regime {r.get('regime')!r}: {min_alt}")
 
         processed.append({
             "regime": r.get("regime"),
-            "min_alt": min(min_alt, max_alt),
-            "max_alt": max(min_alt, max_alt),
-            "rate": rate
+            "min_alt": min(min_alt, max_alt).to("ft"),
+            "max_alt": max(min_alt, max_alt).to("ft"),
+            "rate": rate.to("ft/min")
         })
         
     for p in processed:
@@ -64,127 +64,23 @@ def _get_aircraft_performance(
 
 
 def generate_4d_trajectory(
-    df_ofp: pl.DataFrame,
+    df_ofp: pd.DataFrame,
     aircraft_type: str,
     filepath_perf_data: Path,
     resolution_min: float = 1.0,
     strategy: str = "leveloff",
     colname_wp: str = "waypoint",
-    colname_timeto: str = "timeto",
+    colname_timecum: str = "timecum",
     colname_alt: str = "alt",
     colname_lat: str = "lat",
     colname_lon: str = "lon",
     timestamp_start: pd.Timestamp = pd.Timestamp('2025-01-01 00:00:00'),
-) -> pl.DataFrame:
+) -> pd.DataFrame:
     r"""
     Generate a four-dimensional (4D) trajectory from a flight plan.
 
-    The function interpolates positions (lat/lon) and altitudes (alt) mapping to a high-resolution timestamp
-    given the times to arrive at specific waypoints. Missing `timeto` values are inferred primarily by utilizing
-    the vertical rates defined in `perf_data` to calculate the required time to bridge altitude changes between
-    known points.
-
     If an aircraft climbs or descends and reaches the altitude of the next waypoint before arriving at the waypoint's
     target time (inferred or given), it will level off at that target altitude until the waypoint is reached.
-
-    ??? info "Data Pipeline"
-
-        | waypoint | timeto | alt    | ...   |
-        |----------|--------|--------|-------|
-        | WP1      | 5      | 0      | ...   |
-        | WP2      | 5      | CLB    | ...   |
-        | WP3      | 3      | 10000  | ...   |
-        | WP4      | ...    | 10000  | ...   |
-
-        gives two separate flight segments, one with a climb:
-
-        | flight segment | waypoint | timeto | alt    | ... |
-        |----------------|----------|--------|--------|-----|
-        | segment 1      | WP1      | 5      | 0      | ... |
-        | segment 1      | WP2      | 5      | CLB    | ... |
-        | segment 1      | WP3      | ...    | 10000  | ... |
-
-        and one with constant altitude:
-
-        | flight segment | waypoint | timeto | alt    | ... |
-        |----------------|----------|--------|--------|-----|
-        | segment 2      | WP3      | 3      | 10000  | ... |
-        | segment 2      | WP4      | ...    | 10000  | ... |
-
-        With the aircraft performance data:
-
-        | climb regime      | min_alt | max_alt | rate |
-        |-------------------|---------|---------|------|
-        | initial climb     | 0       | 5000    | 2000 |
-        | final climb       | 5000    | 15000   | 1000 |
-
-        | descent regime    | min_alt | max_alt | rate  |
-        |-------------------|---------|---------|-------|
-        | initial descent   | 15000   | 5000    | -2000 |
-        | final descent     | 5000    | 0       | -1000 |
-
[…]
