+++ CHANGES.md
@@ -3,13 +3,17 @@
 The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## `2.1.0` (15. March 2026)
+## `3.0.0` (15. March 2026)
 
 ## Improvements
 
-- Added a new `mapping` utility module for mapping-related calculations.
-- Added a new `ofp` utility module for flight plan-related calculations.
+- Added a new `utility.mapping` module for mapping-related calculations.
+- Added a new `utility.ofp` module for flight plan-related calculations.
 - Updated the installation instructions to include optional dependencies.
+
+### Breaking Changes
+
+- Renamed the `utility.math` module to `utility.mathematics` to avoid namespace conflicts with the Python built-in `math` module.
 
 ## `2.0.1` (20. February 2026)
 
