### 2026-03-15 6207f2f
attempted to make dataframe compatible with pint
