+++ src/jetfuelburn/utility/ofp.py
@@ -25,6 +25,48 @@ def _get_aircraft_performance(
     phase: str,
     alt: float | ureg.Quantity,
 ):
+    """
+    Look up the climb or descent rate for a given aircraft type and altitude.
+
+    Reads a YAML performance data file and returns the rate of climb (positive)
+    or rate of descent (negative) applicable to the supplied altitude, according
+    to the altitude-band regime defined for the aircraft and flight phase.
+
+    Parameters
+    ----------
+    filepath_perf_data : Path
+        Path to the YAML file containing aircraft performance data.
+        The file must follow the schema used by the EUROCONTROL APD dataset
+        (see `src/jetfuelburn/data/EurocontrolAPD/data.yaml`).
+    aircraft_type : str
+        ICAO aircraft type designator (e.g. ``'B123'``), matching a top-level
+        key in the YAML file.
+    phase : str
+        Flight phase; must be either ``'climb'`` or ``'descent'``.
+    alt : pint.Quantity
+        Current altitude with a length dimension (e.g. ``30000 * ureg.ft``).
+        The ``@ureg.check`` decorator enforces dimensional correctness.
+
+    Returns
+    -------
+    pint.Quantity
+        Rate of climb or descent in ``ft/min``.  Climb rates are positive;
+        descent rates are negative.
+
+    Raises
+    ------
+    ValueError
+        If *phase* is not ``'climb'`` or ``'descent'``.
+    ValueError
+        If *aircraft_type* is not present in the YAML file.
+    ValueError
+        If the flight phase is missing for the given *aircraft_type*.
+    ValueError
+        If a regime in the YAML file has ``min_alt == max_alt`` (degenerate band).
+    ValueError
+        If *alt* does not fall within any altitude band defined for the aircraft
+        and flight phase.
+    """
     with open(filepath_perf_data, "r") as f:
         data = yaml.safe_load(f)
 
@@ -67,13 +109,14 @@ def generate_4d_trajectory(
     df_ofp: pd.DataFrame,
     aircraft_type: str,
     filepath_perf_data: Path,
-    resolution_min: float = 1.0,
+    time_resolution: ureg.Quantity = 1.0 * ureg.minute,
     strategy: str = "leveloff",
     colname_wp: str = "waypoint",
     colname_timecum: str = "timecum",
-    colname_alt: str = "alt",
     colname_lat: str = "lat",
     colname_lon: str = "lon",
+    colname_alt: str = "alt",
+    unit_alt: str = "ft",
     timestamp_start: pd.Timestamp = pd.Timestamp('2025-01-01 00:00:00'),
 ) -> pd.DataFrame:
     r"""
@@ -81,6 +124,62 @@ def generate_4d_trajectory(
 
     If an aircraft climbs or descends and reaches the altitude of the next waypoint before arriving at the waypoint's
     target time (inferred or given), it will level off at that target altitude until the waypoint is reached.
+
+    Parameters
+    ----------
+    df_ofp : pd.DataFrame
+        Operational flight plan (OFP) as a :class:`pandas.DataFrame`.
+        Must contain at minimum the columns named by *colname_wp*,
+        *colname_timecum*, *colname_alt*, *colname_lat*, and *colname_lon*.
+        The altitude column may contain numeric values (in feet) **or** the
+        string tokens ``'CLB'`` / ``'DSC'`` for climb/descent waypoints whose
+        exact altitude is not yet known.
+    aircraft_type : str
+        ICAO aircraft type designator (e.g. ``'B123'``), passed directly to
+        :func:`_get_aircraft_performance`.
+    filepath_perf_data : Path
+        Path to the YAML performance data file; forwarded to
+        :func:`_get_aircraft_performance`.
+    time_resolution : float, optional
[…]
