+++ docs/api/mathematics.md
@@ -1,0 +1,1 @@
+::: jetfuelburn.utility.mathematics
