+++ src/jetfuelburn/utility/ofp.py
@@ -66,6 +66,20 @@ def _get_aircraft_performance(
     ValueError
         If *alt* does not fall within any altitude band defined for the aircraft
         and flight phase.
+
+    Example
+    -------
+    ```pyodide install='jetfuelburn'
+    import jetfuelburn
+    from jetfuelburn import ureg
+    
+    jetfuelburn._get_aircraft_performance(
+        jetfuelburn.data.EurocontrolAPD.data.yaml,
+        "B123",
+        "climb",
+        30000 * ureg.ft,
+    )
+    ```
     """
     with open(filepath_perf_data, "r") as f:
         data = yaml.safe_load(f)
@@ -140,9 +154,10 @@ def generate_4d_trajectory(
     filepath_perf_data : Path
         Path to the YAML performance data file; forwarded to
         :func:`_get_aircraft_performance`.
-    time_resolution : float, optional
-        Resampling resolution in minutes. The output trajectory is resampled
-        to this time resolution using linear interpolation. Default is ``1.0``.
+    time_resolution : pint.Quantity, optional
+        Resampling resolution as a Pint time quantity (e.g. ``1 * ureg.minute``).
+        The output trajectory is resampled to this resolution using linear
+        interpolation. Default is ``1 * ureg.minute``.
     strategy : str, optional
         Altitude interpolation strategy. Currently only ``'leveloff'`` is
         implemented (default). With this strategy the aircraft levels off at
@@ -160,6 +175,10 @@ def generate_4d_trajectory(
         Name of the latitude column in *df_ofp*. Default ``'lat'``.
     colname_lon : str, optional
         Name of the longitude column in *df_ofp*. Default ``'lon'``.
+    unit_alt : str, optional
+        Pint-compatible unit string for altitude values in *df_ofp* and in the
+        output (e.g. ``'ft'``, ``'m'``). Altitude columns are read and written
+        as plain :class:`float` in this unit. Default is ``'ft'``.
     timestamp_start : pd.Timestamp, optional
         UTC departure timestamp. Cumulative flight times from *colname_timecum*
         are added as :class:`pandas.Timedelta` offsets to this value to produce
@@ -201,6 +220,27 @@ def generate_4d_trajectory(
     In this representation, the angle of the trajectory corresponds to the rate of climb (ROC). 
     If the rate of climb is such that the aircraft would reach the altitude defined at the next waypoint 
     before reaching the actual waypoint, the aircraft will level off at that altitude until the waypoint is reached.
+
+    Example
+    -------
+    ```pyodide install='jetfuelburn'
+    import jetfuelburn
+    from jetfuelburn import ureg
+    df_ofp = pd.DataFrame({
+        "waypoint": ["DEP", "MID", "ARR"],
+        "alt": [0, "CLB", 5000],
+        "timecum": [0, 60, 120],
+        "lat": [47.0, 47.5, 48.0],
+        "lon": [8.0, 8.5, 9.0],
+    })
+    result = jetfuelburn.generate_4d_trajectory(
+        df_ofp,
+        "B123",
+        jetfuelburn.DATA_YAML,
+        time_resolution=1 * ureg.minute,
+        timestamp_start=pd.Timestamp("2025-01-01 00:00:00"),
+    )
+    ```
     """
     if len(df_ofp) == 0:
         raise ValueError("Empty flight plan provided.")
@@ -209,10 +249,6 @@ def generate_4d_trajectory(
             raise ValueError(f"Flight plan must contain column {col}.")
 
     df = df_ofp.copy()
-
-    """
-    add example here later
-    """
 
     df['timestamp'] = timestamp_start + pd.to_timedelta(df[colname_timecum], unit="min")
 
@@ -236,8 +272,7 @@ def generate_4d_trajectory(
[…]
