+++ docs/api/mapping.md
@@ -1,6 +1,6 @@
 !!! warning "Optional Dependencies"
 
-    To use this module, you need to [install `jetfuelburn` with all optional dependencies](installation.md):
+    To use this module, you need to [install `jetfuelburn` with all optional dependencies](../installation.md):
 
     ```bash
     pip install jetfuelburn[optionaldependencies]
