### 2026-03-16 034015e
updated functions
