+++ verify_ofp.py
@@ -2,6 +2,7 @@ import pandas as pd
 import numpy as np
 import polars as pl
 from pathlib import Path
+from jetfuelburn import ureg
 from jetfuelburn.utility.ofp import generate_4d_trajectory
 
 # Create a mock OFP dataframe (Polars as expected by the function's type hint, 
@@ -13,7 +14,7 @@ from jetfuelburn.utility.ofp import generate_4d_trajectory
 df_ofp = pd.DataFrame({
     'waypoint': ['WP1', 'WP2', 'WP3'],
     'alt': [0.0, 'CLB', 10000.0],
-    'timeto': [0.0, 10.0, 10.0],
+    'timecum': [0.0, 10.0, 20.0],
     'lat': [37.0, 38.0, 39.0],
     'lon': [-122.0, -121.0, -120.0]
 })
@@ -23,10 +24,10 @@ import yaml
 mock_data = {
     "B123": {
         "climb": [
-            {"min_alt": 0, "max_alt": 20000, "rate": "2000 ft/min", "regime": "climb"}
+            {"min_alt": "0 ft", "max_alt": "20000 ft", "rate": "2000 ft/min", "regime": "climb"}
         ],
         "descent": [
-            {"min_alt": 20000, "max_alt": 0, "rate": "-1500 ft/min", "regime": "descent"}
+            {"min_alt": "20000 ft", "max_alt": "0 ft", "rate": "-1500 ft/min", "regime": "descent"}
         ]
     }
 }
@@ -38,8 +39,8 @@ try:
     df_4d = generate_4d_trajectory(
         df_ofp=df_ofp,
         aircraft_type="B123",
-        filepath_perf_data=perf_data_path,
-        resolution_min=1.0
+        perf_data_path=perf_data_path,
+        time_resolution=1.0 * ureg.minute
     )
     print("Success status: Pass")
     print(df_4d[['timestamp', 'alt']].head(15))
