### 2026-03-16 51ea843
updated functions
