+++ tests/test_mapping.py
@@ -2,8 +2,9 @@ import pytest
 import os
 import tempfile
 import csv
-import ipyleaflet
-from jetfuelburn.utility.mapping import plot_ofp
+import pandas as pd
+import plotly.graph_objects as go
+from jetfuelburn.utility.mapping import plot_ofp_2d, plot_ofp_1d
 
 
 class TestMapping:
@@ -46,129 +47,95 @@ class TestMapping:
         if os.path.exists(temp_path):
             os.remove(temp_path)
 
-    def test_plot_ofp_returns_map(self, sample_csv):
-        """Test that plot_ofp returns an ipyleaflet Map."""
-        m = plot_ofp(sample_csv)
-        assert isinstance(m, ipyleaflet.Map)
+    def test_plot_ofp_2d_returns_figure(self, sample_csv):
+        """Test that plot_ofp_2d returns a Plotly Figure."""
+        fig = plot_ofp_2d(sample_csv)
+        assert isinstance(fig, go.Figure)
 
-    def test_plot_ofp_layer_counts(self, sample_csv):
-        """Test that the correct number of layers are added to the map.
+    def test_plot_ofp_2d_trace_counts(self, sample_csv):
+        """Test that the correct number of traces are added to the figure.
 
-        With 2 valid rows (ZRH, SFO) every row is either start or end, so:
-          1 TileLayer (base)
-        + 1 Polyline
-        + 1 Marker  (ZRH – start endpoint)
-        + 1 Marker  (ZRH – label DivIcon)
-        + 1 Marker  (SFO – end endpoint)
-        + 1 Marker  (SFO – label DivIcon)
-        = 6 layers
+        With 2 valid rows (ZRH, SFO):
+          1 trace for the path line
+        + 1 trace for ZRH
+        + 1 trace for SFO
+        = 3 traces
         """
-        m = plot_ofp(sample_csv)
+        fig = plot_ofp_2d(sample_csv)
+        assert len(fig.data) == 3
+        # Check that we have Scattermapbox traces
+        assert all(isinstance(t, go.Scattermapbox) for t in fig.data)
 
-        layer_types = [type(layer).__name__ for layer in m.layers]
-        print(f"Layers found: {layer_types}")
-
-        assert any(isinstance(l, ipyleaflet.Polyline) for l in m.layers)
-        assert any(isinstance(l, ipyleaflet.Marker) for l in m.layers)
-        assert not any(isinstance(l, ipyleaflet.CircleMarker) for l in m.layers)
-        assert len(m.layers) == 6
-
-    def test_plot_ofp_three_rows(self):
-        """Test layer counts when there is a middle waypoint (non-endpoint CircleMarker).
-
-        With 3 valid rows:
-          1 TileLayer
-        + 1 Polyline
-        + 1 Marker  (start endpoint)
-        + 1 Marker  (start label)
-        + 1 CircleMarker (middle)
-        + 1 Marker  (middle label)
-        + 1 Marker  (end endpoint)
-        + 1 Marker  (end label)
-        = 8 layers
-        """
-        content = [
-            ["waypoint", "lat", "lon"],
-            ["ZRH", 47.4582, 8.5555],
-            ["MID", 50.0, -10.0],
-            ["SFO", 37.6188, -122.3751],
-        ]
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".csv", delete=False, newline="") as f:
-            writer = csv.writer(f)
-            writer.writerows(content)
-            temp_path = f.name
-
-        try:
-            m = plot_ofp(temp_path)
-            assert any(isinstance(l, ipyleaflet.CircleMarker) for l in m.layers)
-            assert len(m.layers) == 8
-        finally:
-            if os.path.exists(temp_path):
-                os.remove(temp_path)
-
[…]

[…]
