+++ pyproject.toml
@@ -48,7 +48,7 @@ testing = [
 ]
 optionaldependencies = [
     "pandas",
-    "plotly",
+    "plotly>=5.24.0",
 ]
 
 [tool.setuptools]

[…]
