+++ tests/test_ofp.py
@@ -1,3 +1,4 @@
+# %%
 import math
 import textwrap
 import tempfile
@@ -20,47 +21,14 @@ DATA_YAML = Path(__file__).parent.parent / "src" / "jetfuelburn" / "data" / "Eur
 OFP_CSV = Path(__file__).parent / "data" / "ofp" / "ofp_basic.csv"
 """Path to the OFP test fixture."""
 
+PERF_YAML = Path(__file__).parent / "data" / "ofp" / "performance.yaml"
+"""Path to the performance test fixture."""
+
 
 @pytest.fixture()
-def tmp_yaml(tmp_path: Path) -> Path:
-    """
-    Write a minimal, self-contained performance YAML to a temporary file
-    and return its path.  The aircraft key is ``'TEST'``.
-
-    Climb bands
-    -----------
-    initial_climb : 0–5 000 ft   @ +1 000 ft/min
-    high_climb    : 5 000–35 000 ft @ +500 ft/min
-
-    Descent bands
-    -------------
-    high_descent : 35 000–5 000 ft @ -1 000 ft/min
-    approach     : 5 000–0 ft      @ -500 ft/min
-    """
-    content = textwrap.dedent("""\
-        TEST:
-          climb:
-          - regime: initial_climb
-            min_alt: 0 ft
-            max_alt: 5000 ft
-            rate: 1000 ft/min
-          - regime: high_climb
-            min_alt: 5000 ft
-            max_alt: 35000 ft
-            rate: 500 ft/min
-          descent:
-          - regime: high_descent
-            min_alt: 5000 ft
-            max_alt: 35000 ft
-            rate: -1000 ft/min
-          - regime: approach
-            min_alt: 0 ft
-            max_alt: 5000 ft
-            rate: -500 ft/min
-    """)
-    p = tmp_path / "perf.yaml"
-    p.write_text(content)
-    return p
+def tmp_yaml() -> Path:
+    """Return the path to the static performance YAML test fixture."""
+    return PERF_YAML
 
 
 @pytest.fixture()
@@ -79,36 +47,54 @@ class TestGetAircraftPerformance:
 
     # --- happy-path lookups ------------------------------------------------
 
-    def test_climb_lower_band(self, tmp_yaml: Path):
+    def test_climb_lower_band(
+        self,
+        tmp_yaml: Path,
+    ):
         """Altitude in the lower climb band returns the correct rate."""
         rate = _get_aircraft_performance(tmp_yaml, "TEST", "climb", 2500 * ureg.ft)
         assert rate.check("[length]/[time]")
         assert math.isclose(rate.to("ft/min").magnitude, 1000.0)
 
-    def test_climb_upper_band(self, tmp_yaml: Path):
+    def test_climb_upper_band(
+        self,
+        tmp_yaml: Path,
+    ):
         """Altitude in the higher climb band returns the correct rate."""
         rate = _get_aircraft_performance(tmp_yaml, "TEST", "climb", 20000 * ureg.ft)
         assert rate.check("[length]/[time]")
         assert math.isclose(rate.to("ft/min").magnitude, 500.0)
 
-    def test_descent_upper_band(self, tmp_yaml: Path):
+    def test_descent_upper_band(
+        self,
+        tmp_yaml: Path,
+    ):
         """Altitude in the higher descent band returns a negative rate."""
         rate = _get_aircraft_performance(tmp_yaml, "TEST", "descent", 20000 * ureg.ft)
         assert rate.check("[length]/[time]")
         assert math.isclose(rate.to("ft/min").magnitude, -1000.0)
 
-    def test_descent_lower_band(self, tmp_yaml: Path):
+    def test_descent_lower_band(
+        self,
+        tmp_yaml: Path,
+    ):
[…]
