+++ src/jetfuelburn/utility/ofp.py
@@ -1,3 +1,4 @@
+# %%
 try:
     import pandas as pd
 except ImportError as e:
@@ -141,9 +142,9 @@ def _get_aircraft_performance(
 
 
 def generate_4d_trajectory(
-    df_ofp: pd.DataFrame,
+    df_ofp: pd.DataFrame | str | Path,
     aircraft_type: str,
-    perf_data_path: Path,
+    perf_data_path: Path | str,
     time_resolution: ureg.Quantity = 1.0 * ureg.minute,
     strategy: str = "leveloff",
     colname_wp: str = "waypoint",
@@ -263,6 +264,9 @@ def generate_4d_trajectory(
     )
     ```
     """
+    if isinstance(df_ofp, (str, Path)):
+        df_ofp = pd.read_csv(df_ofp)
+
     if len(df_ofp) == 0:
         raise ValueError("Empty flight plan provided.")
     for col in [colname_wp, colname_timecum, colname_alt, colname_lat, colname_lon]:
@@ -297,58 +301,90 @@ def generate_4d_trajectory(
     next_num = df['alt_filled'].bfill().shift(-1)
     df['next_alt'] = next_num.fillna(df['alt_filled'].ffill())
 
-    for idx, row in df.iterrows():
-        if idx == 0:
-            continue
-
-        segment_initial_alt = df.at[idx - 1, 'alt_filled']
-        segment_time = (df.at[idx, 'timestamp'] - df.at[idx - 1, 'timestamp']).total_seconds() / 60 * ureg.minute
-
-        if pd.isnull(row['alt_filled']):
-            if segment_initial_alt < row['next_alt']: # climb segment
-                if segment_initial_alt >= row['next_alt']: # level-off
-                    df.at[idx, 'alt_filled'] = row['next_alt']
-                else: # continued climb
-                    rate_of_climb = _get_aircraft_performance(
-                        perf_data_path=perf_data_path,
-                        aircraft_type=aircraft_type,
-                        phase="climb",
-                        alt=segment_initial_alt * ureg(unit_alt),
-                    )
-                    segment_final_alt = (segment_initial_alt * ureg(unit_alt) + rate_of_climb * segment_time).to(unit_alt).magnitude
-                    if segment_final_alt > row['next_alt']: # level-off
-                        df.at[idx, 'alt_filled'] = row['next_alt']
-                    else:
-                        df.at[idx, 'alt_filled'] = segment_final_alt
-            elif segment_initial_alt > row['next_alt']: # descent segment
-                if segment_initial_alt <= row['next_alt']: # level-off
-                    df.at[idx, 'alt_filled'] = row['next_alt']
-                else: # continued descent
-                    rate_of_descent = _get_aircraft_performance(
-                        perf_data_path=perf_data_path,
-                        aircraft_type=aircraft_type,
-                        phase="descent",
-                        alt=segment_initial_alt * ureg(unit_alt),
-                    )
-                    segment_final_alt = (segment_initial_alt * ureg(unit_alt) + rate_of_descent * segment_time).to(unit_alt).magnitude
-                    if segment_final_alt < row['next_alt']: # level-off
-                        df.at[idx, 'alt_filled'] = row['next_alt']
-                    else:
-                        df.at[idx, 'alt_filled'] = segment_final_alt
-            else: # level flight
-                df.at[idx, 'alt_filled'] = row['next_alt']
+    # Simulation logic to generate high-resolution 4D trajectory
+    trajectory_points = []
     
-    list_interpolation_columns = ['alt_filled', colname_lat, colname_lon]
+    curr_t = df.at[0, 'timestamp']
[…]
