+++ tests/test_ofp.py
@@ -10,12 +10,18 @@ import pytest
 from jetfuelburn import ureg
 from jetfuelburn.utility.ofp import _get_aircraft_performance, generate_4d_trajectory
 
-
 # ---------------------------------------------------------------------------
 # Helpers / shared fixtures
 # ---------------------------------------------------------------------------
 
-DATA_YAML = Path(__file__).parent.parent / "src" / "jetfuelburn" / "data" / "EurocontrolAPD" / "data.yaml"
+DATA_YAML = (
+    Path(__file__).parent.parent
+    / "src"
+    / "jetfuelburn"
+    / "data"
+    / "EurocontrolAPD"
+    / "data.yaml"
+)
 """Path to the real EUROCONTROL APD YAML file shipped with the package."""
 
 OFP_CSV = Path(__file__).parent / "data" / "ofp" / "ofp_basic.csv"
@@ -179,13 +185,15 @@ class TestGenerate4DTrajectory:
         tmp_yaml: Path,
     ):
         """A DataFrame missing a required column raises :class:`ValueError`."""
-        df = pd.DataFrame({
-            "waypoint": ["A", "B"],
-            "timecum": [0, 10],
-            "lat": [0.0, 1.0],
-            "lon": [0.0, 1.0],
-            # 'alt' is intentionally omitted
-        })
+        df = pd.DataFrame(
+            {
+                "waypoint": ["A", "B"],
+                "timecum": [0, 10],
+                "lat": [0.0, 1.0],
+                "lon": [0.0, 1.0],
+                # 'alt' is intentionally omitted
+            }
+        )
         with pytest.raises(ValueError, match="alt"):
             generate_4d_trajectory(df, "TEST", tmp_yaml)
 
@@ -196,13 +204,15 @@ class TestGenerate4DTrajectory:
         tmp_yaml: Path,
     ):
         """The return value is a :class:`pandas.DataFrame`."""
-        df = pd.DataFrame({
-            "waypoint": ["DEP", "ARR"],
-            "alt": [0, 5000],
-            "timecum": [0, 60],
-            "lat": [47.0, 48.0],
-            "lon": [8.0, 9.0],
-        })
+        df = pd.DataFrame(
+            {
+                "waypoint": ["DEP", "ARR"],
+                "alt": [0, 5000],
+                "timecum": [0, 60],
+                "lat": [47.0, 48.0],
+                "lon": [8.0, 9.0],
+            }
+        )
         result = generate_4d_trajectory(df, "TEST", tmp_yaml)
         assert isinstance(result, pd.DataFrame)
 
@@ -211,13 +221,15 @@ class TestGenerate4DTrajectory:
         tmp_yaml: Path,
     ):
         """The output DataFrame contains the ``alt_filled`` column."""
-        df = pd.DataFrame({
-            "waypoint": ["DEP", "ARR"],
-            "alt": [0, 5000],
-            "timecum": [0, 60],
-            "lat": [47.0, 48.0],
-            "lon": [8.0, 9.0],
-        })
+        df = pd.DataFrame(
+            {
+                "waypoint": ["DEP", "ARR"],
+                "alt": [0, 5000],
+                "timecum": [0, 60],
+                "lat": [47.0, 48.0],
+                "lon": [8.0, 9.0],
+            }
+        )
         result = generate_4d_trajectory(df, "TEST", tmp_yaml)
         assert "alt_filled" in result.columns
 
@@ -235,14 +247,18 @@ class TestGenerate4DTrajectory:
         DataFrame; the resampled timestamps are available in the ``'timestamp'``
         column (carried through the merge from ``df_resampled``).
         """
-        df = pd.DataFrame({
-            "waypoint": ["DEP", "ARR"],
-            "alt": [0, 5000],
[…]
