+++ src/jetfuelburn/utility/ofp.py
@@ -18,7 +18,7 @@ from jetfuelburn import ureg
     None,
     None,
     "[length]",
-)    
+)
 def _get_aircraft_performance(
     perf_data_path: Path,
     aircraft_type: str,
@@ -33,7 +33,7 @@ def _get_aircraft_performance(
     to the altitude-band regime defined for the aircraft and flight phase.
 
     Performance data must be provided in the following format:
-    
+
     ```yaml
     B123:
         climb:
@@ -94,7 +94,7 @@ def _get_aircraft_performance(
     ```pyodide install='jetfuelburn'
     import jetfuelburn
     from jetfuelburn import ureg
-    
+
     jetfuelburn._get_aircraft_performance(
         jetfuelburn.data.EurocontrolAPD.data.yaml,
         "B123",
@@ -111,11 +111,19 @@ def _get_aircraft_performance(
 
     if aircraft_type not in data:
         available = sorted(data.keys()) if data else []
-        raise ValueError(f"Aircraft type {aircraft_type!r} not found in {perf_data_path}. Available: {available}")
-    
+        raise ValueError(
+            f"Aircraft type {aircraft_type!r} not found in {perf_data_path}. Available: {available}"
+        )
+
     aircraft_info = data[aircraft_type]
-    if aircraft_info is None or phase not in aircraft_info or aircraft_info[phase] is None:
-        raise ValueError(f"Flight phase {phase!r} not found for aircraft type {aircraft_type!r} in {perf_data_path}")
+    if (
+        aircraft_info is None
+        or phase not in aircraft_info
+        or aircraft_info[phase] is None
+    ):
+        raise ValueError(
+            f"Flight phase {phase!r} not found for aircraft type {aircraft_type!r} in {perf_data_path}"
+        )
 
     regimes = aircraft_info[phase]
 
@@ -124,21 +132,27 @@ def _get_aircraft_performance(
         min_alt = ureg(str(r["min_alt"])).to("ft")
         max_alt = ureg(str(r["max_alt"])).to("ft")
         rate = ureg(str(r["rate"])).to("ft/min")
-        
-        if min_alt == max_alt:
-            raise ValueError(f"Degenerate altitude band (min == max) in regime {r.get('regime')!r}: {min_alt}")
 
-        processed.append({
-            "regime": r.get("regime"),
-            "min_alt": min(min_alt, max_alt).to("ft"),
-            "max_alt": max(min_alt, max_alt).to("ft"),
-            "rate": rate.to("ft/min")
-        })
-        
+        if min_alt == max_alt:
+            raise ValueError(
+                f"Degenerate altitude band (min == max) in regime {r.get('regime')!r}: {min_alt}"
+            )
+
+        processed.append(
+            {
+                "regime": r.get("regime"),
+                "min_alt": min(min_alt, max_alt).to("ft"),
+                "max_alt": max(min_alt, max_alt).to("ft"),
+                "rate": rate.to("ft/min"),
+            }
+        )
+
     for p in processed:
         if p["min_alt"] <= alt <= p["max_alt"]:
             return p["rate"]
-    raise ValueError(f"Altitude {alt} not found in any altitude band for aircraft type {aircraft_type!r} in {perf_data_path}")
+    raise ValueError(
+        f"Altitude {alt} not found in any altitude band for aircraft type {aircraft_type!r} in {perf_data_path}"
+    )
 
 
 def generate_4d_trajectory(
@@ -153,7 +167,7 @@ def generate_4d_trajectory(
     colname_lon: str = "lon",
     colname_alt: str = "alt",
     unit_alt: str = "ft",
-    timestamp_start: pd.Timestamp = pd.Timestamp('2025-01-01 00:00:00'),
[…]
