+++ src/jetfuelburn/data/EurocontrolAPD/__init__.py
@@ -2,5 +2,5 @@
 
 # Note:
 # The data of this database is under copyright of EUROCONTROL
-# The example included in 'data.yaml' is only a fictious sample 
+# The example included in 'data.yaml' is only a fictious sample
 # to test the functionality of the 'jetfuelburn.utility.ofp' module.

[…]
