+++ CHANGES.md
@@ -3,7 +3,7 @@
 The format of this log is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## `3.0.0` (15. March 2026)
+## `3.0.0` (16. March 2026)
 
 ## Improvements
 
