### 2026-03-16 ee84992
updated plotting functions
