### 2026-03-17 42fe66a
working on extracting the movements per aircraft type from the DOT statistics
