+++ src/jetfuelburn/statistics.py
@@ -449,3 +449,18 @@ class usdot:
         fuelburn = (aircraft_data["Fuel/Revenue Seat Distance"] * ureg("kg/km")) * R
         fuelburn = fuelburn.to("kg")
         return fuelburn
+    def calculate_movements(
+            year: int,
+            acft: str,
+    ) -> dict:
+        if year not in usdot._years:
+            raise ValueError(f"No data available for year '{year}'.")
+        if acft not in usdot._aircraft_data[year]:
+            raise ValueError(
+                f"US DOT Aircraft Designator '{acft}' not found in model data."
+            )
+        else:
+            aircraft_data = usdot._aircraft_data[year][acft]
+
+        movements = (aircraft_data["Number of flights performed"])
+        return movements
