### 2026-03-17 07790f8
working on extracting the movements per aircraft type from the DOT statistics
