### 2026-03-18 b4328d5
Extracted the movements per aircraft type from the DOT statistics. Now on to the graphs
