### 2026-03-18 52ad057
Extracted the movements per aircraft type from the DOT statistics and made a graph with the top 15 aircraft types
