+++ processing/usdot/usdot.py
@@ -124,6 +124,9 @@ def process_data_usdot_t2(
     df_t2["Number of flights performed"] = (
         df_t2["REV_ACRFT_DEP_PERF"]
     )
+    df_t2["Revenue PAX Miles"] = (
+        df_t2["REV_PAX_MILES"]
+    )
 
     # SANITY CHECKS
 
@@ -138,6 +141,7 @@ def process_data_usdot_t2(
         "Fuel/Available Weight Distance",
         "Fuel/Revenue Weight Distance",
         "Number of flights performed",
+        "Revenue PAX Miles",
     ]
     df_t2 = df_t2[list_return_columns]
 
@@ -149,6 +153,7 @@ def process_data_usdot_t2(
         "Fuel/Available Weight Distance": "mean",
         "Fuel/Revenue Weight Distance": "mean",
         "Number of flights performed": "sum",
+        "Revenue PAX Miles": "sum",
     }
     df_t2 = df_t2.groupby(
         by="Aircraft Designation (US DOT Schedule T2)",
@@ -217,3 +222,48 @@ df_dequantified.to_json(
     orient="index",
     indent=4,
 )
+
+import plotly.graph_objects as go
+
+# Schritt 1: Jedes Jahr separat laden und speichern
+df24 = process_data_usdot_t2(
+    path_csv_aircraft_types="data/L_AIRCRAFT_TYPE.csv",
+    path_csv_t2="data/T_SCHEDULE_T2_2024.csv",
+).pint.dequantify()
+df24.columns = df24.columns.droplevel(1)
+
+df18 = process_data_usdot_t2(
+    path_csv_aircraft_types="data/L_AIRCRAFT_TYPE.csv",
+    path_csv_t2="data/T_SCHEDULE_T2_2018.csv",
+).pint.dequantify()
+df18.columns = df18.columns.droplevel(1)
+
+df13 = process_data_usdot_t2(
+    path_csv_aircraft_types="data/L_AIRCRAFT_TYPE.csv",
+    path_csv_t2="data/T_SCHEDULE_T2_2013.csv",
+).pint.dequantify()
+df13.columns = df13.columns.droplevel(1)
+
+# Schritt 2: Top 15 Flugzeugtypen aus 2024 als Referenz
+top_types = (
+    df24["Number of flights performed"]
+    .sort_values(ascending=False)
+    .head(15)
+    .index.tolist()
+)
+
+# Schritt 3: Diagramm erstellen
+datasets = {"2013": df13, "2018": df18, "2024": df24}
+
+fig = go.Figure()
+for year, df in datasets.items():
+    df_filtered = df.loc[df.index.isin(top_types), "Number of flights performed"]
+    fig.add_trace(go.Bar(name=year, x=df_filtered.index, y=df_filtered.values))
+
+fig.update_layout(
+    barmode="group",
+    title="Flight Movements by Aircraft Type",
+    xaxis_tickangle=-45,
+    yaxis_title="Movements",
+)
+fig.show()
