### 2026-03-19 fc2dceb
added 2025 to the database (only until September 2025)
