### 2026-03-19 1e7bfc7
added 2025 to the database (only until September 2025)
