### 2026-03-19 7ba246e
added 2025 to the database (only until September 2025)
