### 2026-03-24 5be574a
Added avg trip distance, avg flight time and freight per km to the graphs
