### 2026-03-24 e2bb930
Added avg trip distance, avg flight time and freight per km to the graphs
