+++ processing/usdot/usdot.py
@@ -58,6 +58,10 @@ def process_data_usdot_t2(
         'REV_TON_MILES_240': 'REV_TON_MILES',
         'AVL_TON_MILES_280': 'AVL_TON_MILES',
         'AIRCRAFT_FUELS_921': 'AIRCRAFT_FUELS',
+        'REV_ACRFT_MILES_FLOWN_410': 'REV_ACRFT_MILES_FLOWN',
+        'REV_ACRFT_HRS_AIRBORNE_610': 'REV_ACRFT_HRS_AIRBORNE',
+        'REV_TON_MILES_FREIGHT_247': 'REV_TON_MILES_FREIGHT',
+        'REV_TON_MILES_MAIL_249': 'REV_TON_MILES_MAIL',
         'CARRIER_GROUP': 'CARRIER_GROUP',
         'AIRCRAFT_CONFIG': 'AIRCRAFT_CONFIG',
         'AIRCRAFT_TYPE': 'AIRCRAFT_TYPE',
@@ -71,6 +75,10 @@ def process_data_usdot_t2(
         "REV_TON_MILES": "pint[miles*short_ton]",
         "AVL_TON_MILES": "pint[miles*short_ton]",
         "AIRCRAFT_FUELS": "pint[gallons]",
+        "REV_ACRFT_MILES_FLOWN": "pint[miles]",
+        "REV_ACRFT_HRS_AIRBORNE": "pint[hours]",
+        "REV_TON_MILES_FREIGHT": "pint[miles*short_ton]",
+        "REV_TON_MILES_MAIL": "pint[miles*short_ton]",
         "CARRIER_GROUP": "pint[]",
         "AIRCRAFT_CONFIG": "pint[]",
         "AIRCRAFT_TYPE": "pint[]",
@@ -88,6 +96,9 @@ def process_data_usdot_t2(
     df_t2["REV_PAX_MILES"] = df_t2["REV_PAX_MILES"].pint.to(ureg("km"))
     df_t2["REV_TON_MILES"] = df_t2["REV_TON_MILES"].pint.to(ureg("km*kg"))
     df_t2["AVL_TON_MILES"] = df_t2["AVL_TON_MILES"].pint.to(ureg("km*kg"))
+    df_t2["REV_ACRFT_MILES_FLOWN"] = df_t2["REV_ACRFT_MILES_FLOWN"].pint.to(ureg("km"))
+    df_t2["REV_TON_MILES_FREIGHT"] = df_t2["REV_TON_MILES_FREIGHT"].pint.to(ureg("km*kg"))
+    df_t2["REV_TON_MILES_MAIL"] = df_t2["REV_TON_MILES_MAIL"].pint.to(ureg("km*kg"))
 
     # DATA FILTERING
 
@@ -101,6 +112,10 @@ def process_data_usdot_t2(
         "AVL_TON_MILES",
         "AIRCRAFT_FUELS",
         "REV_ACRFT_DEP_PERF",
+        "REV_ACRFT_MILES_FLOWN",
+        "REV_ACRFT_HRS_AIRBORNE",
+        "REV_TON_MILES_FREIGHT",
+        "REV_TON_MILES_MAIL",
     ]
     df_t2[list_numeric_columns] = df_t2[list_numeric_columns].replace(
         to_replace=0, value=pd.NA
@@ -124,6 +139,16 @@ def process_data_usdot_t2(
 
     df_t2["Revenue PAX km"] = df_t2["REV_PAX_MILES"]
 
+    df_t2["Average trip distance"] = (
+            df_t2["REV_ACRFT_MILES_FLOWN"] / df_t2["REV_ACRFT_DEP_PERF"]
+    )
[…]
