### 2026-03-27 70e30ae
Jupyter Notebook for the comparison of Seymour, Sacchi, Montlaur and Yanto models.
