### 2026-04-01 36965fb
Added distance, cargo and time functions to statistics.py Added 3 years to the US DOT data. Added .idea to .gitignore
