### 2026-04-01 ff995e4
improved descent behavior in generate_4d_trajectory function, added test for descent from TOD
