+++ tests/test_ofp.py
@@ -424,18 +424,22 @@ class TestGenerate4DTrajectory:
         the aircraft must still be at 35000 ft at t=10 min.
         The old (buggy) behaviour would already be at 25000 ft by t=10 min.
         """
-        df = pd.DataFrame({
-            "waypoint": ["DEP", "MID", "ARR"],
-            "alt": [35000, "DSC", 0],
-            "timecum": [0, 60, 120],
-            "lat": [47.0, 47.5, 48.0],
-            "lon": [8.0, 8.5, 9.0],
-        })
-        result = generate_4d_trajectory(df, "TEST", tmp_yaml, time_resolution=1 * ureg.minute)
+        df = pd.DataFrame(
+            {
+                "waypoint": ["DEP", "MID", "ARR"],
+                "alt": [35000, "DSC", 0],
+                "timecum": [0, 60, 120],
+                "lat": [47.0, 47.5, 48.0],
+                "lon": [8.0, 8.5, 9.0],
+            }
+        )
+        result = generate_4d_trajectory(
+            df, "TEST", tmp_yaml, time_resolution=1 * ureg.minute
+        )
         # At t=10 min the aircraft should still be at cruise altitude (TOD not yet reached)
         ts_check = pd.Timestamp("2025-01-01 00:10:00")
         row = result[result["timestamp"] == ts_check]
         assert len(row) > 0, "Expected a row at t=10 min"
-        assert row["alt_filled"].iloc[0] > 30000, (
-            "Aircraft should still be near cruise altitude at t=10 min (TOD not yet reached)"
-        )
+        assert (
+            row["alt_filled"].iloc[0] > 30000
+        ), "Aircraft should still be near cruise altitude at t=10 min (TOD not yet reached)"
