+++ src/jetfuelburn/reducedorder.py
@@ -2085,6 +2085,11 @@ class eea_emission_inventory_2009:
     ----------
     - [EMEP/EEA air pollutant emission inventory guidebook - 2009, Part B, Section 1 (Energy), Subsection 1.A.3.a (`Aviation_annex.zip`)](https://www.eea.europa.eu/en/analysis/publications/emep-eea-emission-inventory-guidebook-2009)
 
+    Notes
+    -----
+    Data from `1.A.3.a Aviation vs2.3spreadsheet2-1updated2007.xlsx` [used with permission](../_static/permissions/opensky_network_data_access_confirmation.pdf) 
+    from the Swedish Defence Research Agency (foi.se).
+
     Warnings
     --------
     Note that only EEA data between 1999-2009 was based on Piano simulations, as documented in the `Aviation_annex.zip` file.
